Crate scrypto

The Scrypto Standard Library

The Scrypto Standard Library is the foundation of Scrypto blueprints, a set of minimal and shared abstractions on top of Radix Engine. It enables asset-oriented programming for feature-rich DeFi dApps.

If you know the name of what you’re looking for, the fastest way to find it is to use the search bar at the top of the page.

Modules

address

RE bech32 address library.

api

blueprints

component

Scrypto component abstraction.

constants

crypto

RE crypto library

crypto_utils

Scrypto crypto utilities abstraction.

data

RE scrypto data model.

engine

Scrypto engine abstraction.

macros 🔒

math

RE math library.

modules

Scrypto module abstraction.

network

RE network identifier model.

prelude

Scrypto preludes.

resource

Scrypto resource abstraction.

runtime

Scrypto runtime abstraction.

schema

time

RE time library.

types

Macros

component_royalties

component_royalty_config

debug

enable_function_auth

enable_method_auth

enable_package_royalties

error

extern_blueprint_internal

external_functions

external_methods

include_code

Includes the WASM file of a Scrypto package.

include_schema

Includes the schema file of a Scrypto package.

info

internal_add_role

internal_component_royalty_entry

main_accessibility

method_accessibilities

method_accessibility

role_list

roles

Roles macro for main module

this_package

to_role_key

trace

warn

Functions

set_up_panic_hook

Sets up panic hook.

Attribute Macros

blueprint

Declares a blueprint.

Derive Macros

NonFungibleData

Derive code that describe a non-fungible data structure.