Module scrypto::api

Modules

actor_api

actor_index_api

actor_key_value_entry_api

actor_sorted_index_api

blueprint_api

field_api

key_value_entry_api

key_value_store_api

node_modules

object_api

system_modules

Structs

AttachedModuleIdIter

An iterator over the variants of Self

EventFlags

FieldValue

GenericArgs

KVEntry

LocalKeyValueStoreDataSchema

LockFlags

ModuleIdIter

An iterator over the variants of Self

RemoteKeyValueStoreDataSchema

Enums

AttachedModuleId

KeyValueStoreDataSchema

Less flexible than previous revision, as mixed type origin is not allowed, but better for client-side optimization

ModuleId

Constants

ACTOR_REF_AUTH_ZONE

ACTOR_REF_GLOBAL

ACTOR_REF_OUTER

ACTOR_REF_SELF

ACTOR_STATE_OUTER_OBJECT

ACTOR_STATE_SELF

KV_STORE_DATA_SCHEMA_VARIANT_LOCAL

KV_STORE_DATA_SCHEMA_VARIANT_REMOTE

Traits

ClientActorApi

Api which exposes methods in the context of the actor

ClientActorIndexApi

Api to manage an iterable index

ClientActorKeyValueEntryApi

ClientActorSortedIndexApi

ClientApi

Interface of the system, for blueprints and Node modules.

ClientBlueprintApi

ClientCostingApi

ClientCryptoUtilsApi

ClientExecutionTraceApi

ClientFieldApi

A high level api to read/write fields

ClientKeyValueEntryApi

ClientKeyValueStoreApi

ClientObjectApi

A high level interface to manipulate objects in the actor’s call frame

ClientTransactionRuntimeApi

FieldPayloadMarker

KeyValueEntryPayloadMarker

KeyValueKeyPayloadMarker

SortedIndexEntryPayloadMarker

SortedIndexKeyPayloadMarker

Type Aliases

ActorRefHandle

ActorStateHandle

CollectionIndex

FieldHandle

FieldIndex

KeyValueEntryHandle

ObjectModuleId

Notes: This is to be deprecated, please use ModuleId instead