Trait scrypto::api::ClientActorIndexApi

pub trait ClientActorIndexApi<E> {
    // Required methods
    fn actor_index_insert(
        &mut self,
        object_handle: u32,
        collection_index: u8,
        key: Vec<u8, Global>,
        buffer: Vec<u8, Global>
    ) -> Result<(), E>;
    fn actor_index_remove(
        &mut self,
        object_handle: u32,
        collection_index: u8,
        key: Vec<u8, Global>
    ) -> Result<Option<Vec<u8, Global>>, E>;
    fn actor_index_scan_keys(
        &mut self,
        object_handle: u32,
        collection_index: u8,
        limit: u32
    ) -> Result<Vec<Vec<u8, Global>, Global>, E>;
    fn actor_index_drain(
        &mut self,
        object_handle: u32,
        collection_index: u8,
        limit: u32
    ) -> Result<Vec<(Vec<u8, Global>, Vec<u8, Global>), Global>, E>;

    // Provided methods
    fn actor_index_insert_typed<K, V>(
        &mut self,
        object_handle: u32,
        collection_index: u8,
        key: K,
        value: V
    ) -> Result<(), E>
       where K: ScryptoEncode,
             V: ScryptoEncode { ... }
    fn actor_index_remove_typed<V>(
        &mut self,
        object_handle: u32,
        collection_index: u8,
        key: Vec<u8, Global>
    ) -> Result<Option<V>, E>
       where V: ScryptoDecode { ... }
    fn actor_index_scan_keys_typed<K>(
        &mut self,
        object_handle: u32,
        collection_index: u8,
        limit: u32
    ) -> Result<Vec<K, Global>, E>
       where K: ScryptoDecode { ... }
    fn actor_index_drain_typed<K, V>(
        &mut self,
        object_handle: u32,
        collection_index: u8,
        limit: u32
    ) -> Result<Vec<(K, V), Global>, E>
       where K: ScryptoDecode,
             V: ScryptoDecode { ... }
}

Api to manage an iterable index

Required Methods

fn actor_index_insert( &mut self, object_handle: u32, collection_index: u8, key: Vec<u8, Global>, buffer: Vec<u8, Global> ) -> Result<(), E>

Inserts an entry into an index

fn actor_index_remove( &mut self, object_handle: u32, collection_index: u8, key: Vec<u8, Global> ) -> Result<Option<Vec<u8, Global>>, E>

Removes an entry from an index

fn actor_index_scan_keys( &mut self, object_handle: u32, collection_index: u8, limit: u32 ) -> Result<Vec<Vec<u8, Global>, Global>, E>

Scans arbitrary elements of count from an index

fn actor_index_drain( &mut self, object_handle: u32, collection_index: u8, limit: u32 ) -> Result<Vec<(Vec<u8, Global>, Vec<u8, Global>), Global>, E>

Removes and returns arbitrary elements of count from an index

Provided Methods

fn actor_index_insert_typed<K, V>( &mut self, object_handle: u32, collection_index: u8, key: K, value: V ) -> Result<(), E>where K: ScryptoEncode, V: ScryptoEncode,

Inserts an entry into an index

fn actor_index_remove_typed<V>( &mut self, object_handle: u32, collection_index: u8, key: Vec<u8, Global> ) -> Result<Option<V>, E>where V: ScryptoDecode,

Removes an entry from an index

fn actor_index_scan_keys_typed<K>( &mut self, object_handle: u32, collection_index: u8, limit: u32 ) -> Result<Vec<K, Global>, E>where K: ScryptoDecode,

Scans arbitrary elements of count from an index

fn actor_index_drain_typed<K, V>( &mut self, object_handle: u32, collection_index: u8, limit: u32 ) -> Result<Vec<(K, V), Global>, E>where K: ScryptoDecode, V: ScryptoDecode,

Removes and returns arbitrary elements of count from an index