Module scrypto::blueprints::resource

Structs

AuthZoneAssertAccessRuleInput

AuthZoneCreateProofOfAllInput

AuthZoneCreateProofOfAmountInput

AuthZoneCreateProofOfNonFungiblesInput

AuthZoneDrainInput

AuthZoneDropProofsInput

AuthZoneDropRegularProofsInput

AuthZoneDropSignatureProofsInput

AuthZonePopInput

AuthZonePushInput

AuthZoneRef

Bucket

BucketCreateProofOfAllInput

BucketGetAmountInput

BucketGetNonFungibleLocalIdsInput

BucketGetResourceAddressInput

BucketPutInput

BucketTakeAdvancedInput

BucketTakeInput

BucketTakeNonFungiblesInput

BurnRoles

DepositRoles

FreezeRoles

FungibleBucket

FungibleBucketCreateProofOfAmountInput

FungibleBucketLockAmountInput

FungibleBucketUnlockAmountInput

FungibleProof

FungibleResourceManagerCreateInput

FungibleResourceManagerCreateManifestInput

FungibleResourceManagerCreateWithInitialSupplyInput

FungibleResourceManagerCreateWithInitialSupplyManifestInput

FungibleResourceManagerMintInput

FungibleResourceRoles

FungibleVault

FungibleVaultCreateProofOfAmountInput

FungibleVaultLockFeeInput

FungibleVaultLockFungibleAmountInput

FungibleVaultUnlockFungibleAmountInput

LiquidFungibleResource

LiquidNonFungibleResource

LiquidNonFungibleVault

LocalNonFungibleDataSchema

LockedFungibleResource

LockedNonFungibleResource

MethodKey

MintRoles

ModuleRoleKey

NonFungibleBucket

NonFungibleBucketContainsNonFungibleInput

NonFungibleBucketCreateProofOfNonFungiblesInput

NonFungibleBucketLockNonFungiblesInput

NonFungibleBucketUnlockNonFungiblesInput

NonFungibleDataUpdateRoles

NonFungibleGlobalId

Represents the global id of a non-fungible.

NonFungibleProof

NonFungibleProofGetLocalIdsInput

NonFungibleResourceManagerCreateGenericInput

NonFungibleResourceManagerCreateInput

NonFungibleResourceManagerCreateManifestInput

NonFungibleResourceManagerCreateRuidWithInitialSupplyGenericInput

For typed value, to skip any codec

NonFungibleResourceManagerCreateRuidWithInitialSupplyInput

NonFungibleResourceManagerCreateRuidWithInitialSupplyManifestInput

For manifest

NonFungibleResourceManagerCreateWithInitialSupplyGenericInput

For typed value, to skip any codec

NonFungibleResourceManagerCreateWithInitialSupplyInput

NonFungibleResourceManagerCreateWithInitialSupplyManifestInput

For manifest

NonFungibleResourceManagerExistsInput

NonFungibleResourceManagerGetNonFungibleInput

NonFungibleResourceManagerMintGenericInput

For typed value, to skip any codec

NonFungibleResourceManagerMintInput

NonFungibleResourceManagerMintManifestInput

For manifest

NonFungibleResourceManagerMintRuidGenericInput

For typed value, to skip any codec

NonFungibleResourceManagerMintRuidInput

NonFungibleResourceManagerMintRuidManifestInput

For manifest

NonFungibleResourceManagerMintSingleRuidGenericInput

For typed value, to skip any codec

NonFungibleResourceManagerMintSingleRuidInput

NonFungibleResourceManagerUpdateDataGenericInput

For typed value, to skip any codec

NonFungibleResourceManagerUpdateDataInput

NonFungibleResourceManagerUpdateDataManifestInput

For manifest

NonFungibleResourceRoles

NonFungibleVault

NonFungibleVaultBurnNonFungiblesInput

NonFungibleVaultContainsNonFungibleInput

NonFungibleVaultCreateProofOfNonFungiblesInput

NonFungibleVaultGetNonFungibleLocalIdsInput

NonFungibleVaultLockNonFungiblesInput

NonFungibleVaultRecallNonFungiblesInput

NonFungibleVaultTakeNonFungiblesInput

NonFungibleVaultUnlockNonFungiblesInput

OwnedWorktop

OwnerRoleEntry

Proof

ProofCloneInput

ProofDropInput

ProofGetAmountInput

ProofGetResourceAddressInput

RecallRoles

RemoteNonFungibleDataSchema

ResourceManagerBurnInput

ResourceManagerCreateEmptyBucketInput

ResourceManagerCreateEmptyVaultInput

ResourceManagerDropEmptyBucketInput

ResourceManagerGetAmountForWithdrawalInput

ResourceManagerGetResourceTypeInput

ResourceManagerGetTotalSupplyInput

ResourceManagerPackageBurnInput

ResourceOrNonFungibleList

RoleAssignmentInit

RoleKey

RoleList

Vault

VaultBurnInput

VaultFreezeFlags

VaultFreezeInput

VaultFrozenFlag

VaultGetAmountInput

VaultPutInput

VaultRecallInput

VaultTakeAdvancedInput

VaultTakeInput

VaultUnfreezeInput

WithdrawRoles

WorktopAssertContainsAmountInput

WorktopAssertContainsInput

WorktopAssertContainsNonFungiblesInput

WorktopDrainInput

WorktopDropInput

WorktopPutInput

WorktopTakeAllInput

WorktopTakeInput

WorktopTakeNonFungiblesInput

Enums

AccessRule

AccessRuleNode

GlobalCaller

MethodAccessibility

NonFungibleDataSchema

OwnerRole

Front end data structure for specifying owner role

OwnerRoleUpdater

ParseNonFungibleGlobalIdError

Represents an error when parsing non-fungible address.

ProofRule

Resource Proof Rules

ResourceError

ResourceFeature

ResourceOrNonFungible

ResourceType

Represents the type of a resource.

WithdrawStrategy

Define the withdraw strategy when request amount does not match underlying resource divisibility.

Constants

AUTH_ZONE_ASSERT_ACCESS_RULE_IDENT

AUTH_ZONE_BLUEPRINT

AUTH_ZONE_CREATE_PROOF_OF_ALL_IDENT

AUTH_ZONE_CREATE_PROOF_OF_AMOUNT_IDENT

AUTH_ZONE_CREATE_PROOF_OF_NON_FUNGIBLES_IDENT

AUTH_ZONE_DRAIN_IDENT

AUTH_ZONE_DROP_PROOFS_IDENT

AUTH_ZONE_DROP_REGULAR_PROOFS_IDENT

AUTH_ZONE_DROP_SIGNATURE_PROOFS_IDENT

AUTH_ZONE_POP_IDENT

AUTH_ZONE_PUSH_IDENT

BUCKET_CREATE_PROOF_OF_ALL_IDENT

BUCKET_GET_AMOUNT_IDENT

BUCKET_GET_RESOURCE_ADDRESS_IDENT

BUCKET_PUT_IDENT

BUCKET_TAKE_ADVANCED_IDENT

BUCKET_TAKE_IDENT

BURNER_ROLE

BURNER_UPDATER_ROLE

DEPOSITOR_ROLE

DEPOSITOR_UPDATER_ROLE

FREEZER_ROLE

FREEZER_UPDATER_ROLE

FUNGIBLE_BUCKET_BLUEPRINT

FUNGIBLE_BUCKET_CREATE_PROOF_OF_AMOUNT_IDENT

FUNGIBLE_BUCKET_LOCK_AMOUNT_IDENT

FUNGIBLE_BUCKET_UNLOCK_AMOUNT_IDENT

FUNGIBLE_PROOF_BLUEPRINT

FUNGIBLE_RESOURCE_MANAGER_BLUEPRINT

FUNGIBLE_RESOURCE_MANAGER_CREATE_IDENT

FUNGIBLE_RESOURCE_MANAGER_CREATE_WITH_INITIAL_SUPPLY_IDENT

FUNGIBLE_RESOURCE_MANAGER_MINT_IDENT

FUNGIBLE_VAULT_BLUEPRINT

FUNGIBLE_VAULT_CREATE_PROOF_OF_AMOUNT_IDENT

FUNGIBLE_VAULT_LOCK_FEE_IDENT

FUNGIBLE_VAULT_LOCK_FUNGIBLE_AMOUNT_IDENT

FUNGIBLE_VAULT_UNLOCK_FUNGIBLE_AMOUNT_IDENT

MINTER_ROLE

MINTER_UPDATER_ROLE

NON_FUNGIBLE_BUCKET_BLUEPRINT

NON_FUNGIBLE_BUCKET_CONTAINS_NON_FUNGIBLE_IDENT

NON_FUNGIBLE_BUCKET_CREATE_PROOF_OF_NON_FUNGIBLES_IDENT

NON_FUNGIBLE_BUCKET_GET_NON_FUNGIBLE_LOCAL_IDS_IDENT

NON_FUNGIBLE_BUCKET_LOCK_NON_FUNGIBLES_IDENT

NON_FUNGIBLE_BUCKET_TAKE_NON_FUNGIBLES_IDENT

NON_FUNGIBLE_BUCKET_UNLOCK_NON_FUNGIBLES_IDENT

NON_FUNGIBLE_DATA_SCHEMA_VARIANT_LOCAL

NON_FUNGIBLE_DATA_SCHEMA_VARIANT_REMOTE

NON_FUNGIBLE_DATA_UPDATER_ROLE

NON_FUNGIBLE_DATA_UPDATER_UPDATER_ROLE

NON_FUNGIBLE_PROOF_BLUEPRINT

NON_FUNGIBLE_PROOF_GET_LOCAL_IDS_IDENT

NON_FUNGIBLE_RESOURCE_MANAGER_BLUEPRINT

NON_FUNGIBLE_RESOURCE_MANAGER_CREATE_IDENT

NON_FUNGIBLE_RESOURCE_MANAGER_CREATE_RUID_WITH_INITIAL_SUPPLY_IDENT

NON_FUNGIBLE_RESOURCE_MANAGER_CREATE_WITH_INITIAL_SUPPLY_IDENT

NON_FUNGIBLE_RESOURCE_MANAGER_EXISTS_IDENT

NON_FUNGIBLE_RESOURCE_MANAGER_GET_NON_FUNGIBLE_IDENT

NON_FUNGIBLE_RESOURCE_MANAGER_MINT_IDENT

NON_FUNGIBLE_RESOURCE_MANAGER_MINT_RUID_IDENT

NON_FUNGIBLE_RESOURCE_MANAGER_MINT_SINGLE_RUID_IDENT

NON_FUNGIBLE_RESOURCE_MANAGER_UPDATE_DATA_IDENT

NON_FUNGIBLE_VAULT_BLUEPRINT

NON_FUNGIBLE_VAULT_BURN_NON_FUNGIBLES_IDENT

NON_FUNGIBLE_VAULT_CONTAINS_NON_FUNGIBLE_IDENT

NON_FUNGIBLE_VAULT_CREATE_PROOF_OF_NON_FUNGIBLES_IDENT

NON_FUNGIBLE_VAULT_GET_NON_FUNGIBLE_LOCAL_IDS_IDENT

NON_FUNGIBLE_VAULT_LOCK_NON_FUNGIBLES_IDENT

NON_FUNGIBLE_VAULT_RECALL_NON_FUNGIBLES_IDENT

NON_FUNGIBLE_VAULT_TAKE_NON_FUNGIBLES_IDENT

NON_FUNGIBLE_VAULT_UNLOCK_NON_FUNGIBLES_IDENT

OWNER_ROLE

PROOF_CLONE_IDENT

PROOF_DROP_IDENT

PROOF_GET_AMOUNT_IDENT

PROOF_GET_RESOURCE_ADDRESS_IDENT

RECALLER_ROLE

RECALLER_UPDATER_ROLE

RESOURCE_MANAGER_BURN_IDENT

RESOURCE_MANAGER_CREATE_EMPTY_BUCKET_IDENT

RESOURCE_MANAGER_CREATE_EMPTY_VAULT_IDENT

RESOURCE_MANAGER_DROP_EMPTY_BUCKET_IDENT

RESOURCE_MANAGER_GET_AMOUNT_FOR_WITHDRAWAL_IDENT

RESOURCE_MANAGER_GET_RESOURCE_TYPE_IDENT

RESOURCE_MANAGER_GET_TOTAL_SUPPLY_IDENT

RESOURCE_MANAGER_PACKAGE_BURN_IDENT

SELF_ROLE

VAULT_BURN_IDENT

VAULT_FREEZE_IDENT

VAULT_GET_AMOUNT_IDENT

VAULT_PUT_IDENT

VAULT_RECALL_IDENT

VAULT_TAKE_ADVANCED_IDENT

VAULT_TAKE_IDENT

VAULT_UNFREEZE_IDENT

WITHDRAWER_ROLE

WITHDRAWER_UPDATER_ROLE

WORKTOP_ASSERT_CONTAINS_AMOUNT_IDENT

WORKTOP_ASSERT_CONTAINS_IDENT

WORKTOP_ASSERT_CONTAINS_NON_FUNGIBLES_IDENT

WORKTOP_BLUEPRINT

WORKTOP_DRAIN_IDENT

WORKTOP_DROP_IDENT

WORKTOP_PUT_IDENT

WORKTOP_TAKE_ALL_IDENT

WORKTOP_TAKE_IDENT

WORKTOP_TAKE_NON_FUNGIBLES_IDENT

Traits

AccessRuleVisitor

ForWithdrawal

FromPublicKey

Functions

check_fungible_amount

check_non_fungible_amount

global_caller

A requirement for the global ancestor of the actor who made the latest global call to either be:

package_of_direct_caller

A requirement for the immediate caller’s package to equal the given package.

require

require_all_of

require_amount

require_any_of

require_n_of

Type Aliases

AuthZoneAssertAccessRuleOutput

AuthZoneCreateProofOfAllOutput

AuthZoneCreateProofOfAmountOutput

AuthZoneCreateProofOfNonFungiblesOutput

AuthZoneDrainOutput

AuthZoneDropProofsOutput

AuthZoneDropRegularProofsOutput

AuthZoneDropSignatureProofsOutput

AuthZonePopOutput

AuthZonePushOutput

BucketCreateProofOfAllOutput

BucketGetAmountOutput

BucketGetNonFungibleLocalIdsOutput

BucketGetResourceAddressOutput

BucketPutOutput

BucketTakeAdvancedOutput

BucketTakeNonFungiblesOutput

BucketTakeOutput

FungibleBucketCreateProofOfAmountOutput

FungibleBucketLockAmountOutput

FungibleBucketUnlockAmountOutput

FungibleResourceManagerCreateOutput

FungibleResourceManagerCreateWithInitialSupplyOutput

FungibleResourceManagerMintOutput

FungibleVaultCreateProofOfAmountOutput

FungibleVaultLockFeeOutput

FungibleVaultLockFungibleAmountOutput

FungibleVaultUnlockFungibleAmountOutput

NonFungibleBucketContainsNonFungibleOutput

NonFungibleBucketCreateProofOfNonFungiblesOutput

NonFungibleBucketLockNonFungiblesOutput

NonFungibleBucketUnlockNonFungiblesOutput

NonFungibleProofGetLocalIdsOutput

NonFungibleResourceManagerCreateOutput

NonFungibleResourceManagerCreateRuidWithInitialSupplyOutput

NonFungibleResourceManagerCreateWithInitialSupplyOutput

NonFungibleResourceManagerExistsOutput

NonFungibleResourceManagerGetNonFungibleOutput

NonFungibleResourceManagerMintOutput

NonFungibleResourceManagerMintRuidOutput

NonFungibleResourceManagerMintSingleRuidOutput

NonFungibleResourceManagerUpdateDataOutput

NonFungibleVaultBurnNonFungiblesOutput

NonFungibleVaultContainsNonFungibleOutput

NonFungibleVaultCreateProofOfNonFungiblesOutput

NonFungibleVaultGetNonFungibleLocalIdsOutput

NonFungibleVaultLockNonFungiblesOutput

NonFungibleVaultRecallNonFungiblesOutput

NonFungibleVaultTakeNonFungiblesOutput

NonFungibleVaultUnlockNonFungiblesOutput

ProofCloneOutput

ProofDropOutput

ProofGetAmountOutput

ProofGetResourceAddressOutput

ResourceManagerBurnOutput

ResourceManagerCreateEmptyBucketOutput

ResourceManagerCreateEmptyVaultOutput

ResourceManagerDropEmptyBucketOutput

ResourceManagerGetAmountForWithdrawalOutput

ResourceManagerGetResourceTypeOutput

ResourceManagerGetTotalSupplyOutput

ResourceManagerPackageBurnOutput

VaultBurnOutput

VaultFreezeOutput

VaultGetAmountOutput

VaultPutOutput

VaultRecallOutput

VaultTakeAdvancedOutput

VaultTakeOutput

VaultUnfreezeOutput

WorktopAssertContainsAmountOutput

WorktopAssertContainsNonFungiblesOutput

WorktopAssertContainsOutput

WorktopDrainOutput

WorktopDropOutput

WorktopPutOutput

WorktopTakeAllOutput

WorktopTakeNonFungiblesOutput

WorktopTakeOutput