Module scrypto::buffer

Scrypto data encoding, decoding and exchange.

Constants

SCRYPTO_NAME_ACCOUNT

SCRYPTO_NAME_ADDRESS

SCRYPTO_NAME_BID

SCRYPTO_NAME_BIG_DECIMAL

SCRYPTO_NAME_BLUEPRINT

SCRYPTO_NAME_BUCKET

SCRYPTO_NAME_BUCKET_REF

SCRYPTO_NAME_COMPONENT

SCRYPTO_NAME_DECIMAL

SCRYPTO_NAME_H256

SCRYPTO_NAME_LAZY_MAP

SCRYPTO_NAME_MID

SCRYPTO_NAME_NON_FUNGIBLE_KEY

SCRYPTO_NAME_PACKAGE

SCRYPTO_NAME_RESOURCE_DEF

SCRYPTO_NAME_RID

SCRYPTO_NAME_VAULT

SCRYPTO_NAME_VID

SCRYPTO_TYPE_ADDRESS

SCRYPTO_TYPE_BID

SCRYPTO_TYPE_BIG_DECIMAL

SCRYPTO_TYPE_DECIMAL

SCRYPTO_TYPE_H256

SCRYPTO_TYPE_MID

SCRYPTO_TYPE_NON_FUNGIBLE_KEY

SCRYPTO_TYPE_RID

SCRYPTO_TYPE_VID

Functions

scrypto_alloc⚠

Allocates a chunk of memory that is not tracked by Rust ownership system.

scrypto_consume⚠

Consumes a memory chunk.

scrypto_decode

Decodes an instance of T from a slice.

scrypto_encode

Encodes a data structure into byte array.

scrypto_encode_for_radix_engine

Encodes a data structure into byte array for radix engine.

scrypto_free⚠

Releases an allocated memory.

scrypto_wrap

Wraps a byte array into a pointer.