Module scrypto::component

Scrypto component abstraction.

Re-exports

pub use component_access_rules::Mutability::*;

Structs

BorrowedPackage

Represents a published package.

Component

Represents an instantiated component.

ComponentAccessRules

ComponentInfoSubstate

ComponentStateSubstate

ComponentSystem

Represents the Radix Engine component subsystem.

GlobalComponentRef

KeyValueStore

A scalable key-value map which loads entries on demand.

KeyValueStoreEntrySubstate

Enums

Mutability

ParseComponentError

Represents an error when decoding key value store.

ParseKeyValueStoreError

Represents an error when decoding key value store.

Traits

ComponentState

Represents the state of a component.

GlobalComponent

A separate trait for standardized calls so that component methods don’t name clash TODO: unify with LocalComponent and use Own and GlobalRef Deref structures

LocalComponent

A separate trait for standardized calls so that component methods don’t name clash

Functions

component_system

Returns the component subsystem.

init_component_system

Initializes component subsystem.