Struct scrypto::component::component::Component

pub struct Component(pub ComponentId);

Represents an instantiated component.

Tuple Fields
0: ComponentId
Implementations

impl Component

pub fn call<T: ScryptoDecode>(&self, method: &str, args: Vec<u8>) -> T

Invokes a method on this component.

pub fn package_address(&self) -> PackageAddress

Returns the package ID of this component.

pub fn blueprint_name(&self) -> String

Returns the blueprint name of this component.

pub fn add_access_check(&mut self, access_rules: AccessRules) -> &mut Self

Add access check on the component.

pub fn set_royalty_config(&mut self, royalty_config: RoyaltyConfig) -> &mut Self

Set the royalty configuration of the component.

pub fn metadata<K: AsRef<str>, V: AsRef<str>>(
&mut self,
name: K,
value: V
) -> &mut Self

pub fn globalize(self) -> ComponentAddress

pub fn globalize_with_owner(
self,
owner_badge: NonFungibleAddress
) -> ComponentAddress

Globalize with owner badge. This will add additional access rules to protect native methods, such as metadata and royalty.

pub fn access_rules_chain(&self) -> Vec<ComponentAccessRules>

Returns the layers of access rules on this component.

impl Component

pub fn to_vec(&self) -> Vec<u8>