Struct scrypto::component::kv_store::KeyValueStore

pub struct KeyValueStore<K: ScryptoEncode + ScryptoDecode + ScryptoDescribe, V: ScryptoEncode + ScryptoDecode + ScryptoDescribe> {
    pub id: Own,
    pub key: PhantomData<K>,
    pub value: PhantomData<V>,
}

A scalable key-value map which loads entries on demand.

Fields
id: Ownkey: PhantomData<K>value: PhantomData<V>
Implementations

impl<K: ScryptoEncode + ScryptoDecode + ScryptoDescribe, V: ScryptoEncode + ScryptoDecode + ScryptoDescribe> KeyValueStore<K, V>

pub fn new() -> Self

Creates a new key value store.

pub fn get(&self, key: &K) -> Option<KeyValueEntryRef<'_, V>>

Returns the value that is associated with the given key.

pub fn get_mut(&mut self, key: &K) -> Option<KeyValueEntryRefMut<'_, V>>

pub fn insert(&self, key: K, value: V)

Inserts a new key-value pair into this map.

pub fn remove(&self, key: &K) -> Option<V>

Remove an entry from the map and return the original value if it exists