Module scrypto::engine::types

Types and functions shared by both Scrypto and Radix Engine.

Re-exports

pub use crate::constants::*;

Structs

ComponentAddress

An instance of a blueprint, which lives in the ledger state.

Decimal

Decimal represents a 128 bit representation of a fixed-scale decimal number.

EcdsaPrivateKey

Represents an ECDSA private key.

EcdsaPublicKey

Represents an ECDSA public key.

EcdsaSignature

Represents an ECDSA signature.

Hash

Represents a 32-byte hash digest.

NonFungibleAddress

Identifier for a non-fungible unit.

NonFungibleId

Represents a key for a non-fungible resource

PackageAddress

A collection of blueprints, compiled and published as a single unit.

ResourceAddress

Represents a resource address.

ScryptoActorInfo

Represents the running entity.

Enums

Level

Represents the level of a log message.

MintParams

Represents the minting parameters

ResourceType

Represents the type of a resource.

Type Definitions

BucketId

LazyMapId

ProofId

VaultId