Struct scrypto::math::integer::U8

#[repr(transparent)]
pub struct U8(pub u8);

Provides safe integer arithmetic.

Operations like +, ‘-’, ‘*’, or ‘/’ sometimes produce overflow which is detected and results in a panic, instead of silently wrapping around.

The bit length of output type will be the greater one in the math operation, and if any of the types was signed, then the resulting type will be signed too, otherwise the output type is unsigned.

The underlying value can be retrieved through the .0 index of the U8 tuple.

Layout

U8 will have the same methods and traits as the built-in counterpart.

Tuple Fields
0: u8
Implementations

impl U8

pub const MIN: Self = Self(u8::MIN)

Returns the smallest value that can be represented by this integer type.

Examples

Please note that this example is shared between large integer types. Which explains why I8 is used here.

Basic usage:

use scrypto::prelude::*;

assert_eq!(U8::MIN, < U8 > :: MIN );

pub const MAX: Self = Self(u8::MAX)

Returns the largest value that can be represented by this integer type.

Examples

Please note that this example is shared between large integer types. Which explains why I8 is used here.

Basic usage:

use scrypto::prelude::*;

assert_eq!(U8::MAX, < U8 > :: MAX );

pub const BITS: u32 = 8u32

Returns the size of this integer type in bits.

Examples

Basic usage:

use scrypto::prelude::*;

assert_eq!(U8::BITS, < U8 > :: BITS);

impl U8

pub const fn count_ones(self) -> u32

Returns the number of ones in the binary representation of self.

Examples

Basic usage:

use scrypto::prelude::*;

let n = U8::by(0b01001100u8);

assert_eq!(n.count_ones(), 3);

pub const fn count_zeros(self) -> u32

Returns the number of zeros in the binary representation of self.

Examples

Please note that this example is shared between large integer types. Which explains why I8 is used here.

Basic usage:

use scrypto::prelude::*;

assert_eq!(I8::by(0i8).count_zeros(), 8);

pub const fn trailing_zeros(self) -> u32

Returns the number of trailing zeros in the binary representation of self.

Examples

Please note that this example is shared between large integer types. Which explains why I8 is used here.

Basic usage:

use scrypto::prelude::*;

let n = I8::by(0b0101000u8);

assert_eq!(n.trailing_zeros(), 3);

pub const fn swap_bytes(self) -> Self

Reverses the byte order of the integer.

Examples

Please note that this example is shared between large integer types. Which explains why I16 is used here.

Basic usage:

use scrypto::prelude::*;

let n: I16  = I16::by(0b0000000_01010101i16);
assert_eq!(n, I16::by(85));

let m = n.swap_bytes();

assert_eq!(m, I16::by(0b01010101_00000000i16));
assert_eq!(m, I16::by(21760i16));

pub const fn reverse_bits(self) -> Self

Reverses the bit pattern of the integer.

Examples

Please note that this example is shared between large integer types. Which explains why I16 is used here.

Basic usage:

use scrypto::prelude::*;

let n: I16  = I16::by(0b0000000_01010101i16);
assert_eq!(n, I16::by(85i16));

let m = n.reverse_bits();

assert_eq!(m.0 as u16, 0b10101010_00000000);
assert_eq!(m, I16::by(-22016i16));

pub const fn from_be(x: Self) -> Self

Converts an integer from big endian to the target’s endianness.

On big endian this is a no-op. On little endian the bytes are swapped.

Examples

Please note that this example is shared between large integer types. Which explains why I8 is used here.

Basic usage:

use scrypto::prelude::*;

let n = U8::by(0x1Ai8);

if cfg!(target_endian = "big") {
   assert_eq!(U8::from_be(n), n)
} else {
   assert_eq!(U8::from_be(n), n.swap_bytes())
}

pub const fn from_le(x: Self) -> Self

Converts an integer from little endian to the target’s endianness.

On little endian this is a no-op. On big endian the bytes are swapped.

Examples

Please note that this example is shared between large integer types. Which explains why I8 is used here.

Basic usage:

use scrypto::prelude::*;

let n = U8::by(0x1Ai8);

if cfg!(target_endian = "little") {
   assert_eq!(U8::from_le(n), n)
} else {
   assert_eq!(U8::from_le(n), n.swap_bytes())
}

pub const fn to_be(self) -> Self

Converts self to big endian from the target’s endianness.

On big endian this is a no-op. On little endian the bytes are swapped.

Examples

Please note that this example is shared between large integer types. Which explains why I8 is used here.

Basic usage:

use scrypto::prelude::*;

let n = U8::by(0x1Ai8);

if cfg!(target_endian = "big") {
    assert_eq!(n.to_be(), n)
} else {
    assert_eq!(n.to_be(), n.swap_bytes())
}

pub const fn to_le(self) -> Self

Converts self to little endian from the target’s endianness.

On little endian this is a no-op. On big endian the bytes are swapped.

Examples

Please note that this example is shared between large integer types. Which explains why I8 is used here.

Basic usage:

use scrypto::prelude::*;

let n = U8::by(0x1Ai8);

if cfg!(target_endian = "little") {
    assert_eq!(n.to_le(), n)
} else {
    assert_eq!(n.to_le(), n.swap_bytes())
}

pub fn leading_zeros(self) -> u32

Returns the number of leading zeros in the binary representation of self.

impl U8

pub fn from_le_bytes(val: [u8; 1]) -> Self

impl U8

pub fn to_le_bytes(&self) -> [u8; 1]

pub fn to_vec(&self) -> Vec<u8>

impl U8

pub fn is_power_of_two(self) -> bool

Returns true if and only if self == 2^k for some k.

pub fn next_power_of_two(self) -> Self

Returns the smallest power of two greater than or equal to self.

When return value overflows (i.e., self > (1 << (N-1)) for type uN), overflows to 2^N = 0.