Struct scrypto::math::precise_decimal::PreciseDecimal

pub struct PreciseDecimal(pub I512);

PreciseDecimal represents a 512 bit representation of a fixed-scale decimal number.

The finite set of values are of the form m / 10^64, where m is an integer such that -2^(512 - 1) <= m < 2^(512 - 1).

Unless otherwise specified, all operations will panic if underflow/overflow.

Tuple Fields
0: I512
Implementations

impl PreciseDecimal

pub const MIN: Self = _

The min value of PreciseDecimal.

pub const MAX: Self = _

The max value of PreciseDecimal.

pub const BITS: usize = 512usize

The bit length of number storing PreciseDecimal.

pub const SCALE: u32 = 64u32

The fixed scale used by PreciseDecimal.

pub const ZERO: Self = _

pub const ONE: Self = _

pub fn zero() -> Self

Returns PreciseDecimal of 0.

pub fn one() -> Self

Returns PreciseDecimal of 1.

pub fn is_zero(&self) -> bool

Whether this decimal is zero.

pub fn is_positive(&self) -> bool

Whether this decimal is positive.

pub fn is_negative(&self) -> bool

Whether this decimal is negative.

pub fn abs(&self) -> PreciseDecimal

Returns the absolute value.

pub fn floor(&self) -> Self

Returns the largest integer that is equal to or less than this number.

pub fn ceiling(&self) -> Self

Returns the smallest integer that is equal to or greater than this number.

pub fn round(&self, decimal_places: u32, mode: RoundingMode) -> Self

pub fn powi(&self, exp: i64) -> Self

Calculates power usingexponentiation by squaring.

impl PreciseDecimal

pub fn to_vec(&self) -> Vec<u8>