Module scrypto::model

Scrypto RE node model.

Structs

AccessRules

AccessRulesAddAccessCheckInvocation

AccessRulesGetLengthInvocation

AccessRulesSetGroupAccessRuleInvocation

AccessRulesSetGroupMutabilityInvocation

AccessRulesSetMethodAccessRuleInvocation

AccessRulesSetMethodMutabilityInvocation

AuthZoneAssertAccessRuleInvocation

AuthZoneClearInvocation

AuthZoneCreateProofByAmountInvocation

AuthZoneCreateProofByIdsInvocation

AuthZoneCreateProofInvocation

AuthZoneDrainInvocation

AuthZonePopInvocation

AuthZonePushInvocation

Bucket

Represents a transient resource container.

BucketCreateProofInvocation

BucketGetAmountInvocation

BucketGetNonFungibleIdsInvocation

BucketGetResourceAddressInvocation

BucketPutInvocation

BucketTakeInvocation

BucketTakeNonFungiblesInvocation

ClockCompareCurrentTimeInvocation

ClockCreateInvocation

ClockGetCurrentTimeInvocation

ClockSetCurrentTimeInvocation

ComponentClaimRoyaltyInvocation

ComponentGlobalizeInvocation

ComponentGlobalizeWithOwnerInvocation

ComponentSetRoyaltyConfigInvocation

EpochManagerCreateInvocation

EpochManagerGetCurrentEpochInvocation

EpochManagerSetEpochInvocation

MetadataGetInvocation

MetadataSetInvocation

NonFungibleAddress

Identifier for a non-fungible unit.

PackageClaimRoyaltyExecutable

PackageClaimRoyaltyInvocation

PackagePublishInvocation

PackagePublishWithOwnerInvocation

PackageSetRoyaltyConfigExecutable

PackageSetRoyaltyConfigInvocation

Proof

Represents a proof of owning some resource.

ProofCloneInvocation

ProofGetAmountInvocation

ProofGetNonFungibleIdsInvocation

ProofGetResourceAddressInvocation

ResourceManagerBucketBurnInvocation

ResourceManagerBurnInvocation

ResourceManagerCreateBucketInvocation

ResourceManagerCreateInvocation

ResourceManagerCreateVaultInvocation

ResourceManagerCreateWithOwnerInvocation

ResourceManagerGetNonFungibleInvocation

ResourceManagerGetResourceTypeInvocation

ResourceManagerGetTotalSupplyInvocation

ResourceManagerMintInvocation

ResourceManagerNonFungibleExistsInvocation

ResourceManagerSetVaultAuthMutabilityInvocation

ResourceManagerUpdateNonFungibleDataInvocation

ResourceManagerUpdateVaultAuthInvocation

RoyaltyConfig

RoyaltyConfigBuilder

Vault

VaultCreateProofByAmountInvocation

VaultCreateProofByIdsInvocation

VaultCreateProofInvocation

VaultGetAmountInvocation

VaultGetNonFungibleIdsInvocation

VaultGetResourceAddressInvocation

VaultLockFeeInvocation

VaultPutInvocation

VaultRecallInvocation

VaultRecallNonFungiblesInvocation

VaultTakeInvocation

VaultTakeNonFungiblesInvocation

WorktopAssertContainsAmountInvocation

WorktopAssertContainsInvocation

WorktopAssertContainsNonFungiblesInvocation

WorktopDrainInvocation

WorktopPutInvocation

WorktopTakeAllInvocation

WorktopTakeAmountInvocation

WorktopTakeNonFungiblesInvocation

Enums

AccessRule

AccessRuleEntry

AccessRuleKey

AccessRuleNode

ComponentAddress

An instance of a blueprint, which lives in the ledger state.

MintParams

NonFungibleId

Represents a key for a non-fungible resource

NonFungibleIdType

Represents type of non-fungible id

PackageAddress

A collection of blueprints, compiled and published as a single unit.

ParseBucketError

Represents an error when decoding bucket.

ParseNonFungibleAddressError

Represents an error when parsing non-fungible address.

ParseNonFungibleIdError

Represents an error when decoding non-fungible id.

ParseNonFungibleIdTypeError

ParseProofError

Represents an error when decoding proof.

ParseVaultError

Represents an error when decoding vault.

ParsedScryptoInvocation

ProofRule

ProofValidationError

Represents an error when validating proof.

ProofValidationMode

Specifies the validation mode that should be used for validating a Proof.

ResourceAddress

Represents a resource address.

ResourceMethodAuthKey

ResourceType

Represents the type of a resource.

ScryptoInvocation

SoftCount

SoftDecimal

SoftResource

SoftResourceOrNonFungible

SoftResourceOrNonFungibleList

SystemAddress

TimePrecision

VaultMethodAuthKey

Constants

NON_FUNGIBLE_ID_MAX_LENGTH

Traits

FromPublicKey

NonFungibleData

Represents the data structure of a non-fungible.

Functions

require

require_all_of

require_amount

require_any_of

require_n_of