Module scrypto::prelude

Scrypto preludes.

Re-exports

pub use crate::engine::scrypto_env::ScryptoVmV1Api;

pub use radix_engine_interface::radix_engine_common;

pub use crate::component::*;

pub use crate::crypto_utils::*;

pub use crate::engine::*;

pub use crate::modules::*;

pub use crate::resource::*;

pub use crate::runtime::*;

Modules

actor_api

actor_index_api

actor_key_value_entry_api

actor_sorted_index_api

additional_attributes

Documentation for Additional Attributes

auth

blueprint_api

bnum_integer

Definitions of safe integers and uints.

decimal

field_api

key_value_entry_api

key_value_store_api

metadata

node_modules

object_api

precise_decimal

radix_engine_common

rounding_mode

royalty

system_modules

traits

vec

A contiguous growable array type with heap-allocated contents, written Vec<T>.

Macros

access_and_or

access_rule_node

btreemap

btreeset

burn_roles

component_royalties

component_royalty_config

create_well_known_lookup

debug

dec

Creates a Decimal from literals. It is a compile-time macro. It allows to declare constants and statics. Example: const D1: Decimal = dec!(“1111.11111”) const D2: Decimal = dec!(“-1111.11111”) const D3: Decimal = dec!(1) const D4: Decimal = dec!(-1_0000_000_u128)

define_single_versioned

This macro is intended for creating a data model which supports versioning. This is useful for creating an SBOR data model which can be updated in future. In future, enum variants can be added, and automatically mapped to.

define_versioned

This macro is intended for creating a data model which supports versioning. This is useful for creating an SBOR data model which can be updated in future. In future, enum variants can be added, and automatically mapped to.

define_wrapped_hash

deposit_roles

enable_function_auth

enable_method_auth

enable_package_royalties

error

extern_blueprint_internal

format

Creates a String using interpolation of runtime expressions.

freeze_roles

hashmap

hashset

i

Creates a safe integer from literals. You must specify the type of the integer you want to create.

include_code

Includes the WASM file of a Scrypto package.

include_schema

Includes the schema file of a Scrypto package.

indexmap

indexset

info

internal_add_role

internal_component_royalty_entry

internal_roles

look_up_type

main_accessibility

manifest_args

metadata

metadata_init

metadata_init_set_entry

metadata_roles

method_accessibilities

method_accessibility

mint_roles

non_fungible_data_update_roles

numeric_validation_match

pdec

Creates a PreciseDecimal from literals. It is a compile-time macro. It allows to declare constants and statics. Example: const D1: PreciseDecimal = pdec!(“1111.11111”) const D2: PreciseDecimal = pdec!(“-1111.11111”) const D3: PreciseDecimal = pdec!(1) const D4: PreciseDecimal = pdec!(-1_0000_000_u128)

recall_roles

return_if_error

return_type_mismatch_error

role_entry

role_list

roles

Roles macro for main module

roles2

rule

scrypto_args

terminal_value

terminal_value_from_body

this_package

to_manifest_value_and_unwrap

to_role_key

trace

vec

Creates a Vec containing the arguments.

warn

withdraw_roles

Structs

AddressBech32Decoder

Represents a decoder which understands how to decode Scrypto addresses in Bech32.

AddressBech32Encoder

Represents an encoder which understands how to encode Scrypto addresses in Bech32.

AddressDisplayContext

ArrayData

AttachedModuleIdIter

An iterator over the variants of Self

AuthAddresses

AuthZoneAssertAccessRuleInput

AuthZoneCreateProofOfAllInput

AuthZoneCreateProofOfAmountInput

AuthZoneCreateProofOfNonFungiblesInput

AuthZoneDrainInput

AuthZoneDropProofsInput

AuthZoneDropRegularProofsInput

AuthZoneDropSignatureProofsInput

AuthZonePopInput

AuthZonePushInput

AuthZoneRef

BTreeMap

An ordered map based on a B-Tree.

BTreeSet

An ordered set based on a B-Tree.

Bls12381G1PrivateKey

Bls12381G1PublicKey

Represents a BLS12-381 G1 public key.

Bls12381G2Signature

Represents a BLS12-381 G2 signature (variant with 96-byte signature and 48-byte public key)

BlueprintEventSchemaInit

BlueprintFunctionsSchemaInit

BlueprintHooksInit

BlueprintId

BlueprintInfo

Core object state, persisted in TypeInfoSubstate.

BlueprintKeyValueSchema

BlueprintSchemaInit

BlueprintStateSchemaInit

BlueprintTypeIdentifier

BlueprintTypeSchemaInit

BorrowError

An error returned by RefCell::try_borrow.

BorrowMutError

An error returned by RefCell::try_borrow_mut.

Box

A pointer type that uniquely owns a heap allocation of type T.

Bucket

BucketCreateProofOfAllInput

BucketGetAmountInput

BucketGetNonFungibleLocalIdsInput

BucketGetResourceAddressInput

BucketPutInput

BucketTakeAdvancedInput

BucketTakeInput

BucketTakeNonFungiblesInput

Buffer

BurnRoles

BytesNonFungibleLocalId

Bytes, of length between 1 and 64.

Cell

A mutable memory location.

CheckedOrigin

CheckedUrl

ComponentAddress

Address to a global component

ComponentRoyaltyConfig

Royalty rules

ContextDisplayable

ContextSerializable

Decimal

Decimal represents a 192 bit representation of a fixed-scale decimal number.

DepositRoles

Ed25519PrivateKey

Ed25519PublicKey

Represents an ED25519 public key.

Ed25519PublicKeyHash

Ed25519Signature

Represents an ED25519 signature.

EnumVariantData

EnumVariantValue

Epoch

A type-safe consensus epoch number.

EventFlags

EventTypeIdentifier

Identifies a specific event schema emitter by some emitter RENode.

FieldSchema

FieldValue

FixedEnumVariant

FnIdentifier

FreezeRoles

FullyScopedTypeId

An identifier for a type under a given node’s schema context in the Radix network.

FunctionSchemaInit

FungibleBucket

FungibleBucketCreateProofOfAmountInput

FungibleBucketLockAmountInput

FungibleBucketUnlockAmountInput

FungibleProof

FungibleResourceManagerCreateInput

FungibleResourceManagerCreateManifestInput

FungibleResourceManagerCreateWithInitialSupplyInput

FungibleResourceManagerCreateWithInitialSupplyManifestInput

FungibleResourceManagerMintInput

FungibleResourceRoles

FungibleVault

FungibleVaultCreateProofOfAmountInput

FungibleVaultLockFeeInput

FungibleVaultLockFungibleAmountInput

FungibleVaultUnlockFungibleAmountInput

GenericArgs

GlobalAddress

Address to a global entity

GlobalAddressPhantom

GlobalAddressReservation

Hash

Represents a 32-byte hash digest.

HashAccumulator

Represents a 32-byte hash accumulator.

HrpSet

Represents an HRP set (typically corresponds to a network).

I192

Provides safe integer arithmetic.

I256

Provides safe integer arithmetic.

I320

Provides safe integer arithmetic.

I384

Provides safe integer arithmetic.

I448

Provides safe integer arithmetic.

I512

Provides safe integer arithmetic.

I768

Provides safe integer arithmetic.

IndexedScryptoValue

Instant

Represents a Unix timestamp, capturing the seconds since the unix epoch.

IntegerNonFungibleLocalId

Unsigned integers, up to u64.

InternalAddress

Address to a local entity

KVEntry

KeyValueStoreGenericSubstitutions

KeyValueStoreInfo

KeyValueStoreInit

KeyValueStoreInitEntry

LengthValidation

Represents additional validation that should be performed on the size.

LinkedList

A doubly-linked list with owned nodes.

LiquidFungibleResource

LiquidNonFungibleResource

LiquidNonFungibleVault

LocalNonFungibleDataSchema

LocatedValidationError

LockedFungibleResource

LockedNonFungibleResource

MapData

MetadataCreateInput

MetadataCreateWithDataInput

MetadataGetInput

MetadataLockInput

MetadataRemoveInput

MetadataRoles

MetadataSetInput

MethodKey

MintRoles

ModuleConfig

ModuleRoleKey

NonFungibleBucket

NonFungibleBucketContainsNonFungibleInput

NonFungibleBucketCreateProofOfNonFungiblesInput

NonFungibleBucketLockNonFungiblesInput

NonFungibleBucketUnlockNonFungiblesInput

NonFungibleDataUpdateRoles

NonFungibleGlobalId

Represents the global id of a non-fungible.

NonFungibleProof

NonFungibleProofGetLocalIdsInput

NonFungibleResourceManagerCreateGenericInput

NonFungibleResourceManagerCreateInput

NonFungibleResourceManagerCreateManifestInput

NonFungibleResourceManagerCreateRuidWithInitialSupplyGenericInput

For typed value, to skip any codec

NonFungibleResourceManagerCreateRuidWithInitialSupplyInput

NonFungibleResourceManagerCreateRuidWithInitialSupplyManifestInput

For manifest

NonFungibleResourceManagerCreateWithInitialSupplyGenericInput

For typed value, to skip any codec

NonFungibleResourceManagerCreateWithInitialSupplyInput

NonFungibleResourceManagerCreateWithInitialSupplyManifestInput

For manifest

NonFungibleResourceManagerExistsInput

NonFungibleResourceManagerGetNonFungibleInput

NonFungibleResourceManagerMintGenericInput

For typed value, to skip any codec

NonFungibleResourceManagerMintInput

NonFungibleResourceManagerMintManifestInput

For manifest

NonFungibleResourceManagerMintRuidGenericInput

For typed value, to skip any codec

NonFungibleResourceManagerMintRuidInput

NonFungibleResourceManagerMintRuidManifestInput

For manifest

NonFungibleResourceManagerMintSingleRuidGenericInput

For typed value, to skip any codec

NonFungibleResourceManagerMintSingleRuidInput

NonFungibleResourceManagerUpdateDataGenericInput

For typed value, to skip any codec

NonFungibleResourceManagerUpdateDataInput

NonFungibleResourceManagerUpdateDataManifestInput

For manifest

NonFungibleResourceRoles

NonFungibleVault

NonFungibleVaultBurnNonFungiblesInput

NonFungibleVaultContainsNonFungibleInput

NonFungibleVaultCreateProofOfNonFungiblesInput

NonFungibleVaultGetNonFungibleLocalIdsInput

NonFungibleVaultLockNonFungiblesInput

NonFungibleVaultRecallNonFungiblesInput

NonFungibleVaultTakeNonFungiblesInput

NonFungibleVaultUnlockNonFungiblesInput

NovelTypeMetadata

This is the struct used in the Schema

NumericValidation

Represents additional validation that should be performed on the numeric value.

OnceCell

A cell which can be written to only once.

OwnedWorktop

OwnerRoleEntry

PhantomData

Zero-sized type used to mark things that “act like” they own a T.

PreciseDecimal

PreciseDecimal represents a 256 bit representation of a fixed-scale decimal number.

Proof

ProofCloneInput

ProofDropInput

ProofGetAmountInput

ProofGetResourceAddressInput

RawPayload

A wrapper for a full encoded SBOR payload, including the prefix byte.

RawValue

A wrapper for a reference to a valid partial SBOR payload representing a single value.

Rc

A single-threaded reference-counting pointer. ‘Rc’ stands for ‘Reference Counted’.

RecallRoles

Ref

Wraps a borrowed reference to a value in a RefCell box. A wrapper type for an immutably borrowed value from a RefCell<T>.

RefCell

A mutable memory location with dynamically checked borrow rules

RefMut

A wrapper type for a mutably borrowed value from a RefCell<T>.

RemoteNonFungibleDataSchema

ResourceManagerBurnInput

ResourceManagerCreateEmptyBucketInput

ResourceManagerCreateEmptyVaultInput

ResourceManagerDropEmptyBucketInput

ResourceManagerGetAmountForWithdrawalInput

ResourceManagerGetResourceTypeInput

ResourceManagerGetTotalSupplyInput

ResourceManagerPackageBurnInput

ResourceOrNonFungibleList

RoleAssignmentCreateInput

RoleAssignmentGetInput

RoleAssignmentInit

RoleAssignmentLockOwnerInput

RoleAssignmentSetInput

RoleAssignmentSetOwnerInput

RoleKey

RoleList

SborFixedEnumVariant

SborPath

A series of indexes which describes some value in the sbor tree

SborPathBuf

SborRawPayload

A wrapper for a full encoded SBOR payload, including the prefix byte.

SborRawValue

A wrapper for a reference to a valid partial SBOR payload representing a single value.

SchemaContext

SchemaV1

An array of custom type kinds, and associated extra information which can attach to the type kinds

Secp256k1PrivateKey

Secp256k1PublicKey

Represents an ECDSA Secp256k1 public key.

Secp256k1PublicKeyHash

Secp256k1Signature

Represents an ECDSA Secp256k1 signature.

String

A UTF-8–encoded, growable string.

TupleData

TypeData

Combines all data about a Type:

TypeMetadata

This enables the type to be represented as eg JSON Also used to facilitate type reconstruction

U192

Provides safe integer arithmetic.

U256

Provides safe integer arithmetic.

U320

Provides safe integer arithmetic.

U384

Provides safe integer arithmetic.

U448

Provides safe integer arithmetic.

U512

Provides safe integer arithmetic.

U768

Provides safe integer arithmetic.

UncheckedOrigin

UncheckedUrl

UnsafeCell

The core primitive for interior mutability in Rust.

Vault

VaultBurnInput

VaultFreezeFlags

VaultFreezeInput

VaultFrozenFlag

VaultGetAmountInput

VaultPutInput

VaultRecallInput

VaultTakeAdvancedInput

VaultTakeInput

VaultUnfreezeInput

Vec

A contiguous growable array type, written as Vec<T>, short for ‘vector’.

VecDecoder

A Decoder abstracts the logic for decoding basic types.

VecDeque

A double-ended queue implemented with a growable ring buffer.

LazyCellExperimental

A value which is initialized on the first access.

LocalKeyValueStoreDataSchema

LockFlags

ManifestAddressReservation

ManifestArgs

ManifestBlobRef

ManifestBucket

ManifestCustomTerminalValueRef

ManifestDecimal

ManifestPreciseDecimal

ManifestProof

ManifestValueDisplayContext

Note - this is quite similar to ManifestDecompilationDisplayContext

ModuleIdIter

An iterator over the variants of Self

NetworkDefinition

Network Definition is intended to be the actual definition of a network

NodeId

The unique identifier of a (stored) node.

NonIterMap

A thin wrapper around a HashMap, which guarantees that a HashMap usage will not result in a non-deterministic execution (simply by disallowing the iteration over its elements).

ObjectInfo

Own

PackageAddress

Address to a global package

PartitionNumber

The unique identifier of a node module.

PartitionOffset

RUIDNonFungibleLocalId

RUID, v4, variant 1, big endian. See https://www.rfc-editor.org/rfc/rfc4122

ReceiverInfo

RefTypes

Reference

RemoteKeyValueStoreDataSchema

ResourceAddress

Address to a global resource

Round

A type-safe consensus round number within a single epoch.

SchemaHash

Represents a particular schema under a package

ScopedTypeId

An identifier for a type in the context of a schema.

ScryptoCustomExtension

ScryptoCustomSchema

ScryptoCustomTerminalValueRef

ScryptoValueDisplayContext

Slice

StringNonFungibleLocalId

A string matching [_0-9a-zA-Z]{1,64}.

SyncUnsafeCellExperimental

UnsafeCell, but Sync.

TypeAggregator

UtcDateTime

A UtcDateTime represents a Unix timestamp on the UTC Calendar.

VecEncoder

An Encoder abstracts the logic for writing core types into a byte buffer.

WellKnownTypeId

WithdrawRoles

WorktopAssertContainsAmountInput

WorktopAssertContainsInput

WorktopAssertContainsNonFungiblesInput

WorktopDrainInput

WorktopDropInput

WorktopPutInput

WorktopTakeAllInput

WorktopTakeInput

WorktopTakeNonFungiblesInput

Enums

AccessControllerPartitionOffset

AccessRule

AccessRuleNode

AccountPartitionOffset

AddressBech32DecodeError

AddressBech32EncodeError

AttachedModuleId

AuthZoneField

BlueprintCollectionSchema

BlueprintHook

BootLoaderField

ChildNames

ClientCostingEntry

ComponentField

ComponentRoyaltyPartitionOffset

Condition

ConsensusManagerPartitionOffset

ContentValidationError

Cow

A clone-on-write smart pointer.

DateTimeError

DecodeError

Represents an error occurred during decoding.

Emitter

EncodeError

Represents an error occurred during encoding.

EntityType

An enum which represents the different addressable entities.

FallToOwner

FieldTransience

FungibleBucketField

FungibleProofField

FungibleResourceManagerPartitionOffset

FungibleVaultPartitionOffset

GenericBound

GenericMetadataValue

GenericSubstitution

A reference to the type to substitute with for the case of generics.

GlobalCaller

KeyValueStoreDataSchema

Less flexible than previous revision, as mixed type origin is not allowed, but better for client-side optimization

Level

Represents the level of a log message.

LocalTypeId

The TypeId which is local to a given Schema. This is the SchemaTypeLink used in a linearized Schema to link TypeKinds.

ManifestAddress

Any address supported by manifest, both global and local.

ManifestCustomExtension

ManifestCustomTraversal

ManifestCustomValue

ManifestCustomValueKind

ManifestExpression

ManifestNonFungibleLocalId

ManifestNonFungibleLocalIdValidationError

ManifestToRustValueError

MetadataConversionError

MetadataPartitionOffset

MethodAccessibility

ModuleId

MultiResourcePoolPartitionOffset

NoCustomExtension

NoCustomSchema

NoCustomTerminalValueRef

NoCustomTraversal

NoCustomTypeKind

NoCustomTypeValidation

NoCustomValue

NoCustomValueKind

NonFungibleBucketField

NonFungibleDataSchema

NonFungibleIdType

Represents type of non-fungible id

NonFungibleLocalId

Represents the local id of a non-fungible.

NonFungibleProofField

NonFungibleResourceManagerPartitionOffset

NonFungibleVaultPartitionOffset

ObjectType

OneResourcePoolPartitionOffset

Option

The Option type. See the module level documentation for more.

OuterObjectInfo

OwnValidation

OwnerRole

Front end data structure for specifying owner role

OwnerRoleUpdater

PackageCode

PackagePartitionOffset

PackageRoyalty

Royalty rules

PackageRoyaltyConfig

Royalty rules

ParseBlsPublicKeyError

Represents an error when retrieving BLS public key from hex or when aggregating.

ParseBlsSignatureError

Represents an error when retrieving BLS signature from hex or when aggregating.

ParseComponentAddressError

ParseDecimalError

Represents an error when parsing Decimal from another type.

ParseEd25519PublicKeyError

Represents an error when parsing ED25519 public key from hex.

ParseEd25519SignatureError

ParseError

The ParseError enum is a collection of all the possible reasons an enum can fail to parse from a string.

ParseGlobalAddressError

ParseHashError

Represents an error when parsing hash.

ParseI192Error

ParseI256Error

ParseI320Error

ParseI384Error

ParseI448Error

ParseI512Error

ParseI768Error

ParseInternalAddressError

ParseManifestAddressReservationError

Represents an error when parsing ManifestAddressReservation.

ParseManifestBlobRefError

Represents an error when parsing ManifestBlobRef.

ParseManifestBucketError

Represents an error when parsing ManifestBucket.

ParseManifestDecimalError

Represents an error when parsing ManifestDecimal.

ParseManifestExpressionError

Represents an error when parsing ManifestExpression.

ParseManifestPreciseDecimalError

Represents an error when parsing ManifestPreciseDecimal.

ParseManifestProofError

Represents an error when parsing ManifestProof.

ParseNetworkError

ParseNonFungibleGlobalIdError

Represents an error when parsing non-fungible address.

ParseNonFungibleLocalIdError

Represents an error when decoding non-fungible id.

ParseOwnError

ParsePackageAddressError

ParsePreciseDecimalError

Represents an error when parsing PreciseDecimal from another type.

ParseReferenceError

ParseResourceAddressError

ParseSecp256k1PublicKeyError

Represents an error when parsing ED25519 public key from hex.

ParseSecp256k1SignatureError

ParseU192Error

ParseU256Error

ParseU320Error

ParseU384Error

ParseU448Error

ParseU512Error

ParseU768Error

ParseUtcDateTimeError

PayloadValidationError

ProofRule

Resource Proof Rules

PublicKey

Represents any natively supported public key.

PublicKeyHash

The hash of a given public key.

Receiver

ReferenceValidation

ResourceError

ResourceFeature

ResourceOrNonFungible

ResourceType

Represents the type of a resource.

Result

Result is a type that represents either success (Ok) or failure (Err).

RoleAssignmentPartitionOffset

RoundingMode

Defines the rounding strategy.

RoyaltyAmount

RoyaltyField

RustToManifestValueError

RustTypeId

This is a compile-time identifier for a given type, used by the type aggregator to uniquely identify a type.

SborValue

Represents any value conforming to the SBOR value model.

SchemaValidationError

ScryptoCustomTraversal

ScryptoCustomTypeKind

A schema for the values that a codec can decode / views as valid

ScryptoCustomTypeValidation

ScryptoCustomValue

ScryptoCustomValueKind

SubstateKey

The unique identifier of a substate within a node module.

TimeComparisonOperator

TransactionTrackerField

TwoResourcePoolPartitionOffset

TypeInfoField

TypeKind

A schema for the values that a codec can decode / views as valid

TypeRef

TypeValidation

Additional validation to apply to a payload of the given type, beyond validation from the TypeKind’s type structure.

UpdateResult

ValidationError

ValidatorPartitionOffset

Value

Represents any value conforming to the SBOR value model.

ValueKind

VersionedSchema

WithdrawStrategy

Define the withdraw strategy when request amount does not match underlying resource divisibility.

WorktopField

Constants

ACCESS_CONTROLLER_PACKAGE

The native package for access controllers.

ACCESS_RULE_NODE_LIST_TYPE

ACCESS_RULE_NODE_TYPE

ACCESS_RULE_TYPE

ACCOUNT_OWNER_BADGE

The non-fungible badge resource which is used for account ownership, if accounts have been set up with simple account creation, or have been securified.

ACCOUNT_PACKAGE

The native package for accounts.

ACTOR_REF_AUTH_ZONE

ACTOR_REF_GLOBAL

ACTOR_REF_OUTER

ACTOR_REF_SELF

ACTOR_STATE_OUTER_OBJECT

ACTOR_STATE_SELF

ARCHIVE_STORAGE_PRICE_IN_XRD

The price for adding a single byte to archive storage, in XRD. 1 MB = 6 USD This is primarily for transaction payload, events and logs.

ATTACHED_MODULE_ID_TYPE

AUTH_ZONE_ASSERT_ACCESS_RULE_IDENT

AUTH_ZONE_BLUEPRINT

AUTH_ZONE_CREATE_PROOF_OF_ALL_IDENT

AUTH_ZONE_CREATE_PROOF_OF_AMOUNT_IDENT

AUTH_ZONE_CREATE_PROOF_OF_NON_FUNGIBLES_IDENT

AUTH_ZONE_DRAIN_IDENT

AUTH_ZONE_DROP_PROOFS_IDENT

AUTH_ZONE_DROP_REGULAR_PROOFS_IDENT

AUTH_ZONE_DROP_SIGNATURE_PROOFS_IDENT

AUTH_ZONE_POP_IDENT

AUTH_ZONE_PUSH_IDENT

BASIC_SBOR_V1_MAX_DEPTH

BASIC_SBOR_V1_PAYLOAD_PREFIX

BLS12381_CIPHERSITE_V1

BLS12-381 ciphersuite v1 It has following parameters

BLUEPRINT_PAYLOAD_MAX_DEPTH

Depth limit for various types of blueprint payload

BOOT_LOADER_PARTITION

Used only with TRANSACTION_TRACKER Node for boot loading

BOOT_LOADER_RESERVED_NODE_ID_FIRST_BYTE

BUCKET_CREATE_PROOF_OF_ALL_IDENT

BUCKET_GET_AMOUNT_IDENT

BUCKET_GET_RESOURCE_ADDRESS_IDENT

BUCKET_PUT_IDENT

BUCKET_TAKE_ADVANCED_IDENT

BUCKET_TAKE_IDENT

BURNER_ROLE

BURNER_UPDATER_ROLE

COMPONENT_ADDRESS_TYPE

CONSENSUS_MANAGER

The consensus manager native component - in charge of validators, consensus and epochs.

CONSENSUS_MANAGER_PACKAGE

The native package for the consensus manager.

CUSTOM_VALUE_KIND_START

A custom type is an application defined type with special semantics.

DECIMAL_SIZE

DECIMAL_TYPE

DEFAULT_TIP_PERCENTAGE

The default tip percentage to use when building a transaction.

DEPOSITOR_ROLE

DEPOSITOR_UPDATER_ROLE

ED25519_PUBLIC_KEY_HASH_TYPE

ED25519_PUBLIC_KEY_TYPE

ED25519_SIGNATURE_VIRTUAL_BADGE

The non-fungible badge resource which is used for virtual proofs of EdDSA Ed25519 transacton signatures in the transaction processor.

EMPTY

EXECUTION_COST_UNIT_LIMIT

The execution cost unit limit.

EXECUTION_COST_UNIT_LOAN

The execution cost units loaned from system

EXECUTION_COST_UNIT_PRICE_IN_XRD

The price of execution cost unit, in XRD.

FAUCET

The faucet native component - use this on testnets for getting XRD and locking fee.

FAUCET_BLUEPRINT

The name of the faucet blueprint under the FAUCET_PACKAGE.

FAUCET_COMPONENT

The faucet native component - use this on testnets for getting XRD and locking fee.

FAUCET_PACKAGE

The scrypto package for the faucet

FINALIZATION_COST_UNIT_LIMIT

The finalization cost unit limit.

FINALIZATION_COST_UNIT_PRICE_IN_XRD

The price of finalization cost unit, in XRD.

FREEZER_ROLE

FREEZER_UPDATER_ROLE

FUNGIBLE_BUCKET_BLUEPRINT

FUNGIBLE_BUCKET_CREATE_PROOF_OF_AMOUNT_IDENT

FUNGIBLE_BUCKET_LOCK_AMOUNT_IDENT

FUNGIBLE_BUCKET_UNLOCK_AMOUNT_IDENT

FUNGIBLE_PROOF_BLUEPRINT

FUNGIBLE_RESOURCE_MANAGER_BLUEPRINT

FUNGIBLE_RESOURCE_MANAGER_CREATE_IDENT

FUNGIBLE_RESOURCE_MANAGER_CREATE_WITH_INITIAL_SUPPLY_IDENT

FUNGIBLE_RESOURCE_MANAGER_MINT_IDENT

FUNGIBLE_VAULT_BLUEPRINT

FUNGIBLE_VAULT_CREATE_PROOF_OF_AMOUNT_IDENT

FUNGIBLE_VAULT_LOCK_FEE_IDENT

FUNGIBLE_VAULT_LOCK_FUNGIBLE_AMOUNT_IDENT

FUNGIBLE_VAULT_UNLOCK_FUNGIBLE_AMOUNT_IDENT

GENESIS_HELPER

The genesis helper scrypto component - used for sorting out genesis.

GENESIS_HELPER_BLUEPRINT

The name of the genesis helper blueprint under the GENESIS_HELPER_PACKAGE.

GENESIS_HELPER_PACKAGE

The scrypto package for the genesis helper.

GLOBAL_ADDRESS_TYPE

GLOBAL_CALLER_VIRTUAL_BADGE

The non-fungible badge resource which is used for virtual proofs which represent the global ancestor of the actor which made the latest global call.

IDENTITY_OWNER_BADGE

The non-fungible badge resource which is used for identity ownership, if identities have been set up with simple account creation, or have been securified.

IDENTITY_PACKAGE

The native package for identities.

INSTANT_TYPE

INTERNAL_ADDRESS_TYPE

KEY_VALUE_STORE_PAYLOAD_MAX_DEPTH

Depth limit for the key and value of an entry in KeyValueStore

KV_STORE_DATA_SCHEMA_VARIANT_LOCAL

KV_STORE_DATA_SCHEMA_VARIANT_REMOTE

MAIN_BASE_PARTITION

MANIFEST_ADDRESS_DISCRIMINATOR_NAMED

MANIFEST_ADDRESS_DISCRIMINATOR_STATIC

MANIFEST_NON_FUNGIBLE_LOCAL_ID_MAX_LENGTH

MANIFEST_SBOR_V1_MAX_DEPTH

MANIFEST_SBOR_V1_PAYLOAD_PREFIX

0x4d = M in ASCII for Manifest - (77 in decimal)

MANIFEST_VALUE_KIND_ADDRESS

MANIFEST_VALUE_KIND_ADDRESS_RESERVATION

MANIFEST_VALUE_KIND_BLOB

MANIFEST_VALUE_KIND_BUCKET

MANIFEST_VALUE_KIND_DECIMAL

MANIFEST_VALUE_KIND_EXPRESSION

MANIFEST_VALUE_KIND_NON_FUNGIBLE_LOCAL_ID

MANIFEST_VALUE_KIND_PRECISE_DECIMAL

MANIFEST_VALUE_KIND_PROOF

MAX_ACCESS_RULE_DEPTH

The max depth of an access rule, to protect unbounded native stack usage

MAX_ACCESS_RULE_NODES

The max number of access rule nodes in an access rule

MAX_BLUEPRINT_NAME_LEN

The max length of a blueprint identifier

MAX_CALL_DEPTH

The max call depth, used by transaction executor.

MAX_EPOCH_RANGE

The max epoch range Should be ~ 1 month. The below is ~30 days given 5 minute epochs.

MAX_EVENT_NAME_LEN

The max length of an event name

MAX_EVENT_SIZE

The max event size

MAX_EXECUTION_TRACE_DEPTH

MAX_FEATURE_NAME_LEN

The max length of a feature name

MAX_FUNCTION_NAME_LEN

The max length of a function name

MAX_HEAP_SUBSTATE_TOTAL_BYTES

The max total heap substate size.

MAX_INITIAL_TABLE_SIZE

The maximum initial table size

MAX_INVOKE_PAYLOAD_SIZE

The maximum invoke payload size.

MAX_LOG_SIZE

The max log size

MAX_MEMORY_SIZE_IN_PAGES

The maximum memory size (per call frame): 64 * 64KiB = 4MiB

MAX_METADATA_KEY_STRING_LEN

The max SBOR size of metadata key

MAX_METADATA_VALUE_SBOR_LEN

The max SBOR size of metadata value

MAX_NUMBER_OF_BLOBS

MAX_NUMBER_OF_BLUEPRINT_FIELDS

The max number of fields a blueprint can have

MAX_NUMBER_OF_BR_TABLE_TARGETS

The max number of labels of a table jump, excluding the default

MAX_NUMBER_OF_BUFFERS

The maximum number of “live” buffers maintained by Scrypto runtime.

MAX_NUMBER_OF_EVENTS

The max number of events

MAX_NUMBER_OF_FIELDS

MAX_NUMBER_OF_FUNCTIONS

The max number of functions

MAX_NUMBER_OF_FUNCTION_LOCALS

The max number of function local variables

MAX_NUMBER_OF_FUNCTION_PARAMS

The max number of function parameters

MAX_NUMBER_OF_GLOBALS

The max number of global variables

MAX_NUMBER_OF_INTENT_SIGNATURES

MAX_NUMBER_OF_LOGS

The max number of logs

MAX_ORIGIN_LENGTH

The max length of an Origin in metadata

MAX_PANIC_MESSAGE_SIZE

The max panic message size

MAX_PER_FUNCTION_ROYALTY_IN_XRD

The maximum that a package or component owner is allowed to set their method royalty to. 10 USD

MAX_ROLES

The max number of roles in a Role Specification

MAX_ROLE_NAME_LEN

The max length of a role name

MAX_SUBSTATE_KEY_SIZE

The maximum substate key read and write size

MAX_SUBSTATE_VALUE_SIZE

The maximum substate read and write size.

MAX_TIP_PERCENTAGE

The maximum value of tip percentage

MAX_TRACK_SUBSTATE_TOTAL_BYTES

The max total track substate size.

MAX_TRANSACTION_SIZE

The max transaction size

MAX_TYPE_NAME_LEN

The max length of a registered type name

MAX_URL_LENGTH

The max length of a URL in metadata

METADATA_BASE_PARTITION

METADATA_BLUEPRINT

METADATA_CREATE_IDENT

METADATA_CREATE_WITH_DATA_IDENT

METADATA_DISCRIMINATOR_ARRAY_BASE

METADATA_GET_IDENT

METADATA_KV_STORE_PARTITION_OFFSET

METADATA_LOCKER_ROLE

METADATA_LOCKER_UPDATER_ROLE

METADATA_LOCK_IDENT

METADATA_MODULE_PACKAGE

The native package for the metadata module.

METADATA_REMOVE_IDENT

METADATA_SETTER_ROLE

METADATA_SETTER_UPDATER_ROLE

METADATA_SET_IDENT

METADATA_VALUE_BOOLEAN_ARRAY_DISCRIMINATOR

METADATA_VALUE_BOOLEAN_DISCRIMINATOR

METADATA_VALUE_DECIMAL_ARRAY_DISCRIMINATOR

METADATA_VALUE_DECIMAL_DISCRIMINATOR

METADATA_VALUE_GLOBAL_ADDRESS_ARRAY_DISCRIMINATOR

METADATA_VALUE_GLOBAL_ADDRESS_DISCRIMINATOR

METADATA_VALUE_I32_ARRAY_DISCRIMINATOR

METADATA_VALUE_I32_DISCRIMINATOR

METADATA_VALUE_I64_ARRAY_DISCRIMINATOR

METADATA_VALUE_I64_DISCRIMINATOR

METADATA_VALUE_INSTANT_ARRAY_DISCRIMINATOR

METADATA_VALUE_INSTANT_DISCRIMINATOR

METADATA_VALUE_NON_FUNGIBLE_GLOBAL_ID_ARRAY_DISCRIMINATOR

METADATA_VALUE_NON_FUNGIBLE_GLOBAL_ID_DISCRIMINATOR

METADATA_VALUE_NON_FUNGIBLE_LOCAL_ID_ARRAY_DISCRIMINATOR

METADATA_VALUE_NON_FUNGIBLE_LOCAL_ID_DISCRIMINATOR

METADATA_VALUE_ORIGIN_ARRAY_DISCRIMINATOR

METADATA_VALUE_ORIGIN_DISCRIMINATOR

METADATA_VALUE_PUBLIC_KEY_ARRAY_DISCRIMINATOR

METADATA_VALUE_PUBLIC_KEY_DISCRIMINATOR

METADATA_VALUE_PUBLIC_KEY_HASH_ARRAY_DISCRIMINATOR

METADATA_VALUE_PUBLIC_KEY_HASH_DISCRIMINATOR

METADATA_VALUE_STRING_ARRAY_DISCRIMINATOR

METADATA_VALUE_STRING_DISCRIMINATOR

METADATA_VALUE_U8_ARRAY_DISCRIMINATOR

METADATA_VALUE_U8_DISCRIMINATOR

METADATA_VALUE_U32_ARRAY_DISCRIMINATOR

METADATA_VALUE_U32_DISCRIMINATOR

METADATA_VALUE_U64_ARRAY_DISCRIMINATOR

METADATA_VALUE_U64_DISCRIMINATOR

METADATA_VALUE_URL_ARRAY_DISCRIMINATOR

METADATA_VALUE_URL_DISCRIMINATOR

MINTER_ROLE

MINTER_UPDATER_ROLE

MIN_TIP_PERCENTAGE

The minimum value of tip percentage

MODULE_ID_TYPE

NETWORK_FEES_PROPOSER_SHARE_PERCENTAGE

The proposer’s share of network fees (execution, finalization and storage)

NETWORK_FEES_VALIDATOR_SET_SHARE_PERCENTAGE

The validator set’s share of network fees (execution, finalization and storage)

NON_FUNGIBLE_BUCKET_BLUEPRINT

NON_FUNGIBLE_BUCKET_CONTAINS_NON_FUNGIBLE_IDENT

NON_FUNGIBLE_BUCKET_CREATE_PROOF_OF_NON_FUNGIBLES_IDENT

NON_FUNGIBLE_BUCKET_GET_NON_FUNGIBLE_LOCAL_IDS_IDENT

NON_FUNGIBLE_BUCKET_LOCK_NON_FUNGIBLES_IDENT

NON_FUNGIBLE_BUCKET_TAKE_NON_FUNGIBLES_IDENT

NON_FUNGIBLE_BUCKET_UNLOCK_NON_FUNGIBLES_IDENT

NON_FUNGIBLE_DATA_SCHEMA_VARIANT_LOCAL

NON_FUNGIBLE_DATA_SCHEMA_VARIANT_REMOTE

NON_FUNGIBLE_DATA_UPDATER_ROLE

NON_FUNGIBLE_DATA_UPDATER_UPDATER_ROLE

NON_FUNGIBLE_GLOBAL_ID_TYPE

NON_FUNGIBLE_LOCAL_ID_MAX_LENGTH

NON_FUNGIBLE_LOCAL_ID_TYPE

NON_FUNGIBLE_PROOF_BLUEPRINT

NON_FUNGIBLE_PROOF_GET_LOCAL_IDS_IDENT

NON_FUNGIBLE_RESOURCE_MANAGER_BLUEPRINT

NON_FUNGIBLE_RESOURCE_MANAGER_CREATE_IDENT

NON_FUNGIBLE_RESOURCE_MANAGER_CREATE_RUID_WITH_INITIAL_SUPPLY_IDENT

NON_FUNGIBLE_RESOURCE_MANAGER_CREATE_WITH_INITIAL_SUPPLY_IDENT

NON_FUNGIBLE_RESOURCE_MANAGER_EXISTS_IDENT

NON_FUNGIBLE_RESOURCE_MANAGER_GET_NON_FUNGIBLE_IDENT

NON_FUNGIBLE_RESOURCE_MANAGER_MINT_IDENT

NON_FUNGIBLE_RESOURCE_MANAGER_MINT_RUID_IDENT

NON_FUNGIBLE_RESOURCE_MANAGER_MINT_SINGLE_RUID_IDENT

NON_FUNGIBLE_RESOURCE_MANAGER_UPDATE_DATA_IDENT

NON_FUNGIBLE_VAULT_BLUEPRINT

NON_FUNGIBLE_VAULT_BURN_NON_FUNGIBLES_IDENT

NON_FUNGIBLE_VAULT_CONTAINS_NON_FUNGIBLE_IDENT

NON_FUNGIBLE_VAULT_CREATE_PROOF_OF_NON_FUNGIBLES_IDENT

NON_FUNGIBLE_VAULT_GET_NON_FUNGIBLE_LOCAL_IDS_IDENT

NON_FUNGIBLE_VAULT_LOCK_NON_FUNGIBLES_IDENT

NON_FUNGIBLE_VAULT_RECALL_NON_FUNGIBLES_IDENT

NON_FUNGIBLE_VAULT_TAKE_NON_FUNGIBLES_IDENT

NON_FUNGIBLE_VAULT_UNLOCK_NON_FUNGIBLES_IDENT

OPTION_VARIANT_NONE

OPTION_VARIANT_SOME

ORIGIN_TYPE

OWNER_ROLE

OWNER_ROLE_TYPE

OWN_BUCKET_TYPE

OWN_FUNGIBLE_BUCKET_TYPE

OWN_FUNGIBLE_PROOF_TYPE

OWN_FUNGIBLE_VAULT_TYPE

OWN_GLOBAL_ADDRESS_RESERVATION_TYPE

OWN_KEY_VALUE_STORE_TYPE

OWN_NON_FUNGIBLE_BUCKET_TYPE

OWN_NON_FUNGIBLE_PROOF_TYPE

OWN_NON_FUNGIBLE_VAULT_TYPE

OWN_PROOF_TYPE

OWN_TYPE

OWN_VAULT_TYPE

PACKAGE_ADDRESS_TYPE

PACKAGE_OF_DIRECT_CALLER_VIRTUAL_BADGE

The non-fungible badge resource which is used for virtual proofs which represent the package of the immediate caller - ie the actor which made the latest (global or internal) call.

PACKAGE_OWNER_BADGE

The non-fungible badge resource which is used for package ownership when creating packages with the simple package creation set-up.

PACKAGE_PACKAGE

The native package for package deployment.

POOL_PACKAGE

The native package for pools.

PRECISE_DECIMAL_SIZE

PRECISE_DECIMAL_TYPE

PREVIEW_CREDIT_IN_XRD

The free credit amount used by preview. 1000000 XRD = $60000, which should be enough to cover all sensible preview requests.

PROOF_CLONE_IDENT

PROOF_DROP_IDENT

PROOF_GET_AMOUNT_IDENT

PROOF_GET_RESOURCE_ADDRESS_IDENT

PROOF_RULE_TYPE

PUBLIC_KEY_HASH_TYPE

PUBLIC_KEY_TYPE

RECALLER_ROLE

RECALLER_UPDATER_ROLE

REFERENCE_TYPE

RESOURCE_ADDRESS_TYPE

RESOURCE_MANAGER_BURN_IDENT

RESOURCE_MANAGER_CREATE_EMPTY_BUCKET_IDENT

RESOURCE_MANAGER_CREATE_EMPTY_VAULT_IDENT

RESOURCE_MANAGER_DROP_EMPTY_BUCKET_IDENT

RESOURCE_MANAGER_GET_AMOUNT_FOR_WITHDRAWAL_IDENT

RESOURCE_MANAGER_GET_RESOURCE_TYPE_IDENT

RESOURCE_MANAGER_GET_TOTAL_SUPPLY_IDENT

RESOURCE_MANAGER_PACKAGE_BURN_IDENT

RESOURCE_OR_NON_FUNGIBLE_LIST_TYPE

RESOURCE_OR_NON_FUNGIBLE_TYPE

RESOURCE_PACKAGE

The native package for resource managers, proofs, buckets, vaults etc.

RESULT_VARIANT_ERR

RESULT_VARIANT_OK

ROLA_HASHABLE_PAYLOAD_PREFIX

The ROLA hash which is signed is created as hash(ROLA_HASHABLE_PAYLOAD_PREFIX || ..)

ROLE_ASSIGNMENT_BASE_PARTITION

ROLE_ASSIGNMENT_BLUEPRINT

ROLE_ASSIGNMENT_CREATE_IDENT

ROLE_ASSIGNMENT_FIELDS_PARTITION

ROLE_ASSIGNMENT_FIELDS_PARTITION_OFFSET

ROLE_ASSIGNMENT_GET_IDENT

ROLE_ASSIGNMENT_LOCK_OWNER_IDENT

ROLE_ASSIGNMENT_MODULE_PACKAGE

The native package for the role assignment module.

ROLE_ASSIGNMENT_MUTABILITY_PARTITION_OFFSET

ROLE_ASSIGNMENT_ROLE_DEF_PARTITION

ROLE_ASSIGNMENT_ROLE_DEF_PARTITION_OFFSET

ROLE_ASSIGNMENT_SET_IDENT

ROLE_ASSIGNMENT_SET_OWNER_IDENT

ROLE_KEY_TYPE

ROYALTY_AMOUNT_TYPE

ROYALTY_BASE_PARTITION

ROYALTY_CONFIG_PARTITION

ROYALTY_CONFIG_PARTITION_OFFSET

ROYALTY_FIELDS_PARTITION

ROYALTY_FIELDS_PARTITION_OFFSET

ROYALTY_MODULE_PACKAGE

The native package for the royalty module.

SCHEMAS_PARTITION

SCRYPTO_SBOR_V1_MAX_DEPTH

SCRYPTO_SBOR_V1_PAYLOAD_PREFIX

0x5c for [5c]rypto - (92 in decimal)

SECP256K1_PUBLIC_KEY_HASH_TYPE

SECP256K1_PUBLIC_KEY_TYPE

SECP256K1_SIGNATURE_VIRTUAL_BADGE

The non-fungible badge resource which is used for virtual proofs of ECDSA Secp256k1 transacton signatures in the transaction processor.

SELF_ROLE

STATE_STORAGE_PRICE_IN_XRD

The price for adding a single byte to state storage, in XRD. 1 MB = 6 USD

SYSTEM_TRANSACTION_BADGE

The non-fungible badge resource which is used for virtual proofs representing the fact that the current transaction is a system transaction.

TEST_UTILS_PACKAGE

The native package for test utils.

TIPS_PROPOSER_SHARE_PERCENTAGE

The proposer’s share of tips

TIPS_VALIDATOR_SET_SHARE_PERCENTAGE

The validator set’s share of tips

TRANSACTION_HASHABLE_PAYLOAD_PREFIX

The Transaction hash which is signed is created as: hash(TRANSACTION_HASHABLE_PAYLOAD_PREFIX || version prefix according to type of transaction payload || ..)

TRANSACTION_PROCESSOR_PACKAGE

The native package for the transaction processor.

TRANSACTION_TRACKER

The intent hash store component

TRANSACTION_TRACKER_BLUEPRINT

TRANSACTION_TRACKER_PACKAGE

TRANSACTION_VERSION_V1

TRANSIENT_SUBSTATE_DEFAULT_VALUE_MAX_DEPTH

Depth limit for the default value of a transient substate

TYPE_INFO_FIELD_PARTITION

URL_TYPE

USD_PRICE_IN_XRD

The USD price, in XRD. 1 XRD = 0.06 USD

UTC_DATE_TIME_TYPE

VALIDATOR_OWNER_BADGE

The non-fungible badge resource which is used for validator ownership.

VALUE_KIND_ARRAY

VALUE_KIND_BOOL

VALUE_KIND_DECIMAL

VALUE_KIND_ENUM

VALUE_KIND_I8

VALUE_KIND_I16

VALUE_KIND_I32

VALUE_KIND_I64

VALUE_KIND_I128

VALUE_KIND_MAP

VALUE_KIND_NON_FUNGIBLE_LOCAL_ID

VALUE_KIND_OWN

VALUE_KIND_PRECISE_DECIMAL

VALUE_KIND_REFERENCE

VALUE_KIND_STRING

VALUE_KIND_TUPLE

VALUE_KIND_U8

VALUE_KIND_U16

VALUE_KIND_U32

VALUE_KIND_U64

VALUE_KIND_U128

VAULT_BURN_IDENT

VAULT_FREEZE_IDENT

VAULT_GET_AMOUNT_IDENT

VAULT_PUT_IDENT

VAULT_RECALL_IDENT

VAULT_TAKE_ADVANCED_IDENT

VAULT_TAKE_IDENT

VAULT_UNFREEZE_IDENT

WASM_ENGINE_CACHE_SIZE

The number of entries in the engine cache

WASM_EXECUTION_COST_UNITS_BUFFER

WITHDRAWER_ROLE

WITHDRAWER_UPDATER_ROLE

WORKTOP_ASSERT_CONTAINS_AMOUNT_IDENT

WORKTOP_ASSERT_CONTAINS_IDENT

WORKTOP_ASSERT_CONTAINS_NON_FUNGIBLES_IDENT

WORKTOP_BLUEPRINT

WORKTOP_DRAIN_IDENT

WORKTOP_DROP_IDENT

WORKTOP_PUT_IDENT

WORKTOP_TAKE_ALL_IDENT

WORKTOP_TAKE_IDENT

WORKTOP_TAKE_NON_FUNGIBLES_IDENT

XRD

XRD is the native token of the Radix ledger. It is a fungible token, measured in attos (10^-18).

Statics

NO_NETWORK

Traits

AccessRuleVisitor

ArrayMetadataVal

AsMut

Used to do a cheap mutable-to-mutable reference conversion.

AsRef

Used to do a cheap reference-to-reference conversion.

AsStaticRefDeprecated

A cheap reference-to-reference conversion. Used to convert a value to a reference value with 'static lifetime within generic code.

BasicCategorize

BasicDecode

BasicDescribe

BasicEncode

BasicSbor

BasicSborEnum

BasicSborTuple

BlueprintFeature

BlueprintPartitionOffset

BorrowingDecoder

Categorize

The Categorize trait marks a rust type as having a fixed value kind for SBOR encoding/decoding.

Cbrt

CheckedAdd

CheckedDiv

CheckedMul

CheckedNeg

CheckedSub

CheckedTruncate

ClientActorApi

Api which exposes methods in the context of the actor

ClientActorIndexApi

Api to manage an iterable index

ClientActorKeyValueEntryApi

ClientActorSortedIndexApi

ClientApi

Interface of the system, for blueprints and Node modules.

ClientBlueprintApi

ClientCostingApi

ClientCryptoUtilsApi

ClientExecutionTraceApi

ClientFieldApi

A high level api to read/write fields

ClientKeyValueEntryApi

ClientKeyValueStoreApi

ClientObjectApi

A high level interface to manipulate objects in the actor’s call frame

ClientTransactionRuntimeApi

Clone

A common trait for the ability to explicitly duplicate an object.

CloneIntoLatest

CollectionDescriptor

ContextualDisplay

This trait is used where context is required to correctly display a value.

ContextualSerialize

This trait is used where context is required to correctly serialize a value.

Copy

Types whose values can be duplicated simply by copying bits.

CustomExtension

CustomSchema

CustomTypeKind

CustomTypeValidation

CustomValue

Represents a custom SBOR value.

CustomValueKind

Debug

? formatting.

Decode

A data structure that can be decoded from a byte array using SBOR.

Decoder

Default

A trait for giving a type a useful default value.

Describe

The Describe trait allows a type to describe how to interpret and validate a corresponding SBOR payload.

Display

Format trait for an empty format, {}.

DoubleEndedIterator

An iterator able to yield elements from both ends.

Drop

Custom code within the destructor.

Encode

A data structure that can be serialized into a byte array using SBOR.

Encoder

EnumCount

A trait for capturing the number of variants in Enum. This trait can be autoderived by strum_macros.

EnumMessage

Associates additional pieces of information with an Enum. This can be autoimplemented by deriving EnumMessage and annotating your variants with #[strum(message="...")].

EnumProperty

EnumProperty is a trait that makes it possible to store additional information with enum variants. This trait is designed to be used with the macro of the same name in the strum_macros crate. Currently, the only string literals are supported in attributes, the other methods will be implemented as additional attribute types become stabilized.

Eq

Trait for comparisons corresponding to equivalence relations.

ExactSizeIterator

An iterator that knows its exact length.

Extend

Extend a collection with the contents of an iterator.

FieldDescriptor

FieldPayloadMarker

Fn

The version of the call operator that takes an immutable receiver.

FnMut

The version of the call operator that takes a mutable receiver.

FnOnce

The version of the call operator that takes a by-value receiver.

ForWithdrawal

From

Used to do value-to-value conversions while consuming the input value. It is the reciprocal of Into.

FromIterator

Conversion from an Iterator.

FromPrimitive

A generic trait for converting a number to a value.

FromPublicKey

FromStr

Parse a value from a string

HasLatestVersion

A marker trait to indicate that the type is versioned. This can be used for type bounds for requiring that types are versioned.

HasPublicKeyHash

HasSchemaHash

Into

A value-to-value conversion that consumes the input value. The opposite of From.

IntoEnumIterator

This trait designates that an Enum can be iterated over. It can be auto generated using strum_macros on your behalf.

IntoIterator

Conversion into an Iterator.

IsFixedEnumVariant

IsHash

IsNonAutoGeneratedNonFungibleLocalId

Marks the rust type that represents a non-fungible id, of non-auto-generated kind (i.e. String, Integer and Bytes).

IsNonFungibleLocalId

Marks the rust type that represents a non-fungible id, of any kind (i.e. String, Integer, Bytes and RUID).

IsPublicKeyHash

Iterator

A trait for dealing with iterators.

KeyValueEntryPayloadMarker

KeyValueKeyPayloadMarker

ManifestCategorize

ManifestDecode

ManifestEncode

ManifestSborEnum

ManifestSborTuple

MetadataVal

NonFungibleData

Represents the data structure of a non-fungible.

NthRoot

One

Defines a multiplicative identity element for Self.

Ord

Trait for types that form a total order.

PartialEq

Trait for comparisons using the equality operator.

PartialOrd

Trait for types that form a partial order.

Pow

Binary operator for raising a value to a power.

SborCustomValue

Represents a custom SBOR value.

SborEnum

Marker trait for encoding as an Sbor Enum

SborTuple

Marker trait for encoding as an Sbor Tuple

SchemaTypeLink

Marker trait for a link between TypeKinds:

ScryptoCategorize

ScryptoDecode

ScryptoDescribe

ScryptoEncode

ScryptoEvent

ScryptoSbor

ScryptoSborEnum

ScryptoSborTuple

Send

Types that can be transferred across thread boundaries.

Signed

Useful functions for signed numbers (i.e. numbers that can be negative).

SingleMetadataVal

Sized

Types with a constant size known at compile time.

SortedIndexEntryPayloadMarker

SortedIndexKeyPayloadMarker

Sqrt

Sync

Types for which it is safe to share references between threads.

ToMetadataEntry

ToOwned

A generalization of Clone to borrowed data.

ToPrimitive

A generic trait for converting a value to a number.

ToRoleEntry

ToString

A trait for converting a value to a String.

TryFrom

Simple and safe type conversions that may fail in a controlled way under some circumstances. It is the reciprocal of TryInto.

TryInto

An attempted conversion that consumes self, which may or may not be expensive.

Unpin

Types that do not require any pinning guarantees.

ValidatableCustomExtension

ValueVisitor

VariantNames

A trait for retrieving the names of each variant in Enum. This trait can be autoderived by strum_macros.

VersionedSchemaVersion

Zero

Defines an additive identity element for Self.

Functions

access_rule_node_list_type_data

access_rule_node_type_data

access_rule_type_data

aggregate_verify_bls12381_v1

Performs BLS12-381 G2 aggregated signature verification of multiple messages each signed with different key. Domain specifier tag: BLS_SIG_BLS12381G2_XMD:SHA-256_SSWU_RO_POP_

attached_module_id_type_data

basic_decode

Decode an instance of T from a slice.

basic_decode_with_depth_limit

basic_encode

Encode a T into byte array.

basic_encode_with_depth_limit

basic_payload_traverser

Creates a payload traverser from the buffer

bech32_encode_to_fmt

NOTE: The below code is copied with minor alterations from the bech32 crate. These alterations are to avoid using std for allocations, and fit with the sbor no-alloc options.

blake2b_256_hash

check_fungible_amount

check_non_fungible_amount

combine

Combines a u8 with a u8 slice.

component_address

component_address_type_data

copy_u8_array

Copies a slice to a fixed-sized array.

decimal_type_data

drop

Disposes of a value.

ed25519_public_key_hash_type_data

ed25519_public_key_type_data

fast_aggregate_verify_bls12381_v1

Performs BLS12-381 G2 aggregated signature verification one message signed with multiple keys. Domain specifier tag: BLS_SIG_BLS12381G2_XMD:SHA-256_SSWU_RO_POP_

from_manifest_value

generate_full_schema

generate_full_schema_from_single_type

global_address_type_data

global_caller

A requirement for the global ancestor of the actor who made the latest global call to either be:

hash

Computes the hash digest of a message.

hash_map_new

Creates an empty map with capacity 0 and default Hasher

hash_map_with_capacity

Creates an empty map with given capacity and default Hasher

hash_public_key_bytes

hash_set_new

Creates an empty set with capacity 0 and default Hasher

hash_set_with_capacity

Creates an empty set with given capacity and default Hasher

index_map_new

This is safe for std and no-std use cases (unlike IndexMap::new which disappears when std is not in the toolchain - see this article for deep technical reasons)

index_map_with_capacity

This is safe for std and no-std use cases (unlike IndexMap::with_capacity which disappears when std is not in the toolchain - see this article for deep technical reasons)

index_set_new

This is safe for std and no-std use cases (unlike IndexSet::new which disappears when std is not in the toolchain - see this article for deep technical reasons)

index_set_with_capacity

This is safe for std and no-std use cases (unlike IndexSet::with_capacity which disappears when std is not in the toolchain - see this article for deep technical reasons)

instant_type_data

internal_address_type_data

keccak256_hash

local_address

localize_well_known

localize_well_known_type_data

manifest_decode

manifest_decode_with_depth_limit

manifest_encode

manifest_encode_with_depth_limit

module_id_type_data

non_fungible_global_id_type_data

non_fungible_local_id_type_data

origin_type_data

own_bucket_type_data

own_fungible_bucket_type_data

own_fungible_proof_type_data

own_fungible_vault_type_data

own_global_address_reservation_type_data

own_key_value_store_type_data

own_non_fungible_bucket_type_data

own_non_fungible_proof_type_data

own_non_fungible_vault_type_data

own_proof_type_data

own_type_data

own_vault_type_data

owner_role_type_data

package_address

package_address_type_data

package_of_direct_caller

A requirement for the immediate caller’s package to equal the given package.

precise_decimal_type_data

proof_rule_type_data

public_key_hash_type_data

public_key_type_data

recover_secp256k1

reference_type_data

replace_self_package_address

require

require_all_of

require_amount

require_any_of

require_n_of

resolve_local_type_id

resolve_scrypto_well_known_type

resource_address

resource_address_type_data

resource_or_non_fungible_list_type_data

resource_or_non_fungible_type_data

role_key_type_data

royalty_amount_type_data

scrypto_decode

Decodes a data structure from a byte array.

scrypto_decode_with_depth_limit

scrypto_encode

Encodes a data structure into byte array.

scrypto_encode_with_depth_limit

secp256k1_public_key_hash_type_data

secp256k1_public_key_type_data

to_manifest_value

traverse_any

============================================== (DEPRECATED) TRAVERSAL

url_type_data

utc_date_time_type_data

validate_childless_metadata

validate_container

validate_custom_type_validation

validate_enum_metadata

validate_field_names

validate_index

validate_payload_against_schema

validate_schema

validate_schema_enum_variant_name

validate_schema_field_name

validate_schema_type_name

validate_terminal_value

validate_terminal_value_batch

validate_tuple_metadata

validate_type_kind

validate_type_metadata_with_type_kind

verify_bls12381_v1

Performs BLS12-381 G2 signature verification. Domain specifier tag: BLS_SIG_BLS12381G2_XMD:SHA-256_SSWU_RO_POP_

verify_ed25519

verify_secp256k1

Type Aliases

ActorRefHandle

ActorStateHandle

AuthZoneAssertAccessRuleOutput

AuthZoneCreateProofOfAllOutput

AuthZoneCreateProofOfAmountOutput

AuthZoneCreateProofOfNonFungiblesOutput

AuthZoneDrainOutput

AuthZoneDropProofsOutput

AuthZoneDropRegularProofsOutput

AuthZoneDropSignatureProofsOutput

AuthZonePopOutput

AuthZonePushOutput

BasicDecoder

BasicEncoder

BasicOwnedRawPayload

BasicOwnedRawValue

BasicRawPayload

BasicRawValue

BasicSchema

BasicTraverser

BasicTypeData

BasicTypeKind

BasicValue

BasicValueKind

BasicVersionedSchema

Blake2b256

BucketCreateProofOfAllOutput

BucketGetAmountOutput

BucketGetNonFungibleLocalIdsOutput

BucketGetResourceAddressOutput

BucketPutOutput

BucketTakeAdvancedOutput

BucketTakeNonFungiblesOutput

BucketTakeOutput

BufferId

CheckedMetadataValue

CollectionIndex

FieldHandle

FieldIndex

FieldKey

FungibleBucketCreateProofOfAmountOutput

FungibleBucketLockAmountOutput

FungibleBucketUnlockAmountOutput

FungibleResourceManagerCreateOutput

FungibleResourceManagerCreateWithInitialSupplyOutput

FungibleResourceManagerMintOutput

FungibleVaultCreateProofOfAmountOutput

FungibleVaultLockFeeOutput

FungibleVaultLockFungibleAmountOutput

FungibleVaultUnlockFungibleAmountOutput

HashMap

HashSet

IndexMap

IndexSet

KeyValueEntryHandle

ManifestDecoder

ManifestEncoder

ManifestEnumVariantValue

ManifestOwnedRawPayload

ManifestOwnedRawValue

ManifestRawPayload

ManifestRawValue

ManifestTraverser

ManifestValue

ManifestValueKind

MapKey

MetadataCreateOutput

MetadataCreateWithDataOutput

MetadataGetOutput

MetadataInit

MetadataLockOutput

MetadataRemoveOutput

MetadataSetOutput

MetadataValue

NonFungibleBucketContainsNonFungibleOutput

NonFungibleBucketCreateProofOfNonFungiblesOutput

NonFungibleBucketLockNonFungiblesOutput

NonFungibleBucketUnlockNonFungiblesOutput

NonFungibleProofGetLocalIdsOutput

NonFungibleResourceManagerCreateOutput

NonFungibleResourceManagerCreateRuidWithInitialSupplyOutput

NonFungibleResourceManagerCreateWithInitialSupplyOutput

NonFungibleResourceManagerExistsOutput

NonFungibleResourceManagerGetNonFungibleOutput

NonFungibleResourceManagerMintOutput

NonFungibleResourceManagerMintRuidOutput

NonFungibleResourceManagerMintSingleRuidOutput

NonFungibleResourceManagerUpdateDataOutput

NonFungibleVaultBurnNonFungiblesOutput

NonFungibleVaultContainsNonFungibleOutput

NonFungibleVaultCreateProofOfNonFungiblesOutput

NonFungibleVaultGetNonFungibleLocalIdsOutput

NonFungibleVaultLockNonFungiblesOutput

NonFungibleVaultRecallNonFungiblesOutput

NonFungibleVaultTakeNonFungiblesOutput

NonFungibleVaultUnlockNonFungiblesOutput

ObjectModuleId

Notes: This is to be deprecated, please use ModuleId instead

Origin

We should always UncheckedOrigin in Scrypto, as the validation logic is heavy. Thus, this type alias is added.

ProofCloneOutput

ProofDropOutput

ProofGetAmountOutput

ProofGetResourceAddressOutput

RawScryptoPayload

RawScryptoValue

ResourceManagerBurnOutput

ResourceManagerCreateEmptyBucketOutput

ResourceManagerCreateEmptyVaultOutput

ResourceManagerDropEmptyBucketOutput

ResourceManagerGetAmountForWithdrawalOutput

ResourceManagerGetResourceTypeOutput

ResourceManagerGetTotalSupplyOutput

ResourceManagerPackageBurnOutput

RoleAssignmentCreateOutput

RoleAssignmentGetOutput

RoleAssignmentSetOutput

RoleAssignmentSetOwnerOutput

RoleAssingmentLockOwnerOutput

RoleDefinition

Schema

SchemaTypeKind

ScryptoDecoder

ScryptoEncoder

ScryptoOwnedRawPayload

ScryptoOwnedRawValue

ScryptoRawPayload

ScryptoRawValue

ScryptoTraverser

ScryptoTypeData

ScryptoTypeKind

ScryptoValue

ScryptoValueKind

SortedKey

SubstateHandle

TimePrecision

TypeHash

Url

We should always UncheckedUrl in Scrypto, as the validation logic is heavy. Thus, this type alias is added.

ValidatorIndex

An index of a specific validator within the current validator set. To be exact: a ValidatorIndex equal to k references the k-th element returned by the iterator of the IndexMap<ComponentAddress, Validator> in this epoch’s active validator set (which is expected to be sorted by stake, descending). This uniquely identifies the validator, while being shorter than ComponentAddress (we do care about the constant factor of the space taken by LeaderProposalHistory under prolonged liveness break scenarios).

VaultBurnOutput

VaultFreezeOutput

VaultGetAmountOutput

VaultPutOutput

VaultRecallOutput

VaultTakeAdvancedOutput

VaultTakeOutput

VaultUnfreezeOutput

VersionedScryptoSchema

WorktopAssertContainsAmountOutput

WorktopAssertContainsNonFungiblesOutput

WorktopAssertContainsOutput

WorktopDrainOutput

WorktopDropOutput

WorktopPutOutput

WorktopTakeAllOutput

WorktopTakeNonFungiblesOutput

WorktopTakeOutput

Attribute Macros

blueprint

Declares a blueprint.

Derive Macros

AsRefStr

Converts enum variants to &'static str.

AsStaticStr

BasicCategorize

Derive code that returns the value kind - specifically for Basic SBOR.

BasicDecode

Derive code that decodes this data structure from a byte array - specifically for Basic SBOR.

BasicDescribe

Derive code that describes the type - specifically for Basic SBOR.

BasicEncode

Derive code that encodes this data structure - specifically for Basic SBOR.

BasicSbor

A shortcut for BasicCategorize, BasicEncode, BasicDecode, and BasicDescribe derives.

Categorize

Derive code that returns the value kind.

Clone

Derive macro generating an impl of the trait Clone.

Copy

Derive macro generating an impl of the trait Copy.

Debug

Derive macro generating an impl of the trait Debug.

Decode

Derive code that decodes this data structure from a byte array.

Default

Derive macro generating an impl of the trait Default.

Describe

Derive code that describes this type.

Display

Converts enum variants to strings.

Encode

Derive code that encodes this data structure

EnumCount

Add a constant usize equal to the number of variants.

EnumDiscriminants

Generate a new type with only the discriminant names.

EnumIter

Creates a new type that iterates of the variants of an enum.

EnumMessage

Add a verbose message to an enum variant.

EnumProperty

Add custom properties to enum variants.

EnumString

Converts strings to enum variants based on their name.

EnumVariantNames

Implements Strum::VariantNames which adds an associated constant VARIANTS which is an array of discriminant names.

Eq

Derive macro generating an impl of the trait Eq.

FromRepr

Add a function to enum that allows accessing variants by its discriminant

IntoStaticStr

Implements From<MyEnum> for &'static str on an enum.

ManifestCategorize

Derives code for categorizing a struct or enum with Manifest value model.

ManifestDecode

Derives code for decoding a struct or enum with Manifest value model.

ManifestEncode

Derives code for encoding a struct or enum with Manifest value model.

ManifestSbor

A shortcut for ManifestCategorize, ManifestEncode and ManifestDecode derives.

NonFungibleData

Derive code that describe a non-fungible data structure.

Ord

Derive macro generating an impl of the trait Ord. The behavior of this macro is described in detail here.

PartialEq

Derive macro generating an impl of the trait PartialEq. The behavior of this macro is described in detail here.

PartialOrd

Derive macro generating an impl of the trait PartialOrd. The behavior of this macro is described in detail here.

Sbor

A shortcut for Categorize, Encode, Decode, and Describe derives.

ScryptoCategorize

Derives code for categorizing a struct or enum with Scrypto value model.

ScryptoDecode

Derives code for decoding a struct or enum with Scrypto value model.

ScryptoEncode

Derives code for encoding a struct or enum with Scrypto value model.

ScryptoEvent

Derive code for implementing the required logic to mark a type as being an event.

ScryptoSbor

A shortcut for ScryptoCategorize, ScryptoEncode, ScryptoDecode, and [ScryptoDescribe] derives.

ToString

implements std::string::ToString on en enum