Struct scrypto::prelude::decimal::Decimal

pub struct Decimal(pub BnumI256);

Decimal represents a 256 bit representation of a fixed-scale decimal number.

The finite set of values are of the form m / 10^18, where m is an integer such that -2^(256 - 1) <= m < 2^(256 - 1).

Unless otherwise specified, all operations will panic if underflow/overflow.

Tuple Fields
0: BnumI256
Implementations

impl Decimal

pub const MIN: Decimal = Self(BnumI256::MIN)

The min value of Decimal.

pub const MAX: Decimal = Self(BnumI256::MAX)

The max value of Decimal.

pub const BITS: usize = 256usize

The bit length of number storing Decimal.

pub const SCALE: u32 = 18u32

The fixed scale used by Decimal.

pub const ZERO: Decimal = Self(BnumI256::ZERO)

pub const ONE: Decimal = Self(BnumI256::from_digits([10u64.pow(Decimal::SCALE), 0, 0, 0]))

pub fn zero() -> Decimal

Returns Decimal of 0.

pub fn one() -> Decimal

Returns Decimal of 1.

pub fn is_zero(&self) -> bool

Whether this decimal is zero.

pub fn is_positive(&self) -> bool

Whether this decimal is positive.

pub fn is_negative(&self) -> bool

Whether this decimal is negative.

pub fn abs(&self) -> Decimal

Returns the absolute value.

pub fn floor(&self) -> Decimal

Returns the largest integer that is equal to or less than this number.

pub fn ceiling(&self) -> Decimal

Returns the smallest integer that is equal to or greater than this number.

pub fn round(&self, decimal_places: u32, mode: RoundingMode) -> Decimal

pub fn powi(&self, exp: i64) -> Decimal

Calculates power using exponentiation by squaring“.

pub fn sqrt(&self) -> Option<Decimal>

Square root of a Decimal

pub fn cbrt(&self) -> Decimal

Cubic root of a Decimal

pub fn nth_root(&self, n: u32) -> Option<Decimal>

Nth root of a Decimal

impl Decimal

pub fn to_vec(&self) -> Vec<u8, Global>