Module scrypto::prelude::radix_engine_common::prelude

Each module should have its own prelude, which:

Adds preludes of upstream crates

Exports types with specific-enough names which mean they can safely be used downstream.

The idea is that we can just include the current crate’s prelude and avoid messing around with tons of includes. This makes refactors easier, and makes integration into the node less painful.

Modules

btree_map

btree_set

fmt

Utilities for formatting and printing Strings.

hash_map

hash_set

index_map

The methods and macros provided directly in this index_map module (new, with_capacity) work in both std and no-std modes - unlike the corresponding methods on IndexMap itself.

index_set

The methods and macros provided directly in this index_set module (new, with_capacity) work in both std and no-std modes - unlike the corresponding methods on IndexSet itself.

non_iter_map

representations

SBOR textual representations

vec

A contiguous growable array type with heap-allocated contents, written Vec<T>.

Macros

btreemap

btreeset

dec

Creates a Decimal from literals.

define_wrapped_hash

format

Creates a String using interpolation of runtime expressions.

hashmap

hashset

i

Creates a safe integer from literals. You must specify the type of the integer you want to create.

indexmap

indexset

manifest_args

pdec

Creates a PreciseDecimal from literals.

scrypto_args

to_manifest_value_and_unwrap

vec

Creates a Vec containing the arguments.

Structs

AddressBech32Decoder

Represents a decoder which understands how to decode Scrypto addresses in Bech32.

AddressBech32Encoder

Represents an encoder which understands how to encode Scrypto addresses in Bech32.

AddressDisplayContext

BTreeMap

An ordered map based on a B-Tree.

BTreeSet

An ordered set based on a B-Tree.

BnumI256

Provides safe integer arithmetic.

BnumI384

Provides safe integer arithmetic.

BnumI512

Provides safe integer arithmetic.

BnumI768

Provides safe integer arithmetic.

BnumU256

Provides safe integer arithmetic.

BnumU384

Provides safe integer arithmetic.

BnumU512

Provides safe integer arithmetic.

BnumU768

Provides safe integer arithmetic.

Box

A pointer type that uniquely owns a heap allocation of type T.

ContextDisplayable

ContextSerializable

HashAccumulator

Represents a 32-byte hash accumulator.

LazyCellExperimental

A value which is initialized on the first access.

ManifestArgs

NonIterMap

A thin wrapper around a HashMap, which guarantees that a HashMap usage will not result in a non-deterministic execution (simply by disallowing the iteration over its elements).

SyncUnsafeCellExperimental

UnsafeCell, but Sync.

BorrowError

An error returned by RefCell::try_borrow.

BorrowMutError

An error returned by RefCell::try_borrow_mut.

BytesNonFungibleLocalId

Bytes, of length between 1 and 64.

Cell

A mutable memory location.

ComponentAddress

Address to a global component

Decimal

Decimal represents a 256 bit representation of a fixed-scale decimal number.

Ed25519PublicKey

Represents an ED25519 public key.

Ed25519PublicKeyHash

Epoch

A type-safe consensus epoch number.

GlobalAddress

Address to a global entity

GlobalAddressReservation

Hash

Represents a 32-byte hash digest.

HrpSet

Represents an HRP set (typically corresponds to a network).

Instant

Represents a Unix timestamp, capturing the seconds since the unix epoch.

IntegerNonFungibleLocalId

Unsigned integers, up to u64.

InternalAddress

Address to a local entity

ManifestAddressReservation

ManifestBlobRef

ManifestBucket

ManifestCustomTerminalValueRef

ManifestDecimal

ManifestPreciseDecimal

ManifestProof

ManifestValueDisplayContext

Note - this is quite similar to ManifestDecompilationDisplayContext

NetworkDefinition

Network Definition is intended to be the actual definition of a network

NodeId

The unique identifier of a (stored) node.

OnceCell

A cell which can be written to only once.

Own

PackageAddress

Address to a global package

PartitionNumber

The unique identifier of a node module.

PartitionOffset

PhantomData

Zero-sized type used to mark things that “act like” they own a T.

PreciseDecimal

PreciseDecimal represents a 512 bit representation of a fixed-scale decimal number.

RUIDNonFungibleLocalId

RUID, v4, variant 1, big endian. See https://www.rfc-editor.org/rfc/rfc4122

Rc

A single-threaded reference-counting pointer. ‘Rc’ stands for ‘Reference Counted’.

Ref

Wraps a borrowed reference to a value in a RefCell box. A wrapper type for an immutably borrowed value from a RefCell<T>.

RefCell

A mutable memory location with dynamically checked borrow rules

RefMut

A wrapper type for a mutably borrowed value from a RefCell<T>.

Reference

ResourceAddress

Address to a global resource

Round

A type-safe consensus round number within a single epoch.

SborFixedEnumVariant

SborPath

A series of indexes which describes some value in the sbor tree

SborPathBuf

SborRawPayload

A wrapper for a full encoded SBOR payload, including the prefix byte.

SborRawValue

A wrapper for a reference to a valid partial SBOR payload representing a single value.

SchemaPath

Describes a value located in some sbor given a schema for that sbor

ScryptoCustomExtension

ScryptoCustomSchema

ScryptoCustomTerminalValueRef

ScryptoValueDisplayContext

Secp256k1PublicKey

Represents an ECDSA Secp256k1 public key.

Secp256k1PublicKeyHash

String

A UTF-8–encoded, growable string.

StringNonFungibleLocalId

A string matching [_0-9a-zA-Z]{1,64}.

UnsafeCell

The core primitive for interior mutability in Rust.

UtcDateTime

A UtcDateTime represents a Unix timestamp on the UTC Calendar.

Vec

A contiguous growable array type, written as Vec<T>, short for ‘vector’.

Enums

AddressBech32DecodeError

AddressBech32EncodeError

ContentValidationError

Cow

A clone-on-write smart pointer.

DateTimeError

DecodeError

Represents an error ocurred during decoding.

EncodeError

Represents an error occurred during encoding.

EntityType

An enum which represents the different addressable entities.

ManifestAddress

Any address supported by manifest, both global and local.

ManifestCustomExtension

ManifestCustomTraversal

ManifestCustomValue

ManifestCustomValueKind

ManifestExpression

ManifestNonFungibleLocalId

ManifestNonFungibleLocalIdValidationError

ManifestToRustValueError

NonFungibleIdType

Represents type of non-fungible id

NonFungibleLocalId

Represents the local id of a non-fungible.

Option

The Option type. See the module level documentation for more.

OwnValidation

ParseBnumI256Error

ParseBnumI384Error

ParseBnumI512Error

ParseBnumI768Error

ParseBnumU256Error

ParseBnumU384Error

ParseBnumU512Error

ParseBnumU768Error

ParseComponentAddressError

ParseDecimalError

Represents an error when parsing Decimal from another type.

ParseEd25519PublicKeyError

Represents an error when parsing ED25519 public key from hex.

ParseGlobalAddressError

ParseHashError

Represents an error when parsing hash.

ParseInternalAddressError

ParseManifestAddressReservationError

Represents an error when parsing ManifestAddressReservation.

ParseManifestBlobRefError

Represents an error when parsing ManifestBlobRef.

ParseManifestBucketError

Represents an error when parsing ManifestBucket.

ParseManifestDecimalError

Represents an error when parsing ManifestDecimal.

ParseManifestExpressionError

Represents an error when parsing ManifestExpression.

ParseManifestPreciseDecimalError

Represents an error when parsing ManifestPreciseDecimal.

ParseManifestProofError

Represents an error when parsing ManifestProof.

ParseNetworkError

ParseNonFungibleLocalIdError

Represents an error when decoding non-fungible id.

ParseOwnError

ParsePackageAddressError

ParsePreciseDecimalError

Represents an error when parsing PreciseDecimal from another type.

ParseReferenceError

ParseResourceAddressError

ParseSecp256k1PublicKeyError

Represents an error when parsing ED25519 public key from hex.

ParseUtcDateTimeError

PublicKey

Represents any natively supported public key.

PublicKeyHash

The hash of a given public key.

ReferenceValidation

Result

Result is a type that represents either success (Ok) or failure (Err).

RoundingMode

Defines the rounding strategy.

RoyaltyAmount

RustToManifestValueError

SborValue

SchemaPathParseError

SchemaSubPath

ScryptoCustomTraversal

ScryptoCustomTypeKind

A schema for the values that a codec can decode / views as valid

ScryptoCustomTypeValidation

ScryptoCustomValue

ScryptoCustomValueKind

SubstateKey

The unique identifier of a substate within a node module.

TimeComparisonOperator

ValueKind

Constants

ACCESS_CONTROLLER_PACKAGE

The native package for access controllers.

ACCESS_RULES_MODULE_PACKAGE

The native package for the access rules module.

ACCOUNT_OWNER_BADGE

The non-fungible badge resource which is used for account ownership, if accounts have been set up with simple account creation, or have been securified.

ACCOUNT_PACKAGE

The native package for accounts.

CONSENSUS_MANAGER

The consensus manager native component - in charge of validators, consensus and epochs.

CONSENSUS_MANAGER_PACKAGE

The native package for the consensus manager.

DEFAULT_COST_UNIT_LIMIT

The default cost unit limit.

DEFAULT_COST_UNIT_PRICE_IN_XRD

The default cost unit price, in XRD.

DEFAULT_FREE_CREDIT_IN_XRD

The default free credit, for preview only.

DEFAULT_MAX_CALL_DEPTH

The default max call depth, used by transaction executor.

DEFAULT_MAX_COST_UNIT_LIMIT

The maximum value of cost unit limit

DEFAULT_MAX_EPOCH_RANGE

The max epoch range Should be ~ 1 month. The below is ~30 days given 5 minute epochs.

DEFAULT_MAX_EVENT_SIZE

The max event size

DEFAULT_MAX_EXECUTION_TRACE_DEPTH

DEFAULT_MAX_INVOKE_INPUT_SIZE

The default maximum invoke input args size.

DEFAULT_MAX_LOG_SIZE

The max log size

DEFAULT_MAX_METADATA_KEY_STRING_LEN

The max SBOR size of metadata key

DEFAULT_MAX_METADATA_VALUE_SBOR_LEN

The max SBOR size of metadata value

DEFAULT_MAX_NUMBER_OF_EVENTS

The max number of events

DEFAULT_MAX_NUMBER_OF_LOGS

The max number of logs

DEFAULT_MAX_NUMBER_OF_SUBSTATES_IN_HEAP

The default max number of substates in heap.

DEFAULT_MAX_NUMBER_OF_SUBSTATES_IN_TRACK

The default max number of substates in track.

DEFAULT_MAX_PANIC_MESSAGE_SIZE

The max panic message size

DEFAULT_MAX_PER_FUNCTION_ROYALTY_IN_XRD

The default maximum that a package or component owner is allowed to set their method royalty to

DEFAULT_MAX_SUBSTATE_SIZE

The default maximum substate read and write size.

DEFAULT_MAX_TIP_PERCENTAGE

The maximum value of tip percentage

DEFAULT_MAX_TRANSACTION_SIZE

The max transaction size

DEFAULT_MIN_COST_UNIT_LIMIT

The minimum value of cost unit limit

DEFAULT_MIN_TIP_PERCENTAGE

The minimum value of tip percentage

DEFAULT_STATE_EXPANSION_PRICE_IN_XRD

The default price for adding a single byte to the substate store, in XRD.

DEFAULT_SYSTEM_LOAN

The default system loan amount, used by transaction executor.

DEFAULT_TIP_PERCENTAGE

DEFAULT_USD_PRICE_IN_XRD

The default USD price, in XRD

ED25519_SIGNATURE_VIRTUAL_BADGE

The non-fungible badge resource which is used for virtual proofs of EdDSA Ed25519 transacton signatures in the transaction processor.

FAUCET

The faucet native component - use this on testnets for getting XRD and locking fee.

FAUCET_BLUEPRINT

The name of the faucet blueprint under the FAUCET_PACKAGE.

FAUCET_COMPONENT

The faucet native component - use this on testnets for getting XRD and locking fee.

FAUCET_PACKAGE

The scrypto package for the faucet

FEES_PROPOSER_SHARE_PERCENTAGE

The proposer’s share of fees (execution and state expansion)

FEES_VALIDATOR_SET_SHARE_PERCENTAGE

The validator set’s share of fees (execution and state expansion)

GENESIS_HELPER

The genesis helper scrypto component - used for sorting out genesis.

GENESIS_HELPER_BLUEPRINT

The name of the genesis helper blueprint under the GENESIS_HELPER_PACKAGE.

GENESIS_HELPER_PACKAGE

The scrypto package for the genesis helper.

GLOBAL_CALLER_VIRTUAL_BADGE

The non-fungible badge resource which is used for virtual proofs which represent the global ancestor of the actor which made the latest global call.

IDENTITY_OWNER_BADGE

The non-fungible badge resource which is used for identity ownership, if identities have been set up with simple account creation, or have been securified.

IDENTITY_PACKAGE

The native package for identities.

MANIFEST_NON_FUNGIBLE_LOCAL_ID_MAX_LENGTH

MANIFEST_SBOR_V1_MAX_DEPTH

MANIFEST_SBOR_V1_PAYLOAD_PREFIX

0x4d = M in ASCII for Manifest - (77 in decimal)

MANIFEST_VALUE_KIND_ADDRESS

MANIFEST_VALUE_KIND_ADDRESS_RESERVATION

MANIFEST_VALUE_KIND_BLOB

MANIFEST_VALUE_KIND_BUCKET

MANIFEST_VALUE_KIND_DECIMAL

MANIFEST_VALUE_KIND_EXPRESSION

MANIFEST_VALUE_KIND_NON_FUNGIBLE_LOCAL_ID

MANIFEST_VALUE_KIND_PRECISE_DECIMAL

MANIFEST_VALUE_KIND_PROOF

MAX_NUMBER_OF_BLOBS

MAX_NUMBER_OF_INTENT_SIGNATURES

METADATA_MODULE_PACKAGE

The native package for the metadata module.

NATIVE_TOKEN

XRD is the native token of the Radix ledger. It is a fungible token, measured in attos (10^-18).

NON_FUNGIBLE_LOCAL_ID_MAX_LENGTH

PACKAGE_OF_DIRECT_CALLER_VIRTUAL_BADGE

The non-fungible badge resource which is used for virtual proofs which represent the package of the immediate caller - ie the actor which made the latest (global or internal) call.

PACKAGE_OWNER_BADGE

The non-fungible badge resource which is used for package ownership when creating packages with the simple package creation set-up.

PACKAGE_PACKAGE

The native package for package deployment.

POOL_PACKAGE

The native package for pools.

RADIX_TOKEN

XRD is the native token of the Radix ledger. It is a fungible token, measured in attos (10^-18).

RESOURCE_PACKAGE

The native package for resource managers, proofs, buckets, vaults etc.

ROLA_HASHABLE_PAYLOAD_PREFIX

The ROLA hash which is signed is created as hash(ROLA_HASHABLE_PAYLOAD_PREFIX || ..)

ROYALTY_MODULE_PACKAGE

The native package for the royalty module.

SCRYPTO_SBOR_V1_MAX_DEPTH

SCRYPTO_SBOR_V1_PAYLOAD_PREFIX

0x5c for [5c]rypto - (91 in decimal)

SECP256K1_SIGNATURE_VIRTUAL_BADGE

The non-fungible badge resource which is used for virtual proofs of ECDSA Secp256k1 transacton signatures in the transaction processor.

SYSTEM_TRANSACTION_BADGE

The non-fungible badge resource which is used for virtual proofs representing the fact that the current transaction is a system transaction.

TIPS_PROPOSER_SHARE_PERCENTAGE

The proposer’s share of tips

TIPS_VALIDATOR_SET_SHARE_PERCENTAGE

The validator set’s share of tips

TRANSACTION_HASHABLE_PAYLOAD_PREFIX

The Transaction hash which is signed is created as: hash(TRANSACTION_HASHABLE_PAYLOAD_PREFIX || version prefix according to type of transaction payload || ..)

TRANSACTION_PROCESSOR_PACKAGE

The native package for the transaction processor.

TRANSACTION_TRACKER

The intent hash store component

TRANSACTION_TRACKER_BLUEPRINT

TRANSACTION_TRACKER_PACKAGE

TRANSACTION_VERSION_V1

VALIDATOR_OWNER_BADGE

The non-fungible badge resource which is used for validator ownership.

VALUE_KIND_ARRAY

VALUE_KIND_BOOL

VALUE_KIND_DECIMAL

VALUE_KIND_ENUM

VALUE_KIND_I8

VALUE_KIND_I16

VALUE_KIND_I32

VALUE_KIND_I64

VALUE_KIND_I128

VALUE_KIND_MAP

VALUE_KIND_NON_FUNGIBLE_LOCAL_ID

VALUE_KIND_OWN

VALUE_KIND_PRECISE_DECIMAL

VALUE_KIND_REFERENCE

VALUE_KIND_STRING

VALUE_KIND_TUPLE

VALUE_KIND_U8

VALUE_KIND_U16

VALUE_KIND_U32

VALUE_KIND_U64

VALUE_KIND_U128

XRD

XRD is the native token of the Radix ledger. It is a fungible token, measured in attos (10^-18).

Statics

NO_NETWORK

Traits

AsMut

Used to do a cheap mutable-to-mutable reference conversion.

AsRef

Used to do a cheap reference-to-reference conversion.

BasicCategorize

BasicDecode

BasicDescribe

BasicEncode

BasicSbor

Categorize

The Categorize trait marks a rust type as having a fixed value kind for SBOR encoding/decoding.

Cbrt

CheckedMul

CheckedSub

Clone

A common trait for the ability to explicitly duplicate an object.

ContextualDisplay

This trait is used where context is required to correctly display a value.

ContextualSerialize

This trait is used where context is required to correctly serialize a value.

Copy

Types whose values can be duplicated simply by copying bits.

CustomValueKind

Debug

? formatting.

Decode

A data structure that can be decoded from a byte array using SBOR.

Default

A trait for giving a type a useful default value.

Display

Format trait for an empty format, {}.

DoubleEndedIterator

An iterator able to yield elements from both ends.

Drop

Custom code within the destructor.

Encode

A data structure that can be serialized into a byte array using SBOR.

Eq

Trait for equality comparisons which are equivalence relations.

ExactSizeIterator

An iterator that knows its exact length.

Extend

Extend a collection with the contents of an iterator.

Fn

The version of the call operator that takes an immutable receiver.

FnMut

The version of the call operator that takes a mutable receiver.

FnOnce

The version of the call operator that takes a by-value receiver.

From

Used to do value-to-value conversions while consuming the input value. It is the reciprocal of Into.

FromIterator

Conversion from an Iterator.

FromStr

Parse a value from a string

HasPublicKeyHash

Into

A value-to-value conversion that consumes the input value. The opposite of From.

IntoIterator

Conversion into an Iterator.

IsHash

IsNonAutoGeneratedNonFungibleLocalId

Marks the rust type that represents a non-fungible id, of non-auto-generated kind (i.e. String, Integer and Bytes).

IsNonFungibleLocalId

Marks the rust type that represents a non-fungible id, of any kind (i.e. String, Integer, Bytes and RUID).

IsPublicKeyHash

Iterator

A trait for dealing with iterators.

ManifestCategorize

ManifestDecode

ManifestEncode

ManifestSborEnum

ManifestSborTuple

NthRoot

Ord

Trait for types that form a total order.

PartialEq

Trait for equality comparisons.

PartialOrd

Trait for types that form a partial order.

SborCustomValue

SborEnum

Marker trait for encoding as an Sbor Enum

SborTuple

Marker trait for encoding as an Sbor Tuple

ScryptoCategorize

ScryptoDecode

ScryptoDescribe

ScryptoEncode

ScryptoSbor

ScryptoSborEnum

ScryptoSborTuple

Send

Types that can be transferred across thread boundaries.

Sized

Types with a constant size known at compile time.

Sqrt

Sync

Types for which it is safe to share references between threads.

ToOwned

A generalization of Clone to borrowed data.

ToString

A trait for converting a value to a String.

Truncate

TryFrom

Simple and safe type conversions that may fail in a controlled way under some circumstances. It is the reciprocal of TryInto.

TryInto

An attempted conversion that consumes self, which may or may not be expensive.

Unpin

Types that can be safely moved after being pinned.

Functions

basic_decode

Decode an instance of T from a slice.

basic_encode

Encode a T into byte array.

bech32_encode_to_fmt

NOTE: The below code is copied with minor alterations from the bech32 crate. These alterations are to avoid using std for allocations, and fit with the sbor no-alloc options.

blake2b_256_hash

combine

Combines a u8 with a u8 slice.

component_address

copy_u8_array

Copies a slice to a fixed-sized array.

drop

Disposes of a value.

from_manifest_value

hash

Computes the hash digest of a message.

hash_map_new

Creates an empty map with capacity 0 and default Hasher

hash_map_with_capacity

Creates an empty map with given capacity and default Hasher

hash_public_key_bytes

hash_set_new

Creates an empty set with capacity 0 and default Hasher

hash_set_with_capacity

Creates an empty set with given capacity and default Hasher

index_map_new

This is safe for std and no-std use cases (unlike IndexMap::new which disappears when std is not in the toolchain - see this article for deep technical reasons)

index_map_with_capacity

This is safe for std and no-std use cases (unlike IndexMap::with_capacity which disappears when std is not in the toolchain - see this article for deep technical reasons)

index_set_new

This is safe for std and no-std use cases (unlike IndexSet::new which disappears when std is not in the toolchain - see this article for deep technical reasons)

index_set_with_capacity

This is safe for std and no-std use cases (unlike IndexSet::with_capacity which disappears when std is not in the toolchain - see this article for deep technical reasons)

local_address

manifest_decode

manifest_encode

package_address

replace_self_package_address

resource_address

scrypto_decode

Decodes a data structure from a byte array.

scrypto_encode

Encodes a data structure into byte array.

to_manifest_value

Type Definitions

Blake2b256

FieldKey

HashMap

HashSet

IndexMap

IndexSet

ManifestDecoder

ManifestEncoder

ManifestEnumVariantValue

ManifestOwnedRawPayload

ManifestOwnedRawValue

ManifestRawPayload

ManifestRawValue

ManifestTraverser

ManifestValue

ManifestValueKind

MapKey

ScryptoDecoder

ScryptoEncoder

ScryptoOwnedRawPayload

ScryptoOwnedRawValue

ScryptoRawPayload

ScryptoRawValue

ScryptoSchema

ScryptoTraverser

ScryptoTypeData

ScryptoTypeKind

ScryptoValue

ScryptoValueKind

SortedU16Key

ValidatorIndex

An index of a specific validator within the current validator set. To be exact: a ValidatorIndex equal to k references the k-th element returned by the iterator of the IndexMap<ComponentAddress, Validator> in this epoch’s active validator set (which is expected to be sorted by stake, descending). This uniquely identifies the validator, while being shorter than ComponentAddress (we do care about the constant factor of the space taken by LeaderProposalHistory under prolonged liveness break scenarios).

Derive Macros

BasicCategorize

Derive code that returns the value kind - specifically for Basic SBOR.

BasicDecode

Derive code that decodes this data structure from a byte array - specifically for Basic SBOR.

BasicDescribe

Derive code that describes the SBOR type - specifically for Basic SBOR.

BasicEncode

Derive code that encodes this data structure - specifically for Basic SBOR.

BasicSbor

Derive code that implements BasicCategorize, BasicEncode, BasicDecode, and BasicDescribe traits for this struct or enum.

Categorize

Derive code that returns the value kind.

Clone

Derive macro generating an impl of the trait Clone.

Copy

Derive macro generating an impl of the trait Copy.

Debug

Derive macro generating an impl of the trait Debug.

Decode

Derive code that decodes this data structure from a byte array.

Default

Derive macro generating an impl of the trait Default.

Encode

Derive code that encodes this data structure

Eq

Derive macro generating an impl of the trait Eq.

ManifestCategorize

Derives code for categorizing a struct or enum with Manifest value model.

ManifestDecode

Derives code for decoding a struct or enum with Manifest value model.

ManifestEncode

Derives code for encoding a struct or enum with Manifest value model.

ManifestSbor

Derive code that implements ManifestCategorize, ManifestEncode and ManifestDecode traits for this struct or enum.

Ord

Derive macro generating an impl of the trait Ord. The behavior of this macro is described in detail here.

PartialEq

Derive macro generating an impl of the trait PartialEq. The behavior of this macro is described in detail here.

PartialOrd

Derive macro generating an impl of the trait PartialOrd. The behavior of this macro is described in detail here.

Sbor

Derive code that implements Categorize, Encode, Decode, and Describe traits for this struct or enum.

ScryptoCategorize

Derives code for categorizing a struct or enum with Scrypto value model.

ScryptoDecode

Derives code for decoding a struct or enum with Scrypto value model.

ScryptoEncode

Derives code for encoding a struct or enum with Scrypto value model.

ScryptoEvent

Derive code for implementing the required logic to mark a type as being an event.

ScryptoSbor

Derive code that implements ScryptoCategorize, ScryptoEncode, ScryptoDecode, and ScryptoDescribe traits for this struct or enum.