Struct scrypto::prelude::radix_engine_common::prelude::indexmap::map::Slice

pub struct Slice<K, V> {
    pub(crate) entries: [Bucket<K, V>],
}

A dynamically-sized slice of key-value pairs in an IndexMap.

This supports indexed operations much like a [(K, V)] slice, but not any hashed operations on the map keys.

Unlike IndexMap, Slice does consider the order for PartialEq and Eq, and it also implements PartialOrd, Ord, and Hash.

Fields
entries: [Bucket<K, V>]
Implementations

impl<K, V> Slice<K, V>

pub fn len(&self) -> usize

Return the number of key-value pairs in the map slice.

pub fn is_empty(&self) -> bool

Returns true if the map slice contains no elements.

pub fn get_index(&self, index: usize) -> Option<(&K, &V)>

Get a key-value pair by index.

Valid indices are 0 <= index < self.len()

pub fn get_index_mut(&mut self, index: usize) -> Option<(&K, &mut V)>

Get a key-value pair by index, with mutable access to the value.

Valid indices are 0 <= index < self.len()

pub fn get_range<R>(&self, range: R) -> Option<&Slice<K, V>>where R: RangeBounds<usize>,

Returns a slice of key-value pairs in the given range of indices.

Valid indices are 0 <= index < self.len()

pub fn get_range_mut<R>(&mut self, range: R) -> Option<&mut Slice<K, V>>where R: RangeBounds<usize>,

Returns a mutable slice of key-value pairs in the given range of indices.

Valid indices are 0 <= index < self.len()

pub fn first(&self) -> Option<(&K, &V)>

Get the first key-value pair.

pub fn first_mut(&mut self) -> Option<(&K, &mut V)>

Get the first key-value pair, with mutable access to the value.

pub fn last(&self) -> Option<(&K, &V)>

Get the last key-value pair.

pub fn last_mut(&mut self) -> Option<(&K, &mut V)>

Get the last key-value pair, with mutable access to the value.

pub fn split_at(&self, index: usize) -> (&Slice<K, V>, &Slice<K, V>)

Divides one slice into two at an index.

Panics if index > len.

pub fn split_at_mut( &mut self, index: usize ) -> (&mut Slice<K, V>, &mut Slice<K, V>)

Divides one mutable slice into two at an index.

Panics if index > len.

pub fn split_first(&self) -> Option<((&K, &V), &Slice<K, V>)>

Returns the first key-value pair and the rest of the slice, or None if it is empty.

pub fn split_first_mut(&mut self) -> Option<((&K, &mut V), &mut Slice<K, V>)>

Returns the first key-value pair and the rest of the slice, with mutable access to the value, or None if it is empty.

pub fn split_last(&self) -> Option<((&K, &V), &Slice<K, V>)>

Returns the last key-value pair and the rest of the slice, or None if it is empty.

pub fn split_last_mut(&mut self) -> Option<((&K, &mut V), &mut Slice<K, V>)>

Returns the last key-value pair and the rest of the slice, with mutable access to the value, or None if it is empty.

pub fn iter(&self) -> Iter<'_, K, V>

Return an iterator over the key-value pairs of the map slice.

pub fn iter_mut(&mut self) -> IterMut<'_, K, V>

Return an iterator over the key-value pairs of the map slice.

pub fn keys(&self) -> Keys<'_, K, V>

Return an iterator over the keys of the map slice.

pub fn into_keys(self: Box<Slice<K, V>, Global>) -> IntoKeys<K, V>

Return an owning iterator over the keys of the map slice.

pub fn values(&self) -> Values<'_, K, V>

Return an iterator over the values of the map slice.

pub fn values_mut(&mut self) -> ValuesMut<'_, K, V>

Return an iterator over mutable references to the the values of the map slice.

pub fn into_values(self: Box<Slice<K, V>, Global>) -> IntoValues<K, V>

Return an owning iterator over the values of the map slice.