Module scrypto::prelude::radix_engine_common::prelude::schema

SBOR Schema

Modules

basic_well_known_types

prelude

Structs

ArrayData

EnumVariantData

LengthValidation

Represents additional validation that should be performed on the size.

MapData

NovelTypeMetadata

This is the struct used in the Schema

NumericValidation

Represents additional validation that should be performed on the numeric value.

SchemaContext

SchemaV1

An array of custom type kinds, and associated extra information which can attach to the type kinds

TupleData

TypeAggregator

TypeData

Combines all data about a Type:

TypeMetadata

This enables the type to be represented as eg JSON Also used to facilitate type reconstruction

WellKnownTypeId

Enums

ChildNames

LocalTypeId

The TypeId which is local to a given Schema. This is the SchemaTypeLink used in a linearized Schema to link TypeKinds.

RustTypeId

This is a compile-time identifier for a given type, used by the type aggregator to uniquely identify a type.

SchemaValidationError

TypeKind

A schema for the values that a codec can decode / views as valid

TypeValidation

Additional validation to apply to a payload of the given type, beyond validation from the TypeKind’s type structure.

VersionedSchema

Constants

MAX_NUMBER_OF_FIELDS

Traits

CustomExtension

CustomSchema

CustomTypeKind

CustomTypeValidation

Describe

The Describe trait allows a type to describe how to interpret and validate a corresponding SBOR payload.

SchemaTypeLink

Marker trait for a link between TypeKinds:

VersionedSchemaVersion

Functions

generate_full_schema

generate_full_schema_from_single_type

localize_well_known

localize_well_known_type_data

resolve_local_type_id

validate_childless_metadata

validate_custom_type_validation

validate_enum_metadata

validate_field_names

validate_index

validate_schema

validate_schema_enum_variant_name

validate_schema_field_name

validate_schema_type_name

validate_tuple_metadata

validate_type_kind

validate_type_metadata_with_type_kind

Type Aliases

Schema

SchemaTypeKind

TypeHash