Struct scrypto::prelude::radix_engine_common::prelude::UtcDateTime

pub struct UtcDateTime {
    year: u32,
    month: u8,
    day_of_month: u8,
    hour: u8,
    minute: u8,
    second: u8,
}

A UtcDateTime represents a Unix timestamp on the UTC Calendar.

It can represent any date between 1-1-1 00:00:00 and [u32::MAX]-12-31 23:59:59 (inclusive).

In terms of indexing:

Months and days of month are 1-based (i.e. Dec 15th 2022 corresponds to 2022-12-15).

Hour, minute and second are 0-based, based on the 24-hour clock. Midnight is represented as 00:00:00 and 23:59:59 is the last second before midnight. Following Unix timstamp conventions, leap seconds are not supported.

UtcDateTime supports methods for easy conversion to and from the Instant type, which can be queried from the Radix Engine.

Fields
year: u32month: u8day_of_month: u8hour: u8minute: u8second: u8
Implementations

impl UtcDateTime

pub fn new( year: u32, month: u8, day_of_month: u8, hour: u8, minute: u8, second: u8 ) -> Result<UtcDateTime, DateTimeError>

Creates a UtcDateTime from its individual components.

Months and days of month are 1-based, with the following limits:

Valid year range is: 1 to u32::MAXinclusive.

Valid month values are: 1 to 12 inclusive.

Valid day of month values are: 1 to {28, 29, 30, 31} inclusive. The upper limit depends on the month and year, as per the Gregorian calendar.

An attempt to create an invalid date (e.g. 2022-04-31 or 2023-02-29) will result in a corresponding Err(DateTimeError).

Hour, minute and second constitute a 0-based 24-hour clock. Midnight is represented as 00:00:00 and 23:59:59 is the maximum possible time (a second before midnight). Following Unix time conventions, leap seconds are not represented in this system.

pub fn from_instant(instant: &Instant) -> Result<UtcDateTime, DateTimeError>

Creates a UtcDateTime from an Instant.

The minimum supported seconds_since_unix_epoch value is -62135596800 (corresponding to 1-1-1 00:00:00) and the maximum value is 135536014634284799 (corresponding to [u32::Max]-12-31 23:59:59).

pub fn to_instant(&self) -> Instant

Creates an Instant from this UtcDateTime

pub fn year(&self) -> u32

pub fn month(&self) -> u8

pub fn day_of_month(&self) -> u8

pub fn hour(&self) -> u8

pub fn minute(&self) -> u8

pub fn second(&self) -> u8

pub fn add_days(&self, days_to_add: i64) -> Option<UtcDateTime>

pub fn add_hours(&self, hours_to_add: i64) -> Option<UtcDateTime>

pub fn add_minutes(&self, minutes_to_add: i64) -> Option<UtcDateTime>

pub fn add_seconds(&self, seconds_to_add: i64) -> Option<UtcDateTime>