Struct scrypto::prelude::BnumU384

pub struct BnumU384(pub BUint<6>);

Provides safe integer arithmetic.

Operations like +, ‘-’, ‘*’, or ‘/’ sometimes produce overflow which is detected and results in a panic, in of silently wrapping around.

The bit length of output type will be the greater one in the math operation, and if any of the types was signed, then the resulting type will be signed too, otherwise the output type is unsigned.

The underlying value can be retrieved through the .0 index of the BnumU384 tuple.

Layout

BnumU384 will have the same methods and traits as the built-in counterpart.

Tuple Fields
0: BUint<6>
Implementations

impl BnumU384

pub const fn count_ones(self) -> u32

Returns the number of ones in the binary representation of self.

pub const fn count_zeros(self) -> u32

Returns the number of zeros in the binary representation of self.

pub const fn trailing_zeros(self) -> u32

Returns the number of trailing zeros in the binary representation of self.

pub const fn swap_bytes(self) -> BnumU384

Reverses the byte order of the integer.

pub const fn reverse_bits(self) -> BnumU384

Reverses the bit pattern of the integer.

pub const fn leading_zeros(self) -> u32

Returns the number of leading zeros in the binary representation of self.

pub fn rotate_left(self, other: BnumU384) -> BnumU384

Shifts the bits to the left by a specified amount, n, wrapping the truncated bits to the end of the resulting integer.

Please note this isn’t the same operation as the << shifting operator! This method can not overflow as opposed to ‘<<’.

Please note that this example is shared between integer types. Which explains why I128 is used here.

pub fn rotate_right(self, other: BnumU384) -> BnumU384

Shifts the bits to the right by a specified amount, n, wrapping the truncated bits to the beginning of the resulting integer.

Please note this isn’t the same operation as the >> shifting operator! This method can not overflow as opposed to ‘>>’.

pub const fn from_be(x: BnumU384) -> BnumU384

Converts an integer from big endian to the target’s endianness.

On big endian this is a no-op. On little endian the bytes are swapped.

pub const fn from_le(x: BnumU384) -> BnumU384

Converts an integer from little endian to the target’s endianness.

On little endian this is a no-op. On big endian the bytes are swapped.

pub const fn to_be(self) -> BnumU384

Converts self to big endian from the target’s endianness.

On big endian this is a no-op. On little endian the bytes are swapped.

pub const fn to_le(self) -> BnumU384

Converts self to little endian from the target’s endianness.

On little endian this is a no-op. On big endian the bytes are swapped.

impl BnumU384

pub fn from_le_bytes(bytes: &[u8]) -> BnumU384

impl BnumU384

pub fn to_le_bytes(&self) -> [u8; 48]

pub fn to_vec(&self) -> Vec<u8, Global>

impl BnumU384

pub const fn from_digits(digits: [u64; 6]) -> BnumU384

impl BnumU384

pub const MIN: BnumU384 = Self(BUint<6>::MIN)

pub const MAX: BnumU384 = Self(BUint<6>::MAX)

pub const ZERO: BnumU384 = Self(BUint<6>::ZERO)

pub const ONE: BnumU384 = Self(BUint<6>::ONE)

pub const TEN: BnumU384 = Self(BUint<6>::TEN)

pub const BITS: u32 = 384u32

pub const BYTES: u32 = 48u32

pub const N: usize = 6usize

impl BnumU384

pub fn is_power_of_two(self) -> bool

pub fn next_power_of_two(self) -> BnumU384