Struct scrypto::prelude::Decimal

pub struct Decimal(pub I192);

Decimal represents a 192 bit representation of a fixed-scale decimal number.

The finite set of values are of the form m / 10^18, where m is an integer such that -2^(192 - 1) <= m < 2^(192 - 1).

Fractional part: ~60 bits/18 digits Integer part : 132 bits /40 digits Max : 3138550867693340381917894711603833208051.177722232017256447 Min : -3138550867693340381917894711603833208051.177722232017256448

Unless otherwise specified, all operations will panic if underflow/overflow.

Tuple Fields
0: I192
Implementations

impl Decimal

pub const MIN: Decimal = _

The min value of Decimal.

pub const MAX: Decimal = _

The max value of Decimal.

pub const BITS: usize = 192usize

The bit length of number storing Decimal.

pub const SCALE: u32 = 18u32

The fixed scale used by Decimal.

pub const ZERO: Decimal = _

pub const ONE_HUNDREDTH: Decimal = _

pub const ONE_TENTH: Decimal = _

pub const ONE: Decimal = _

pub const TEN: Decimal = _

pub const ONE_HUNDRED: Decimal = _

pub const fn zero() -> Decimal

Returns Decimal of 0.

pub const fn one() -> Decimal

Returns Decimal of 1.

pub fn is_zero(&self) -> bool

Whether this decimal is zero.

pub fn is_positive(&self) -> bool

Whether this decimal is positive.

pub fn is_negative(&self) -> bool

Whether this decimal is negative.

pub fn checked_abs(&self) -> Option<Decimal>

Returns the absolute value.

pub fn checked_floor(&self) -> Option<Decimal>

Returns the largest integer that is equal to or less than this number.

pub fn checked_ceiling(&self) -> Option<Decimal>

Returns the smallest integer that is equal to or greater than this number.

pub fn checked_round<T>( &self, decimal_places: T, mode: RoundingMode ) -> Option<Decimal>
where T: Into<i32>,

Rounds this number to the specified decimal places.

Panics

Panic if the number of decimal places is not within [0..SCALE]

pub fn checked_powi(&self, exp: i64) -> Option<Decimal>

Calculates power using exponentiation by squaring“.

pub fn checked_sqrt(&self) -> Option<Decimal>

Square root of a Decimal

pub fn checked_cbrt(&self) -> Option<Decimal>

Cubic root of a Decimal

pub fn checked_nth_root(&self, n: u32) -> Option<Decimal>

Nth root of a Decimal

impl Decimal

pub fn to_vec(&self) -> Vec<u8>