Trait scrypto::resource::proof::ScryptoUncheckedProof

pub trait ScryptoUncheckedProof {
    type CheckedProofType;

    // Required methods
    fn check(
        self,
        expected_resource_address: ResourceAddress
    ) -> Self::CheckedProofType;
    fn check_with_message<S: ToString>(
        self,
        expected_resource_address: ResourceAddress,
        custom_error_message: S
    ) -> Self::CheckedProofType;
    fn skip_checking(self) -> Self::CheckedProofType;
    fn resource_address(&self) -> ResourceAddress;
    fn drop(self);
    fn clone(&self) -> Self;
    fn authorize<F: FnOnce() -> O, O>(&self, f: F) -> O;

    // Provided method
    fn resource_manager(&self) -> ResourceManager { ... }
}

Required Associated Types

type CheckedProofType

Required Methods

fn check( self, expected_resource_address: ResourceAddress ) -> Self::CheckedProofType

Checks the resource address of this proof and panics if it’s unexpected.

fn check_with_message<S: ToString>( self, expected_resource_address: ResourceAddress, custom_error_message: S ) -> Self::CheckedProofType

Checks the resource address of this proof and panics with custom error message if it’s unexpected.

fn skip_checking(self) -> Self::CheckedProofType

Skips checking and converts this proof into a “checked” proof.

Warning!

Be sure to validate the resource address before reading data from the proof in your custom validation logic!

fn resource_address(&self) -> ResourceAddress

fn drop(self)

fn clone(&self) -> Self

fn authorize<F: FnOnce() -> O, O>(&self, f: F) -> O

Provided Methods

fn resource_manager(&self) -> ResourceManager

Object Safety

This trait is not object safe.