Module scrypto::resource::resource_builder

Structs

FungibleResourceBuilder

A resource builder which builds fungible resources that may or may not have an owner badge.

FungibleResourceWithAuthBuilder

A resource builder which builds fungible resources that do not have an owner-badge where resource behavior is completely handled by the developer.

NonFungibleResourceBuilder

A resource builder which builds non-fungible resources that may or may not have an owner badge.

NonFungibleResourceWithAuthBuilder

A resource builder which builds non-fungible resources that do not have an owner-badge where resource behavior is completely handled by the developer.

ResourceBuilder

Utility for setting up a new resource.

Constants

DIVISIBILITY_MAXIMUM

The maximum divisibility supported.

DIVISIBILITY_NONE

Not divisible.