Struct scrypto::resource::resource_builder::InProgressResourceBuilder

pub struct InProgressResourceBuilder<T: ResourceType, A: ConfiguredAuth> {
    resource_type: T,
    metadata: BTreeMap<String, String>,
    auth: A,
}

Utility for setting up a new resource, which has building in progress.

You start the building process with one of the methods starting with ResourceBuilder::new_.

The allowed methods change depending on which methods have already been called. For example, you can either use owner_non_fungible_badge or set access rules individually, but not both.

You can complete the building process using either create_with_no_initial_supply() or mint_initial_supply(..).

Example

use scrypto::prelude::*;

let bucket = ResourceBuilder::new_fungible()
    .metadata("name", "TestToken")
    .mint_initial_supply(5);

Fields
resource_type: Tmetadata: BTreeMap<String, String>auth: A
Implementations

impl<A: ConfiguredAuth> InProgressResourceBuilder<FungibleResourceType, A>

pub fn divisibility(self, divisibility: u8) -> Self

Set the resource’s divisibility: the number of digits of precision after the decimal point in its balances.

0 means the resource is not divisible (balances are always whole numbers)

18 is the maximum divisibility, and the default.

Examples

use scrypto::prelude::*;

// Only permits whole-number balances.
ResourceBuilder::new_fungible()
   .divisibility(0);

// Only permits balances to 3 decimal places.
ResourceBuilder::new_fungible()
   .divisibility(3);

impl<A: ConfiguredAuth> InProgressResourceBuilder<FungibleResourceType, A>

pub fn mint_initial_supply<T: Into<Decimal>>(self, amount: T) -> Bucket

Creates resource with the given initial supply.

Example

use scrypto::prelude::*;

let bucket = ResourceBuilder::new_fungible()
    .mint_initial_supply(5);

impl<A: ConfiguredAuth> InProgressResourceBuilder<NonFungibleResourceType<StringNonFungibleLocalId>, A>

pub fn mint_initial_supply<T, V>(self, entries: T) -> Bucketwhere
T: IntoIterator<Item = (StringNonFungibleLocalId, V)>,
V: NonFungibleData,

Creates the non-fungible resource, and mints an individual non-fungible for each key/data pair provided.

Example

use scrypto::prelude::*;

#[derive(NonFungibleData)]
struct NFData {
    pub name: String,
    #[mutable]
    pub flag: bool,
}

let bucket = ResourceBuilder::new_string_non_fungible()
    .mint_initial_supply([
        ("One".try_into().unwrap(), NFData { name: "NF One".to_owned(), flag: true }),
        ("Two".try_into().unwrap(), NFData { name: "NF Two".to_owned(), flag: true }),
    ]);

impl<A: ConfiguredAuth> InProgressResourceBuilder<NonFungibleResourceType<IntegerNonFungibleLocalId>, A>

pub fn mint_initial_supply<T, V>(self, entries: T) -> Bucketwhere
T: IntoIterator<Item = (IntegerNonFungibleLocalId, V)>,
V: NonFungibleData,

Creates the non-fungible resource, and mints an individual non-fungible for each key/data pair provided.

Example

use scrypto::prelude::*;

#[derive(NonFungibleData)]
struct NFData {
    pub name: String,
    #[mutable]
    pub flag: bool,
}

let bucket = ResourceBuilder::new_integer_non_fungible()
    .mint_initial_supply([
        (1u64.into(), NFData { name: "NF One".to_owned(), flag: true }),
        (2u64.into(), NFData { name: "NF Two".to_owned(), flag: true }),
    ]);

impl<A: ConfiguredAuth> InProgressResourceBuilder<NonFungibleResourceType<BytesNonFungibleLocalId>, A>

pub fn mint_initial_supply<T, V>(self, entries: T) -> Bucketwhere
T: IntoIterator<Item = (BytesNonFungibleLocalId, V)>,
V: NonFungibleData,

Creates the non-fungible resource, and mints an individual non-fungible for each key/data pair provided.

Example

use scrypto::prelude::*;

#[derive(NonFungibleData)]
struct NFData {
    pub name: String,
    #[mutable]
    pub flag: bool,
}

let bucket = ResourceBuilder::new_bytes_non_fungible()
    .mint_initial_supply([
        (vec![1u8].try_into().unwrap(), NFData { name: "NF One".to_owned(), flag: true }),
        (vec![2u8].try_into().unwrap(), NFData { name: "NF Two".to_owned(), flag: true }),
    ]);

impl<A: ConfiguredAuth> InProgressResourceBuilder<NonFungibleResourceType<UUIDNonFungibleLocalId>, A>

pub fn mint_initial_supply<T, V>(self, entries: T) -> Bucketwhere
T: IntoIterator<Item = V>,
V: NonFungibleData,

Creates the UUID non-fungible resource, and mints an individual non-fungible for each piece of data provided.

The system automatically generates a new UUID NonFungibleLocalId for each non-fungible, and assigns the given data to each.

Example

use scrypto::prelude::*;

#[derive(NonFungibleData)]
struct NFData {
    pub name: String,
    #[mutable]
    pub flag: bool,
}

let bucket = ResourceBuilder::new_uuid_non_fungible()
    .mint_initial_supply([
        (NFData { name: "NF One".to_owned(), flag: true }),
        (NFData { name: "NF Two".to_owned(), flag: true }),
    ]);