Struct scrypto::resource::resource_builder::NonFungibleResourceBuilder

pub struct NonFungibleResourceBuilder {
    metadata: HashMap<String, String>,
    id_type: NonFungibleIdType,
}

A resource builder which builds non-fungible resources that may or may not have an owner badge.

Note

Once one of the methods that set behavior is called, a new NonFungibleResourceWithAuthBuilder is created which commits the developer to building a resource that does not have an owner badge. If none of these methods are called, then the developer has the choice to either building a resource with an owner badge or without one.

Fields
metadata: HashMap<String, String>id_type: NonFungibleIdType
Implementations

impl NonFungibleResourceBuilder

pub fn new(id_type: NonFungibleIdType) -> Self

pub fn metadata<K: AsRef<str>, V: AsRef<str>>(self, name: K, value: V) -> Self

Adds a resource metadata.

If a previous attribute with the same name has been set, it will be overwritten.

pub fn mintable<R: Into<AccessRule>>(
self,
method_auth: AccessRule,
mutability: R
) -> NonFungibleResourceWithAuthBuilder

pub fn burnable<R: Into<AccessRule>>(
self,
method_auth: AccessRule,
mutability: R
) -> NonFungibleResourceWithAuthBuilder

pub fn recallable<R: Into<AccessRule>>(
self,
method_auth: AccessRule,
mutability: R
) -> NonFungibleResourceWithAuthBuilder

pub fn restrict_withdraw<R: Into<AccessRule>>(
self,
method_auth: AccessRule,
mutability: R
) -> NonFungibleResourceWithAuthBuilder

pub fn restrict_deposit<R: Into<AccessRule>>(
self,
method_auth: AccessRule,
mutability: R
) -> NonFungibleResourceWithAuthBuilder

pub fn updateable_metadata<R: Into<AccessRule>>(
self,
method_auth: AccessRule,
mutability: R
) -> NonFungibleResourceWithAuthBuilder

pub fn updateable_non_fungible_data<R: Into<AccessRule>>(
self,
method_auth: AccessRule,
mutability: R
) -> NonFungibleResourceWithAuthBuilder

pub fn initial_supply<T, V>(self, entries: T) -> Bucketwhere
T: IntoIterator<Item = (NonFungibleId, V)>,
V: NonFungibleData,

Creates resource with the given initial supply.

Example

let bucket = ResourceBuilder::new_non_fungible(NonFungibleIdType::UUID)
    .metadata("name", "TestNonFungible")
    .initial_supply([
        (NonFungibleId::from(1u128), "immutable_part", "mutable_part"),
        (NonFungibleId::from(2u128), "another_immutable_part", "another_mutable_part"),
    ]);

pub fn no_initial_supply(self) -> ResourceAddress

Creates resource with no initial supply.

fn build(
self,
mint_params: Option<MintParams>
) -> (ResourceAddress, Option<Bucket>)

pub fn initial_supply_with_owner<T, V>(
self,
entries: T,
owner_badge: NonFungibleAddress
) -> Bucketwhere
T: IntoIterator<Item = (NonFungibleId, V)>,
V: NonFungibleData,

pub fn no_initial_supply_with_owner(
self,
owner_badge: NonFungibleAddress
) -> ResourceAddress