Struct scrypto::resource::resource_builder::NonFungibleResourceWithAuthBuilder

pub struct NonFungibleResourceWithAuthBuilder {
    id_type: NonFungibleIdType,
    metadata: HashMap<String, String>,
    authorization: HashMap<ResourceMethodAuthKey, (AccessRule, AccessRule)>,
}

A resource builder which builds non-fungible resources that do not have an owner-badge where resource behavior is completely handled by the developer.

Typically this resource builder is created from the NonFungibleResourceBuilder once the developer has called one of the methods that set resource behavior. This is done as a static way of committing the developer to this choice by transitioning into a builder which does not offer the initial_supply_with_owner and no_initial_supply_with_owner methods.

Fields
id_type: NonFungibleIdTypemetadata: HashMap<String, String>authorization: HashMap<ResourceMethodAuthKey, (AccessRule, AccessRule)>
Implementations

impl NonFungibleResourceWithAuthBuilder

pub fn metadata<K: AsRef<str>, V: AsRef<str>>(self, name: K, value: V) -> Self

Adds a resource metadata.

If a previous attribute with the same name has been set, it will be overwritten.

pub fn mintable<R: Into<AccessRule>>(
self,
method_auth: AccessRule,
mutability: R
) -> Self

pub fn burnable<R: Into<AccessRule>>(
self,
method_auth: AccessRule,
mutability: R
) -> Self

pub fn recallable<R: Into<AccessRule>>(
self,
method_auth: AccessRule,
mutability: R
) -> Self

pub fn restrict_withdraw<R: Into<AccessRule>>(
self,
method_auth: AccessRule,
mutability: R
) -> Self

pub fn restrict_deposit<R: Into<AccessRule>>(
self,
method_auth: AccessRule,
mutability: R
) -> Self

pub fn updateable_metadata<R: Into<AccessRule>>(
self,
method_auth: AccessRule,
mutability: R
) -> Self

pub fn updateable_non_fungible_data<R: Into<AccessRule>>(
self,
method_auth: AccessRule,
mutability: R
) -> Self

pub fn initial_supply<T, V>(self, entries: T) -> Bucketwhere
T: IntoIterator<Item = (NonFungibleId, V)>,
V: NonFungibleData,

pub fn no_initial_supply(self) -> ResourceAddress

Creates resource with no initial supply.

fn build(
self,
mint_params: Option<MintParams>
) -> (ResourceAddress, Option<Bucket>)