Struct scrypto::resource::Bucket

pub struct Bucket { /* fields omitted */ }

Represents a transient resource container.

Implementations

impl Bucket

pub fn new<A: Into<ResourceDef>>(resource_def: A) -> Self

Creates a new bucket to hold resources of the given definition.

pub fn put(&self, other: Self)

Puts resources from another bucket into this bucket.

pub fn take<A: Into<Decimal>>(&self, amount: A) -> Self

Takes some amount of resources from this bucket.

pub fn present(&self) -> BucketRef

Creates an immutable reference to this bucket.

pub fn amount(&self) -> Decimal

Returns the amount of resources in this bucket.

pub fn resource_def(&self) -> ResourceDef

Returns the resource definition of resources in this bucket.

pub fn resource_address(&self) -> Address

Returns the resource definition address.

pub fn burn(self)

Burns resource within this bucket.

pub fn burn_with_auth(self, auth: BucketRef)

Burns resource within this bucket.

pub fn is_empty(&self) -> bool

Checks if this bucket is empty.

pub fn authorize<F: FnOnce(BucketRef) -> O, O>(&self, f: F) -> O

Uses resources in this bucket as authorization for an operation.

pub fn take_nft(&self, id: u128) -> Bucket

Takes an NFT from this bucket, by id.

Panics

Panics if this is not an NFT bucket or the specified NFT is not found.

pub fn get_nfts<T: NftData>(&self) -> Vec<Nft<T>>

Returns all the NFT units contained.

Panics

Panics if this is not an NFT bucket.

pub fn get_nft_ids(&self) -> Vec<u128>

Get all NFT IDs in this bucket.

Panics

Panics if this is not an NFT bucket.

pub fn get_nft_id(&self) -> u128

Get the NFT id and panic if not singleton.

pub fn get_nft_data<T: NftData>(&self, id: u128) -> T

Returns the data of an NFT unit, both the immutable and mutable parts.

Panics

Panics if this is not an NFT bucket.

pub fn update_nft_data<T: NftData>(
&self,
id: u128,
new_data: T,
auth: BucketRef
)

Updates the mutable part of the data of an NFT unit.

Panics

Panics if this is not an NFT bucket or the specified NFT is not found.