Module scrypto::runtime

Scrypto runtime abstraction.

Modules

clock 🔒

data 🔒

local_auth_zone 🔒

logger 🔒

runtime 🔒

Structs

BytesNonFungibleLocalId

Bytes, of length between 1 and 64.

Clock

The system clock

ComponentStatePointer

DataRef

DataRefMut

IntegerNonFungibleLocalId

Unsigned integers, up to u64.

LocalAuthZone

Represents the auth zone, which is used by system for checking if this component is allowed to

Logger

A utility for logging messages.

Own

RUIDNonFungibleLocalId

RUID, v4, variant 1, big endian. See https://www.rfc-editor.org/rfc/rfc4122

Reference

Runtime

The transaction runtime.

StringNonFungibleLocalId

A string matching [_0-9a-zA-Z]{1,64}.

Enums

ContentValidationError

DataOrigin

NonFungibleIdType

Represents type of non-fungible id

NonFungibleLocalId

Represents the local id of a non-fungible.

ParseNonFungibleLocalIdError

Represents an error when decoding non-fungible id.

ParseOwnError

ParseReferenceError

Constants

NON_FUNGIBLE_LOCAL_ID_MAX_LENGTH

Traits

IsNonAutoGeneratedNonFungibleLocalId

Marks the rust type that represents a non-fungible id, of non-auto-generated kind (i.e. String, Integer and Bytes).

IsNonFungibleLocalId

Marks the rust type that represents a non-fungible id, of any kind (i.e. String, Integer, Bytes and RUID).