Module scrypto::types

Modules

additional_attributes

Documentation for Additional Attributes

Structs

BlueprintId

BlueprintInfo

Core object state, persisted in TypeInfoSubstate.

BlueprintTypeIdentifier

Buffer

ComponentAddress

Address to a global component

ComponentRoyaltyConfig

Royalty rules

Epoch

A type-safe consensus epoch number.

EventTypeIdentifier

Identifies a specific event schema emitter by some emitter RENode.

FnIdentifier

FullyScopedTypeId

An identifier for a type under a given node’s schema context in the Radix network.

GlobalAddress

Address to a global entity

GlobalAddressPhantom

GlobalAddressReservation

IndexedScryptoValue

InternalAddress

Address to a local entity

KeyValueStoreInfo

KeyValueStoreInit

KeyValueStoreInitEntry

NodeId

The unique identifier of a (stored) node.

ObjectInfo

PackageAddress

Address to a global package

PartitionNumber

The unique identifier of a node module.

PartitionOffset

ResourceAddress

Address to a global resource

Round

A type-safe consensus round number within a single epoch.

SchemaHash

Represents a particular schema under a package

ScopedTypeId

An identifier for a type in the context of a schema.

Slice

Enums

AccessControllerPartitionOffset

AccountPartitionOffset

AuthZoneField

BootLoaderField

ClientCostingEntry

ComponentField

ComponentRoyaltyPartitionOffset

ConsensusManagerPartitionOffset

Emitter

EntityType

An enum which represents the different addressable entities.

FungibleBucketField

FungibleProofField

FungibleResourceManagerPartitionOffset

FungibleVaultPartitionOffset

GenericSubstitution

A reference to the type to substitute with for the case of generics.

Level

Represents the level of a log message.

MetadataPartitionOffset

MultiResourcePoolPartitionOffset

NonFungibleBucketField

NonFungibleProofField

NonFungibleResourceManagerPartitionOffset

NonFungibleVaultPartitionOffset

ObjectType

OneResourcePoolPartitionOffset

OuterObjectInfo

PackageCode

PackagePartitionOffset

PackageRoyalty

Royalty rules

PackageRoyaltyConfig

Royalty rules

ParseComponentAddressError

ParseError

The ParseError enum is a collection of all the possible reasons an enum can fail to parse from a string.

ParseGlobalAddressError

ParseInternalAddressError

ParsePackageAddressError

ParseResourceAddressError

RoleAssignmentPartitionOffset

RoyaltyAmount

RoyaltyField

SubstateKey

The unique identifier of a substate within a node module.

TransactionTrackerField

TwoResourcePoolPartitionOffset

TypeInfoField

ValidatorPartitionOffset

WorktopField

Constants

BOOT_LOADER_PARTITION

Used only with TRANSACTION_TRACKER Node for boot loading

BOOT_LOADER_RESERVED_NODE_ID_FIRST_BYTE

MAIN_BASE_PARTITION

METADATA_BASE_PARTITION

METADATA_KV_STORE_PARTITION_OFFSET

ROLE_ASSIGNMENT_BASE_PARTITION

ROLE_ASSIGNMENT_FIELDS_PARTITION

ROLE_ASSIGNMENT_FIELDS_PARTITION_OFFSET

ROLE_ASSIGNMENT_MUTABILITY_PARTITION_OFFSET

ROLE_ASSIGNMENT_ROLE_DEF_PARTITION

ROLE_ASSIGNMENT_ROLE_DEF_PARTITION_OFFSET

ROYALTY_BASE_PARTITION

ROYALTY_CONFIG_PARTITION

ROYALTY_CONFIG_PARTITION_OFFSET

ROYALTY_FIELDS_PARTITION

ROYALTY_FIELDS_PARTITION_OFFSET

SCHEMAS_PARTITION

TYPE_INFO_FIELD_PARTITION

Traits

AsStaticRefDeprecated

A cheap reference-to-reference conversion. Used to convert a value to a reference value with 'static lifetime within generic code.

BlueprintPartitionOffset

CollectionDescriptor

EnumCount

A trait for capturing the number of variants in Enum. This trait can be autoderived by strum_macros.

EnumMessage

Associates additional pieces of information with an Enum. This can be autoimplemented by deriving EnumMessage and annotating your variants with #[strum(message="...")].

EnumProperty

EnumProperty is a trait that makes it possible to store additional information with enum variants. This trait is designed to be used with the macro of the same name in the strum_macros crate. Currently, the only string literals are supported in attributes, the other methods will be implemented as additional attribute types become stabilized.

FieldDescriptor

IntoEnumIterator

This trait designates that an Enum can be iterated over. It can be auto generated using strum_macros on your behalf.

NonFungibleData

Represents the data structure of a non-fungible.

VariantNames

A trait for retrieving the names of each variant in Enum. This trait can be autoderived by strum_macros.

Functions

component_address

local_address

package_address

resource_address

Type Aliases

BufferId

FieldKey

MapKey

SortedKey

SubstateHandle

ValidatorIndex

An index of a specific validator within the current validator set. To be exact: a ValidatorIndex equal to k references the k-th element returned by the iterator of the IndexMap<ComponentAddress, Validator> in this epoch’s active validator set (which is expected to be sorted by stake, descending). This uniquely identifies the validator, while being shorter than ComponentAddress (we do care about the constant factor of the space taken by LeaderProposalHistory under prolonged liveness break scenarios).

Derive Macros

AsRefStr

Converts enum variants to &'static str.

AsStaticStr

Display

Converts enum variants to strings.

EnumCount

Add a constant usize equal to the number of variants.

EnumDiscriminants

Generate a new type with only the discriminant names.

EnumIter

Creates a new type that iterates of the variants of an enum.

EnumMessage

Add a verbose message to an enum variant.

EnumProperty

Add custom properties to enum variants.

EnumString

Converts strings to enum variants based on their name.

EnumVariantNames

Implements Strum::VariantNames which adds an associated constant VARIANTS which is an array of discriminant names.

FromRepr

Add a function to enum that allows accessing variants by its discriminant

IntoStaticStr

Implements From<MyEnum> for &'static str on an enum.

ToString

implements std::string::ToString on en enum