Crate scrypto

The Scrypto Standard Library

The Scrypto Standard Library is the foundation of Scrypto blueprints, a set of minimal and shared abstractions on top of Radix Engine. It enables asset-oriented programming for feature-rich DeFi dApps.

If you know the name of what you’re looking for, the fastest way to find it is to use the search bar at the top of the page.

Modules

abi

Scrypto blueprint ABI.

address

Scrypto address.

buffer

Scrypto data encoding, decoding and exchange.

component

Scrypto component library.

constants

Scrypto constants.

core

Scrypto core library, mainly process and transaction context.

crypto

Scrypto cryptography library.

engine

Radix engine APIs.

math

Scrypto math library.

misc

Miscellaneous functions.

resource

Scrypto resource library.

values

Scrypto values.

Macros

access_and_or

access_rule_node

address

args

Constructs argument list for Scrypto function/method invocation.

args_from_bytes_vec

args_from_value_vec

blueprint

Declares a blueprint.

blueprint_name_from_context

borrow_component

This macro converts a ComponentAddress into a &Component via the Radix Engine component subsystem.

borrow_package

This macro creates a &Package from a PackageAddress via the Radix Engine component subsystem.

borrow_resource_manager

This macro creates a &ResourceManager from a ResourceAddress via the Radix Engine resource subsystem.

construct_address

debug

Logs a DEBUG message.

dec

Creates a Decimal from literals.

error

Logs an ERROR message.

external_blueprint

Generates a bridge/stub to make package calls to a blueprint.

external_component

Generates a bridge/stub to make cross-component calls.

external_interface_members

forward_ref_binop

forward_ref_op_assign

forward_ref_unop

i

Creates a safe integer from literals. You must specify the type of the integer you want to create.

import

Imports a blueprint from its ABI.

include_abi

Includes the ABI file of a Scrypto package.

include_code

Includes the WASM file of a Scrypto package.

info

Logs an INFO message.

native_functions

package_address_from_context

pdec

Creates a PreciseDecimal from literals.

resource_list

rule

this_package

trace

Logs a TRACE message.

warn

Logs a WARN message.

Traits

Decode

A data structure that can be decoded from a byte array using SBOR.

Describe

A data structure that can be described using SBOR types.

Encode

A data structure that can be serialized into a byte array using SBOR.

TypeId

A SBOR type ID.

Derive Macros

Decode

Derive code that decodes this data structure from a byte array.

Describe

Derive code that describes this data structure.

Encode

Derive code that encodes this data structure

NonFungibleData

Derive code that describe a non-fungible data structure.

TypeId

Derive code that returns the type ID.