Crate scrypto

The Scrypto Standard Library

The Scrypto Standard Library is the foundation of Scrypto blueprints, a set of minimal and shared abstractions on top of Radix Engine. It enables asset-oriented programming for feature-rich DeFi dApps.

If you know the name of what you’re looking for, the fastest way to find it is to use the search bar at the top of the page.

Re-exports

pub extern crate radix_engine_derive;

pub extern crate radix_engine_interface;

pub extern crate scrypto_abi;

Modules

abi

Scrypto blueprint ABI.

component

Scrypto component abstraction.

crypto

data

Scrypto data model.

engine

Scrypto RE abstraction.

macros 🔒

math

Scrypto math library.

model

Scrypto RE node model.

prelude

Scrypto preludes.

resource

Scrypto resource abstraction.

runtime

Scrypto runtime abstraction.

Macros

borrow_component

This macro converts a ComponentAddress into a &Component via the Radix Engine component subsystem.

borrow_package

This macro creates a &Package from a PackageAddress via the Radix Engine component subsystem.

borrow_resource_manager

This macro creates a &ResourceManager from a ResourceAddress via the Radix Engine resource subsystem.

debug

Logs a DEBUG message.

error

Logs an ERROR message.

external_blueprint

Generates a bridge/stub to make package calls to a blueprint.

external_blueprint_members

external_component

Generates a bridge/stub to make cross-component calls.

external_component_members

import

Imports a blueprint from its ABI.

include_abi

Includes the ABI file of a Scrypto package.

include_code

Includes the WASM file of a Scrypto package.

info

Logs an INFO message.

resource_list

scrypto_env_native_fn

this_package

trace

Logs a TRACE message.

warn

Logs a WARN message.

Functions

set_up_panic_hook

Sets up panic hook.

Attribute Macros

blueprint

Declares a blueprint.

Derive Macros

LegacyDescribe

Derive code that describes this data structure.

NonFungibleData

Derive code that describe a non-fungible data structure.

ScryptoCategorize

Derives code for categorizing a struct or enum with Scrypto value model.

ScryptoDecode

Derives code for decoding a struct or enum with Scrypto value model.

ScryptoEncode

Derives code for encoding a struct or enum with Scrypto value model.