Crate scrypto

The Scrypto Standard Library

The Scrypto Standard Library is the foundation of Scrypto blueprints, a set of minimal and shared abstractions on top of Radix Engine. It enables asset-oriented programming for feature-rich DeFi dApps.

If you know the name of what you’re looking for, the fastest way to find it is to use the search bar at the top of the page.

Modules

abi

Scrypto blueprint ABI.

buffer

Scrypto data encoding, decoding and exchange.

core

Scrypto core abstractions.

engine

Radix engine APIs.

prelude

Scrypto preludes.

resource

Scrypto resource abstractions.

rust

A facade of types in Rust standard library.

types

Scrypto primitive types.

utils

Utility functions.

Macros

args

Encodes arguments according to Scrypto ABI.

bdec

blueprint

Declares a blueprint.

debug

Logs a DEBUG message.

dec

error

Logs an ERROR message.

import

Imports a blueprint from its ABI.

include_code

Includes package code as a byte array.

info

Logs an INFO message.

trace

Logs a TRACE message.

warn

Logs a WARN message.

Attribute Macros

auth

Defines the authorization rule for a method.

Derive Macros

NonFungibleData

Derive code that describe a non-fungible data structure.