Module scrypto::address

Scrypto address.

Macros

address

construct_address

Structs

Bech32Decoder

Represents a decoder which understands how to decode Scrypto addresses in Bech32.

Bech32Encoder

Represents an encoder which understands how to encode Scrypto addresses in Bech32.

HrpSet

Represents an HRP set (typically corresponds to a network).

Enums

AddressError

Represents an error in addressing.

EntityType

An enum which represents the different addressable entities.

EntityTypeError

Constants

ACCOUNT_COMPONENT_ADDRESS_ENTITY_ID

A unique identifier used in the addressing of Account Component Addresses.

BECH32_DECODER

BECH32_ENCODER

NORMAL_COMPONENT_ADDRESS_ENTITY_ID

A unique identifier used in the addressing of Generic Component Addresses.

PACKAGE_ADDRESS_ENTITY_ID

A unique identifier used in the addressing of Package Addresses.

RESOURCE_ADDRESS_ENTITY_ID

A unique identifier used in the addressing of Resource Addresses.

SYSTEM_COMPONENT_ADDRESS_ENTITY_ID

A unique identifier used in the addressing of System Component Addresses.