Module scrypto::component

Scrypto component library.

Structs

Component

Represents an instantiated component.

ComponentAddress

An instance of a blueprint, which lives in the ledger state.

ComponentSystem

Represents the Radix Engine component subsystem.

LazyMap

A scalable key-value map which loads entries on demand.

LocalComponent

Package

Represents a published package.

PackageAddress

A collection of blueprints, compiled and published as a single unit.

Enums

ParseComponentAddressError

Represents an error when decoding component address.

ParseLazyMapError

Represents an error when decoding lazy map.

ParsePackageAddressError

Represents an error when decoding package address.

Traits

ComponentState

Represents the state of a component.

Functions

component_system

Returns the component subsystem.

init_component_system

Initializes component subsystem.