Module scrypto::component

Scrypto component library.

Structs

BorrowedPackage

Represents a published package.

Component

Represents an instantiated component.

ComponentAddAccessCheckInput

ComponentSystem

Represents the Radix Engine component subsystem.

KeyValueStore

A scalable key-value map which loads entries on demand.

PackagePublishInput

Enums

ComponentAddress

An instance of a blueprint, which lives in the ledger state.

PackageAddress

A collection of blueprints, compiled and published as a single unit.

ParseKeyValueStoreError

Represents an error when decoding key value store.

Traits

ComponentState

Represents the state of a component.

LocalComponent

Functions

component_system

Returns the component subsystem.

init_component_system

Initializes component subsystem.