Module scrypto::prelude

Scrypto preludes.

Re-exports

pub use radix_engine_interface::radix_engine_common;

pub use crate::component::*;

pub use crate::engine::*;

pub use crate::modules::*;

pub use crate::resource::*;

pub use crate::runtime::*;

Modules

additional_attributes

Documentation for Additional Attributes

auth

bnum_integer

Definitions of safe integers and uints.

btree_map

btree_set

decimal

fmt

Utilities for formatting and printing Strings.

hash_map

hash_set

index_map

The methods and macros provided directly in this index_map module (new, with_capacity) work in both std and no-std modes - unlike the corresponding methods on IndexMap itself.

index_set

The methods and macros provided directly in this index_set module (new, with_capacity) work in both std and no-std modes - unlike the corresponding methods on IndexSet itself.

instant

metadata

non_iter_map

precise_decimal

radix_engine_common

rounding_mode

royalty

utc_date_time

vec

A contiguous growable array type with heap-allocated contents, written Vec<T>.

Macros

access_and_or

access_rule_node

btreemap

btreeset

burn_roles

component_royalties

component_royalty_config

debug

Logs a DEBUG message.

dec

Creates a Decimal from literals.

define_wrapped_hash

deposit_roles

enable_function_auth

enable_method_auth

enable_package_royalties

error

Logs an ERROR message.

extern_blueprint_internal

format

Creates a String using interpolation of runtime expressions.

freeze_roles

hashmap

hashset

i

Creates a safe integer from literals. You must specify the type of the integer you want to create.

include_code

Includes the WASM file of a Scrypto package.

include_schema

Includes the schema file of a Scrypto package.

indexmap

indexset

info

Logs an INFO message.

internal_add_role

internal_component_royalty_entry

internal_roles

main_accessibility

manifest_args

metadata

metadata_init

metadata_init_set_entry

metadata_roles

method_accessibilities

method_accessibility

mint_roles

non_fungible_data_update_roles

pdec

Creates a PreciseDecimal from literals.

recall_roles

resource_list

role_entry

role_list

roles

Roles macro for main module

roles2

rule

scrypto_args

this_package

to_manifest_value_and_unwrap

to_role_key

trace

Logs a TRACE message.

vec

Creates a Vec containing the arguments.

warn

Logs a WARN message.

withdraw_roles

Structs

AccessRulesCreateInput

AccessRulesGetRoleInput

AccessRulesLockOwnerRoleInput

AccessRulesSetOwnerRoleInput

AccessRulesSetRoleInput

AddressBech32Decoder

Represents a decoder which understands how to decode Scrypto addresses in Bech32.

AddressBech32Encoder

Represents an encoder which understands how to encode Scrypto addresses in Bech32.

AddressDisplayContext

AuthAddresses

AuthZoneClearInput

AuthZoneClearVirtualProofsInput

AuthZoneCreateProofOfAllInput

AuthZoneCreateProofOfAmountInput

AuthZoneCreateProofOfNonFungiblesInput

AuthZoneDrainInput

AuthZoneDropInput

AuthZonePopInput

AuthZonePushInput

Bucket

BucketCreateProofOfAllInput

BucketGetAmountInput

BucketGetNonFungibleLocalIdsInput

BucketGetResourceAddressInput

BucketPutInput

BucketTakeAdvancedInput

BucketTakeInput

BucketTakeNonFungiblesInput

BurnRoles

DepositRoles

FreezeRoles

FungibleBucket

FungibleBucketCreateProofOfAmountInput

FungibleBucketLockAmountInput

FungibleBucketUnlockAmountInput

FungibleProof

FungibleResourceManagerCreateInput

FungibleResourceManagerCreateManifestInput

FungibleResourceManagerCreateWithInitialSupplyInput

FungibleResourceManagerCreateWithInitialSupplyManifestInput

FungibleResourceManagerMintInput

FungibleResourceRoles

FungibleVaultCreateProofOfAmountInput

FungibleVaultLockFeeInput

FungibleVaultLockFungibleAmountInput

FungibleVaultUnlockFungibleAmountInput

LazyCellExperimental

A value which is initialized on the first access.

BTreeMap

An ordered map based on a B-Tree.

BTreeSet

An ordered set based on a B-Tree.

BlueprintEventSchemaInit

BlueprintFunctionsSchemaInit

BlueprintId

BlueprintIndexSchema

BlueprintKeyValueStoreSchema

BlueprintSchemaInit

BlueprintSortedIndexSchema

BlueprintStateSchemaInit

BnumI256

Provides safe integer arithmetic.

BnumI384

Provides safe integer arithmetic.

BnumI512

Provides safe integer arithmetic.

BnumI768

Provides safe integer arithmetic.

BnumU256

Provides safe integer arithmetic.

BnumU384

Provides safe integer arithmetic.

BnumU512

Provides safe integer arithmetic.

BnumU768

Provides safe integer arithmetic.

BorrowError

An error returned by RefCell::try_borrow.

BorrowMutError

An error returned by RefCell::try_borrow_mut.

Box

A pointer type that uniquely owns a heap allocation of type T.

Buffer

BytesNonFungibleLocalId

Bytes, of length between 1 and 64.

Cell

A mutable memory location.

ComponentAddress

Address to a global component

ContextDisplayable

ContextSerializable

Decimal

Decimal represents a 256 bit representation of a fixed-scale decimal number.

Ed25519PublicKey

Represents an ED25519 public key.

Ed25519PublicKeyHash

Epoch

A type-safe consensus epoch number.

EventTypeIdentifier

Identifies a specific event schema emitter by some emitter RENode.

FieldSchema

FnIdentifier

FunctionIdentifier

FunctionSchemaInit

FungibleVault

GlobalAddress

Address to a global entity

GlobalAddressPhantom

GlobalAddressReservation

Hash

Represents a 32-byte hash digest.

HashAccumulator

Represents a 32-byte hash accumulator.

HrpSet

Represents an HRP set (typically corresponds to a network).

IndexedScryptoValue

InstanceSchema

Instant

Represents a Unix timestamp, capturing the seconds since the unix epoch.

IntegerNonFungibleLocalId

Unsigned integers, up to u64.

InternalAddress

Address to a local entity

KeyValueStoreInfo

KeyValueStoreInit

KeyValueStoreInitEntry

KeyValueStoreSchema

LiquidFungibleResource

LiquidNonFungibleResource

LiquidNonFungibleVault

LockedFungibleResource

LockedNonFungibleResource

ManifestAddressReservation

ManifestArgs

ManifestBlobRef

ManifestBucket

ManifestCustomTerminalValueRef

ManifestDecimal

ManifestPreciseDecimal

ManifestProof

ManifestValueDisplayContext

Note - this is quite similar to ManifestDecompilationDisplayContext

MetadataCreateInput

MetadataCreateWithDataInput

MetadataGetInput

MetadataLockInput

MetadataRemoveInput

MetadataRoles

MetadataSetInput

MethodIdentifier

MethodKey

MintRoles

ModuleConfig

ModuleRoleKey

NetworkDefinition

Network Definition is intended to be the actual definition of a network

NodeId

The unique identifier of a (stored) node.

NonFungibleBucket

NonFungibleBucketCreateProofOfNonFungiblesInput

NonFungibleBucketLockNonFungiblesInput

NonFungibleBucketUnlockNonFungiblesInput

NonFungibleDataSchema

NonFungibleDataUpdateRoles

NonFungibleGlobalId

Represents the global id of a non-fungible.

NonFungibleProof

NonFungibleProofGetLocalIdsInput

NonFungibleResourceManagerCreateInput

NonFungibleResourceManagerCreateManifestInput

NonFungibleResourceManagerCreateRuidWithInitialSupplyInput

NonFungibleResourceManagerCreateRuidWithInitialSupplyManifestInput

NonFungibleResourceManagerCreateWithInitialSupplyInput

NonFungibleResourceManagerCreateWithInitialSupplyManifestInput

NonFungibleResourceManagerExistsInput

NonFungibleResourceManagerGetNonFungibleInput

NonFungibleResourceManagerMintInput

NonFungibleResourceManagerMintManifestInput

NonFungibleResourceManagerMintRuidInput

NonFungibleResourceManagerMintRuidManifestInput

NonFungibleResourceManagerMintSingleRuidInput

NonFungibleResourceManagerUpdateDataInput

NonFungibleResourceRoles

NonFungibleVault

NonFungibleVaultBurnNonFungiblesInput

NonFungibleVaultCreateProofOfNonFungiblesInput

NonFungibleVaultGetNonFungibleLocalIdsInput

NonFungibleVaultLockNonFungiblesInput

NonFungibleVaultRecallNonFungiblesInput

NonFungibleVaultTakeNonFungiblesInput

NonFungibleVaultUnlockNonFungiblesInput

NonIterMap

A thin wrapper around a HashMap, which guarantees that a HashMap usage will not result in a non-deterministic execution (simply by disallowing the iteration over its elements).

ObjectInfo

Origin

OwnedAuthZone

OwnedWorktop

OwnerRoleEntry

PreciseDecimal

PreciseDecimal represents a 512 bit representation of a fixed-scale decimal number.

Proof

ProofCloneInput

ProofDropInput

ProofGetAmountInput

ProofGetResourceAddressInput

RecallRoles

ResourceManagerBurnInput

ResourceManagerCreateEmptyBucketInput

ResourceManagerCreateEmptyVaultInput

ResourceManagerDropEmptyBucketInput

ResourceManagerGetAmountForWithdrawalInput

ResourceManagerGetResourceTypeInput

ResourceManagerGetTotalSupplyInput

ResourceManagerPackageBurnInput

ResourceOrNonFungibleList

RoleKey

RoleList

RolesInit

SborFixedEnumVariant

SborPath

A series of indexes which describes some value in the sbor tree

SborPathBuf

SborRawPayload

A wrapper for a full encoded SBOR payload, including the prefix byte.

SborRawValue

A wrapper for a reference to a valid partial SBOR payload representing a single value.

Secp256k1PublicKey

Represents an ECDSA Secp256k1 public key.

Secp256k1PublicKeyHash

SyncUnsafeCellExperimental

UnsafeCell, but Sync.

OnceCell

A cell which can be written to only once.

Own

PackageAddress

Address to a global package

PartitionNumber

The unique identifier of a node module.

PartitionOffset

PhantomData

Zero-sized type used to mark things that “act like” they own a T.

RUIDNonFungibleLocalId

RUID, v4, variant 1, big endian. See https://www.rfc-editor.org/rfc/rfc4122

Rc

A single-threaded reference-counting pointer. ‘Rc’ stands for ‘Reference Counted’.

ReceiverInfo

Ref

Wraps a borrowed reference to a value in a RefCell box. A wrapper type for an immutably borrowed value from a RefCell<T>.

RefCell

A mutable memory location with dynamically checked borrow rules

RefMut

A wrapper type for a mutably borrowed value from a RefCell<T>.

RefTypes

Reference

ResourceAddress

Address to a global resource

Round

A type-safe consensus round number within a single epoch.

SchemaPath

Describes a value located in some sbor given a schema for that sbor

ScryptoCustomExtension

ScryptoCustomSchema

ScryptoCustomTerminalValueRef

ScryptoValueDisplayContext

Slice

String

A UTF-8–encoded, growable string.

StringNonFungibleLocalId

A string matching [_0-9a-zA-Z]{1,64}.

UnsafeCell

The core primitive for interior mutability in Rust.

Url

UtcDateTime

A UtcDateTime represents a Unix timestamp on the UTC Calendar.

Vault

VaultBurnInput

VaultFreezeFlags

VaultFreezeInput

VaultFrozenFlag

VaultGetAmountInput

VaultPutInput

VaultRecallInput

VaultTakeAdvancedInput

VaultTakeInput

VaultUnfreezeInput

Vec

A contiguous growable array type, written as Vec<T>, short for ‘vector’.

WithdrawRoles

WorktopAssertContainsAmountInput

WorktopAssertContainsInput

WorktopAssertContainsNonFungiblesInput

WorktopDrainInput

WorktopDropInput

WorktopPutInput

WorktopTakeAllInput

WorktopTakeInput

WorktopTakeNonFungiblesInput

Enums

AccessControllerField

AccessRule

AccessRuleNode

AccessRulesField

AccountField

AccountPartitionOffset

AddressBech32DecodeError

AddressBech32EncodeError

AuthZoneField

BlueprintCollectionSchema

ClientCostingEntry

ComponentField

ComponentRoyaltyConfig

Royalty rules

Condition

ConsensusManagerField

ConsensusManagerPartitionOffset

ContentValidationError

Cow

A clone-on-write smart pointer.

DateTimeError

DecodeError

Represents an error ocurred during decoding.

Emitter

EmptyMetadata

EncodeError

Represents an error occurred during encoding.

EntityType

An enum which represents the different addressable entities.

FallToOwner

FnIdent

FungibleBucketField

FungibleProofField

FungibleResourceManagerField

FungibleVaultField

Generic

GlobalCaller

Level

Represents the level of a log message.

ManifestAddress

Any address supported by manifest, both global and local.

ManifestCustomExtension

ManifestCustomTraversal

ManifestCustomValue

ManifestCustomValueKind

ManifestExpression

ManifestNonFungibleLocalId

ManifestNonFungibleLocalIdValidationError

ManifestToRustValueError

MetadataError

MetadataValue

MethodAccessibility

MultiResourcePoolField

NonFungibleBucketField

NonFungibleIdType

Represents type of non-fungible id

NonFungibleLocalId

Represents the local id of a non-fungible.

NonFungibleProofField

NonFungibleResourceManagerField

NonFungibleResourceManagerPartitionOffset

NonFungibleVaultField

NonFungibleVaultPartitionOffset

ObjectBlueprintInfo

OneResourcePoolField

Option

The Option type. See the module level documentation for more.

OwnValidation

OwnerRole

Front end data structure for specifying owner role

OwnerRoleUpdater

PackageCode

PackageField

PackagePartitionOffset

PackageRoyalty

Royalty rules

PackageRoyaltyConfig

Royalty rules

ParseBnumI256Error

ParseBnumI384Error

ParseBnumI512Error

ParseBnumI768Error

ParseBnumU256Error

ParseBnumU384Error

ParseBnumU512Error

ParseBnumU768Error

ParseComponentAddressError

ParseDecimalError

Represents an error when parsing Decimal from another type.

ParseEd25519PublicKeyError

Represents an error when parsing ED25519 public key from hex.

ParseError

The ParseError enum is a collection of all the possible reasons an enum can fail to parse from a string.

ParseGlobalAddressError

ParseHashError

Represents an error when parsing hash.

ParseInternalAddressError

ParseManifestAddressReservationError

Represents an error when parsing ManifestAddressReservation.

ParseManifestBlobRefError

Represents an error when parsing ManifestBlobRef.

ParseManifestBucketError

Represents an error when parsing ManifestBucket.

ParseManifestDecimalError

Represents an error when parsing ManifestDecimal.

ParseManifestExpressionError

Represents an error when parsing ManifestExpression.

ParseManifestPreciseDecimalError

Represents an error when parsing ManifestPreciseDecimal.

ParseManifestProofError

Represents an error when parsing ManifestProof.

ParseNetworkError

ParseNonFungibleGlobalIdError

Represents an error when parsing non-fungible address.

ParseNonFungibleLocalIdError

Represents an error when decoding non-fungible id.

ParseOwnError

ParsePackageAddressError

ParsePreciseDecimalError

Represents an error when parsing PreciseDecimal from another type.

ParseReferenceError

ParseResourceAddressError

ParseSecp256k1PublicKeyError

Represents an error when parsing ED25519 public key from hex.

ParseUtcDateTimeError

ProofRule

Resource Proof Rules

PublicKey

Represents any natively supported public key.

PublicKeyHash

The hash of a given public key.

Receiver

ReferenceValidation

ResourceError

ResourceFeature

ResourceOrNonFungible

ResourceType

Represents the type of a resource.

Result

Result is a type that represents either success (Ok) or failure (Err).

RoundingMode

Defines the rounding strategy.

RoyaltyAmount

RoyaltyField

RustToManifestValueError

SborValue

SchemaPathParseError

SchemaSubPath

ScryptoCustomTraversal

ScryptoCustomTypeKind

A schema for the values that a codec can decode / views as valid

ScryptoCustomTypeValidation

ScryptoCustomValue

ScryptoCustomValueKind

SubstateKey

The unique identifier of a substate within a node module.

TimeComparisonOperator

TransactionTrackerField

TwoResourcePoolField

TypeInfoField

TypeRef

ValidatorField

ValueKind

WithdrawStrategy

Define the withdraw strategy when request amount does not match underlying resource divisibility.

WorktopField

Constants

ACCESS_CONTROLLER_PACKAGE

The native package for access controllers.

ACCESS_RULES_BASE_PARTITION

ACCESS_RULES_BLUEPRINT

ACCESS_RULES_CREATE_IDENT

ACCESS_RULES_FIELDS_PARTITION

ACCESS_RULES_FIELDS_PARTITION_OFFSET

ACCESS_RULES_GET_ROLE_IDENT

ACCESS_RULES_LOCK_OWNER_ROLE_IDENT

ACCESS_RULES_MODULE_PACKAGE

The native package for the access rules module.

ACCESS_RULES_MUTABILITY_PARTITION_OFFSET

ACCESS_RULES_ROLE_DEF_PARTITION

ACCESS_RULES_ROLE_DEF_PARTITION_OFFSET

ACCESS_RULES_SET_OWNER_ROLE_IDENT

ACCESS_RULES_SET_ROLE_IDENT

ACCOUNT_OWNER_BADGE

The non-fungible badge resource which is used for account ownership, if accounts have been set up with simple account creation, or have been securified.

ACCOUNT_PACKAGE

The native package for accounts.

AUTH_ZONE_BLUEPRINT

AUTH_ZONE_CLEAR_EXPORT_NAME

AUTH_ZONE_CLEAR_IDENT

AUTH_ZONE_CLEAR_SIGNATURE_PROOFS_EXPORT_NAME

AUTH_ZONE_CLEAR_SIGNATURE_PROOFS_IDENT

AUTH_ZONE_CREATE_PROOF_OF_ALL_EXPORT_NAME

AUTH_ZONE_CREATE_PROOF_OF_ALL_IDENT

AUTH_ZONE_CREATE_PROOF_OF_AMOUNT_EXPORT_NAME

AUTH_ZONE_CREATE_PROOF_OF_AMOUNT_IDENT

AUTH_ZONE_CREATE_PROOF_OF_NON_FUNGIBLES_EXPORT_NAME

AUTH_ZONE_CREATE_PROOF_OF_NON_FUNGIBLES_IDENT

AUTH_ZONE_DRAIN_EXPORT_NAME

AUTH_ZONE_DRAIN_IDENT

AUTH_ZONE_DROP_EXPORT_NAME

AUTH_ZONE_DROP_IDENT

AUTH_ZONE_POP_EXPORT_NAME

AUTH_ZONE_POP_IDENT

AUTH_ZONE_PUSH_EXPORT_NAME

AUTH_ZONE_PUSH_IDENT

BUCKET_CREATE_PROOF_OF_ALL_IDENT

BUCKET_GET_AMOUNT_IDENT

BUCKET_GET_RESOURCE_ADDRESS_IDENT

BUCKET_PUT_IDENT

BUCKET_TAKE_ADVANCED_IDENT

BUCKET_TAKE_IDENT

BURNER_ROLE

BURNER_UPDATER_ROLE

BURN_FEATURE

CONSENSUS_MANAGER

The consensus manager native component - in charge of validators, consensus and epochs.

CONSENSUS_MANAGER_PACKAGE

The native package for the consensus manager.

DEFAULT_COST_UNIT_LIMIT

The default cost unit limit.

DEFAULT_COST_UNIT_PRICE_IN_XRD

The default cost unit price, in XRD.

DEFAULT_FREE_CREDIT_IN_XRD

The default free credit, for preview only.

DEFAULT_MAX_CALL_DEPTH

The default max call depth, used by transaction executor.

DEFAULT_MAX_COST_UNIT_LIMIT

The maximum value of cost unit limit

DEFAULT_MAX_EPOCH_RANGE

The max epoch range Should be ~ 1 month. The below is ~30 days given 5 minute epochs.

DEFAULT_MAX_EVENT_SIZE

The max event size

DEFAULT_MAX_EXECUTION_TRACE_DEPTH

DEFAULT_MAX_INVOKE_INPUT_SIZE

The default maximum invoke input args size.

DEFAULT_MAX_LOG_SIZE

The max log size

DEFAULT_MAX_METADATA_KEY_STRING_LEN

The max SBOR size of metadata key

DEFAULT_MAX_METADATA_VALUE_SBOR_LEN

The max SBOR size of metadata value

DEFAULT_MAX_NUMBER_OF_EVENTS

The max number of events

DEFAULT_MAX_NUMBER_OF_LOGS

The max number of logs

DEFAULT_MAX_NUMBER_OF_SUBSTATES_IN_HEAP

The default max number of substates in heap.

DEFAULT_MAX_NUMBER_OF_SUBSTATES_IN_TRACK

The default max number of substates in track.

DEFAULT_MAX_PANIC_MESSAGE_SIZE

The max panic message size

DEFAULT_MAX_PER_FUNCTION_ROYALTY_IN_XRD

The default maximum that a package or component owner is allowed to set their method royalty to

DEFAULT_MAX_SUBSTATE_SIZE

The default maximum substate read and write size.

DEFAULT_MAX_TIP_PERCENTAGE

The maximum value of tip percentage

DEFAULT_MAX_TRANSACTION_SIZE

The max transaction size

DEFAULT_MIN_COST_UNIT_LIMIT

The minimum value of cost unit limit

DEFAULT_MIN_TIP_PERCENTAGE

The minimum value of tip percentage

DEFAULT_STATE_EXPANSION_PRICE_IN_XRD

The default price for adding a single byte to the substate store, in XRD.

DEFAULT_SYSTEM_LOAN

The default system loan amount, used by transaction executor.

DEFAULT_TIP_PERCENTAGE

DEFAULT_USD_PRICE_IN_XRD

The default USD price, in XRD

DEPOSITOR_ROLE

DEPOSITOR_UPDATER_ROLE

ED25519_SIGNATURE_VIRTUAL_BADGE

The non-fungible badge resource which is used for virtual proofs of EdDSA Ed25519 transacton signatures in the transaction processor.

FAUCET

The faucet native component - use this on testnets for getting XRD and locking fee.

FAUCET_BLUEPRINT

The name of the faucet blueprint under the FAUCET_PACKAGE.

FAUCET_COMPONENT

The faucet native component - use this on testnets for getting XRD and locking fee.

FAUCET_PACKAGE

The scrypto package for the faucet

FEES_PROPOSER_SHARE_PERCENTAGE

The proposer’s share of fees (execution and state expansion)

FEES_VALIDATOR_SET_SHARE_PERCENTAGE

The validator set’s share of fees (execution and state expansion)

FREEZER_ROLE

FREEZER_UPDATER_ROLE

FUNGIBLE_BUCKET_BLUEPRINT

FUNGIBLE_BUCKET_CREATE_PROOF_OF_AMOUNT_IDENT

FUNGIBLE_BUCKET_LOCK_AMOUNT_IDENT

FUNGIBLE_BUCKET_UNLOCK_AMOUNT_IDENT

FUNGIBLE_PROOF_BLUEPRINT

FUNGIBLE_RESOURCE_MANAGER_BLUEPRINT

FUNGIBLE_RESOURCE_MANAGER_CREATE_IDENT

FUNGIBLE_RESOURCE_MANAGER_CREATE_WITH_INITIAL_SUPPLY_IDENT

FUNGIBLE_RESOURCE_MANAGER_MINT_IDENT

FUNGIBLE_VAULT_BLUEPRINT

FUNGIBLE_VAULT_CREATE_PROOF_OF_AMOUNT_IDENT

FUNGIBLE_VAULT_LOCK_FEE_IDENT

FUNGIBLE_VAULT_LOCK_FUNGIBLE_AMOUNT_IDENT

FUNGIBLE_VAULT_UNLOCK_FUNGIBLE_AMOUNT_IDENT

GENESIS_HELPER

The genesis helper scrypto component - used for sorting out genesis.

GENESIS_HELPER_BLUEPRINT

The name of the genesis helper blueprint under the GENESIS_HELPER_PACKAGE.

GENESIS_HELPER_PACKAGE

The scrypto package for the genesis helper.

GLOBAL_CALLER_VIRTUAL_BADGE

The non-fungible badge resource which is used for virtual proofs which represent the global ancestor of the actor which made the latest global call.

IDENTITY_OWNER_BADGE

The non-fungible badge resource which is used for identity ownership, if identities have been set up with simple account creation, or have been securified.

IDENTITY_PACKAGE

The native package for identities.

MAIN_BASE_PARTITION

MANIFEST_NON_FUNGIBLE_LOCAL_ID_MAX_LENGTH

MANIFEST_SBOR_V1_MAX_DEPTH

MANIFEST_SBOR_V1_PAYLOAD_PREFIX

0x4d = M in ASCII for Manifest - (77 in decimal)

MANIFEST_VALUE_KIND_ADDRESS

MANIFEST_VALUE_KIND_ADDRESS_RESERVATION

MANIFEST_VALUE_KIND_BLOB

MANIFEST_VALUE_KIND_BUCKET

MANIFEST_VALUE_KIND_DECIMAL

MANIFEST_VALUE_KIND_EXPRESSION

MANIFEST_VALUE_KIND_NON_FUNGIBLE_LOCAL_ID

MANIFEST_VALUE_KIND_PRECISE_DECIMAL

MANIFEST_VALUE_KIND_PROOF

MAX_NUMBER_OF_BLOBS

MAX_NUMBER_OF_INTENT_SIGNATURES

METADATA_BLUEPRINT

METADATA_CREATE_IDENT

METADATA_CREATE_WITH_DATA_IDENT

METADATA_DISCRIMINATOR_ARRAY_BASE

METADATA_GET_IDENT

METADATA_KV_STORE_PARTITION

METADATA_LOCKER_ROLE

METADATA_LOCKER_UPDATER_ROLE

METADATA_LOCK_IDENT

METADATA_MODULE_PACKAGE

The native package for the metadata module.

METADATA_REMOVE_IDENT

METADATA_SETTER_ROLE

METADATA_SETTER_UPDATER_ROLE

METADATA_SET_IDENT

METADATA_VALUE_BOOLEAN_ARRAY_DISCRIMINATOR

METADATA_VALUE_BOOLEAN_DISCRIMINATOR

METADATA_VALUE_DECIMAL_ARRAY_DISCRIMINATOR

METADATA_VALUE_DECIMAL_DISCRIMINATOR

METADATA_VALUE_GLOBAL_ADDRESS_ARRAY_DISCRIMINATOR

METADATA_VALUE_GLOBAL_ADDRESS_DISCRIMINATOR

METADATA_VALUE_I32_ARRAY_DISCRIMINATOR

METADATA_VALUE_I32_DISCRIMINATOR

METADATA_VALUE_I64_ARRAY_DISCRIMINATOR

METADATA_VALUE_I64_DISCRIMINATOR

METADATA_VALUE_INSTANT_ARRAY_DISCRIMINATOR

METADATA_VALUE_INSTANT_DISCRIMINATOR

METADATA_VALUE_NON_FUNGIBLE_GLOBAL_ID_ARRAY_DISCRIMINATOR

METADATA_VALUE_NON_FUNGIBLE_GLOBAL_ID_DISCRIMINATOR

METADATA_VALUE_NON_FUNGIBLE_LOCAL_ID_ARRAY_DISCRIMINATOR

METADATA_VALUE_NON_FUNGIBLE_LOCAL_ID_DISCRIMINATOR

METADATA_VALUE_ORIGIN_ARRAY_DISCRIMINATOR

METADATA_VALUE_ORIGIN_DISCRIMINATOR

METADATA_VALUE_PUBLIC_KEY_ARRAY_DISCRIMINATOR

METADATA_VALUE_PUBLIC_KEY_DISCRIMINATOR

METADATA_VALUE_PUBLIC_KEY_HASH_ARRAY_DISCRIMINATOR

METADATA_VALUE_PUBLIC_KEY_HASH_DISCRIMINATOR

METADATA_VALUE_STRING_ARRAY_DISCRIMINATOR

METADATA_VALUE_STRING_DISCRIMINATOR

METADATA_VALUE_U8_ARRAY_DISCRIMINATOR

METADATA_VALUE_U8_DISCRIMINATOR

METADATA_VALUE_U32_ARRAY_DISCRIMINATOR

METADATA_VALUE_U32_DISCRIMINATOR

METADATA_VALUE_U64_ARRAY_DISCRIMINATOR

METADATA_VALUE_U64_DISCRIMINATOR

METADATA_VALUE_URL_ARRAY_DISCRIMINATOR

METADATA_VALUE_URL_DISCRIMINATOR

MINTER_ROLE

MINTER_UPDATER_ROLE

MINT_FEATURE

NATIVE_TOKEN

XRD is the native token of the Radix ledger. It is a fungible token, measured in attos (10^-18).

NON_FUNGIBLE_BUCKET_BLUEPRINT

NON_FUNGIBLE_BUCKET_CREATE_PROOF_OF_NON_FUNGIBLES_IDENT

NON_FUNGIBLE_BUCKET_GET_NON_FUNGIBLE_LOCAL_IDS_IDENT

NON_FUNGIBLE_BUCKET_LOCK_NON_FUNGIBLES_IDENT

NON_FUNGIBLE_BUCKET_TAKE_NON_FUNGIBLES_IDENT

NON_FUNGIBLE_BUCKET_UNLOCK_NON_FUNGIBLES_IDENT

NON_FUNGIBLE_DATA_UPDATER_ROLE

NON_FUNGIBLE_DATA_UPDATER_UPDATER_ROLE

NON_FUNGIBLE_LOCAL_ID_MAX_LENGTH

NON_FUNGIBLE_PROOF_BLUEPRINT

NON_FUNGIBLE_PROOF_GET_LOCAL_IDS_IDENT

NON_FUNGIBLE_RESOURCE_MANAGER_BLUEPRINT

NON_FUNGIBLE_RESOURCE_MANAGER_CREATE_IDENT

NON_FUNGIBLE_RESOURCE_MANAGER_CREATE_RUID_WITH_INITIAL_SUPPLY_IDENT

NON_FUNGIBLE_RESOURCE_MANAGER_CREATE_WITH_INITIAL_SUPPLY_IDENT

NON_FUNGIBLE_RESOURCE_MANAGER_EXISTS_IDENT

NON_FUNGIBLE_RESOURCE_MANAGER_GET_NON_FUNGIBLE_IDENT

NON_FUNGIBLE_RESOURCE_MANAGER_MINT_IDENT

NON_FUNGIBLE_RESOURCE_MANAGER_MINT_RUID_IDENT

NON_FUNGIBLE_RESOURCE_MANAGER_MINT_SINGLE_RUID_IDENT

NON_FUNGIBLE_RESOURCE_MANAGER_UPDATE_DATA_IDENT

NON_FUNGIBLE_VAULT_BLUEPRINT

NON_FUNGIBLE_VAULT_BURN_NON_FUNGIBLES_IDENT

NON_FUNGIBLE_VAULT_CREATE_PROOF_OF_NON_FUNGIBLES_IDENT

NON_FUNGIBLE_VAULT_GET_NON_FUNGIBLE_LOCAL_IDS_IDENT

NON_FUNGIBLE_VAULT_LOCK_NON_FUNGIBLES_IDENT

NON_FUNGIBLE_VAULT_RECALL_NON_FUNGIBLES_IDENT

NON_FUNGIBLE_VAULT_TAKE_NON_FUNGIBLES_IDENT

NON_FUNGIBLE_VAULT_UNLOCK_NON_FUNGIBLES_IDENT

OWNER_ROLE

PACKAGE_OF_DIRECT_CALLER_VIRTUAL_BADGE

The non-fungible badge resource which is used for virtual proofs which represent the package of the immediate caller - ie the actor which made the latest (global or internal) call.

PACKAGE_OWNER_BADGE

The non-fungible badge resource which is used for package ownership when creating packages with the simple package creation set-up.

PACKAGE_PACKAGE

The native package for package deployment.

POOL_PACKAGE

The native package for pools.

PROOF_CLONE_IDENT

PROOF_DROP_IDENT

PROOF_GET_AMOUNT_IDENT

PROOF_GET_RESOURCE_ADDRESS_IDENT

RADIX_TOKEN

XRD is the native token of the Radix ledger. It is a fungible token, measured in attos (10^-18).

RECALLER_ROLE

RECALLER_UPDATER_ROLE

RESOURCE_MANAGER_BURN_IDENT

RESOURCE_MANAGER_CREATE_EMPTY_BUCKET_IDENT

RESOURCE_MANAGER_CREATE_EMPTY_VAULT_IDENT

RESOURCE_MANAGER_DROP_EMPTY_BUCKET_IDENT

RESOURCE_MANAGER_GET_AMOUNT_FOR_WITHDRAWAL_IDENT

RESOURCE_MANAGER_GET_RESOURCE_TYPE_IDENT

RESOURCE_MANAGER_GET_TOTAL_SUPPLY_IDENT

RESOURCE_MANAGER_PACKAGE_BURN_IDENT

RESOURCE_PACKAGE

The native package for resource managers, proofs, buckets, vaults etc.

ROLA_HASHABLE_PAYLOAD_PREFIX

The ROLA hash which is signed is created as hash(ROLA_HASHABLE_PAYLOAD_PREFIX || ..)

ROYALTY_BASE_PARTITION

ROYALTY_CONFIG_PARTITION

ROYALTY_CONFIG_PARTITION_OFFSET

ROYALTY_FIELDS_PARTITION

ROYALTY_FIELDS_PARTITION_OFFSET

ROYALTY_MODULE_PACKAGE

The native package for the royalty module.

SCRYPTO_SBOR_V1_MAX_DEPTH

SCRYPTO_SBOR_V1_PAYLOAD_PREFIX

0x5c for [5c]rypto - (91 in decimal)

SECP256K1_SIGNATURE_VIRTUAL_BADGE

The non-fungible badge resource which is used for virtual proofs of ECDSA Secp256k1 transacton signatures in the transaction processor.

SELF_ROLE

SYSTEM_TRANSACTION_BADGE

The non-fungible badge resource which is used for virtual proofs representing the fact that the current transaction is a system transaction.

TIPS_PROPOSER_SHARE_PERCENTAGE

The proposer’s share of tips

TIPS_VALIDATOR_SET_SHARE_PERCENTAGE

The validator set’s share of tips

TRACK_TOTAL_SUPPLY_FEATURE

TRANSACTION_HASHABLE_PAYLOAD_PREFIX

The Transaction hash which is signed is created as: hash(TRANSACTION_HASHABLE_PAYLOAD_PREFIX || version prefix according to type of transaction payload || ..)

TRANSACTION_PROCESSOR_PACKAGE

The native package for the transaction processor.

TRANSACTION_TRACKER

The intent hash store component

TRANSACTION_TRACKER_BLUEPRINT

TRANSACTION_TRACKER_PACKAGE

TRANSACTION_VERSION_V1

TYPE_INFO_FIELD_PARTITION

VALIDATOR_OWNER_BADGE

The non-fungible badge resource which is used for validator ownership.

VALUE_KIND_ARRAY

VALUE_KIND_BOOL

VALUE_KIND_DECIMAL

VALUE_KIND_ENUM

VALUE_KIND_I8

VALUE_KIND_I16

VALUE_KIND_I32

VALUE_KIND_I64

VALUE_KIND_I128

VALUE_KIND_MAP

VALUE_KIND_NON_FUNGIBLE_LOCAL_ID

VALUE_KIND_OWN

VALUE_KIND_PRECISE_DECIMAL

VALUE_KIND_REFERENCE

VALUE_KIND_STRING

VALUE_KIND_TUPLE

VALUE_KIND_U8

VALUE_KIND_U16

VALUE_KIND_U32

VALUE_KIND_U64

VALUE_KIND_U128

VAULT_BURN_IDENT

VAULT_FREEZE_FEATURE

VAULT_FREEZE_IDENT

VAULT_GET_AMOUNT_IDENT

VAULT_PUT_IDENT

VAULT_RECALL_FEATURE

VAULT_RECALL_IDENT

VAULT_TAKE_ADVANCED_IDENT

VAULT_TAKE_IDENT

VAULT_UNFREEZE_IDENT

WITHDRAWER_ROLE

WITHDRAWER_UPDATER_ROLE

WORKTOP_ASSERT_CONTAINS_AMOUNT_IDENT

WORKTOP_ASSERT_CONTAINS_IDENT

WORKTOP_ASSERT_CONTAINS_NON_FUNGIBLES_IDENT

WORKTOP_BLUEPRINT

WORKTOP_DRAIN_IDENT

WORKTOP_DROP_IDENT

WORKTOP_PUT_IDENT

WORKTOP_TAKE_ALL_IDENT

WORKTOP_TAKE_IDENT

WORKTOP_TAKE_NON_FUNGIBLES_IDENT

XRD

XRD is the native token of the Radix ledger. It is a fungible token, measured in attos (10^-18).

Statics

NO_NETWORK

Traits

ArrayMetadataVal

AsMut

Used to do a cheap mutable-to-mutable reference conversion.

AsRef

Used to do a cheap reference-to-reference conversion.

AsStaticRefDeprecated

A cheap reference-to-reference conversion. Used to convert a value to a reference value with 'static lifetime within generic code.

BasicCategorize

BasicDecode

BasicDescribe

BasicEncode

BasicSbor

Categorize

The Categorize trait marks a rust type as having a fixed value kind for SBOR encoding/decoding.

Cbrt

CheckedMul

CheckedSub

Clone

A common trait for the ability to explicitly duplicate an object.

ContextualDisplay

This trait is used where context is required to correctly display a value.

ContextualSerialize

This trait is used where context is required to correctly serialize a value.

Copy

Types whose values can be duplicated simply by copying bits.

CustomValueKind

Debug

? formatting.

Decode

A data structure that can be decoded from a byte array using SBOR.

Default

A trait for giving a type a useful default value.

Display

Format trait for an empty format, {}.

DoubleEndedIterator

An iterator able to yield elements from both ends.

Drop

Custom code within the destructor.

Encode

A data structure that can be serialized into a byte array using SBOR.

EnumCount

A trait for capturing the number of variants in Enum. This trait can be autoderived by strum_macros.

EnumMessage

Associates additional pieces of information with an Enum. This can be autoimplemented by deriving EnumMessage and annotating your variants with #[strum(message="...")].

EnumProperty

EnumProperty is a trait that makes it possible to store additional information with enum variants. This trait is designed to be used with the macro of the same name in the strum_macros crate. Currently, the only string literals are supported in attributes, the other methods will be implemented as additional attribute types become stabilized.

Eq

Trait for equality comparisons which are equivalence relations.

ExactSizeIterator

An iterator that knows its exact length.

Extend

Extend a collection with the contents of an iterator.

Fn

The version of the call operator that takes an immutable receiver.

FnMut

The version of the call operator that takes a mutable receiver.

FnOnce

The version of the call operator that takes a by-value receiver.

ForWithdrawal

From

Used to do value-to-value conversions while consuming the input value. It is the reciprocal of Into.

FromIterator

Conversion from an Iterator.

FromPrimitive

A generic trait for converting a number to a value.

FromPublicKey

FromStr

Parse a value from a string

HasPublicKeyHash

Into

A value-to-value conversion that consumes the input value. The opposite of From.

IntoEnumIterator

This trait designates that an Enum can be iterated over. It can be auto generated using strum_macros on your behalf.

IntoIterator

Conversion into an Iterator.

IsHash

IsNonAutoGeneratedNonFungibleLocalId

Marks the rust type that represents a non-fungible id, of non-auto-generated kind (i.e. String, Integer and Bytes).

IsNonFungibleLocalId

Marks the rust type that represents a non-fungible id, of any kind (i.e. String, Integer, Bytes and RUID).

IsPublicKeyHash

Iterator

A trait for dealing with iterators.

ManifestCategorize

ManifestDecode

ManifestEncode

ManifestSborEnum

ManifestSborTuple

MetadataVal

NonFungibleData

Represents the data structure of a non-fungible.

NthRoot

One

Defines a multiplicative identity element for Self.

Ord

Trait for types that form a total order.

PartialEq

Trait for equality comparisons.

PartialOrd

Trait for types that form a partial order.

Pow

Binary operator for raising a value to a power.

SborCustomValue

SborEnum

Marker trait for encoding as an Sbor Enum

SborTuple

Marker trait for encoding as an Sbor Tuple

ScryptoCategorize

ScryptoDecode

ScryptoDescribe

ScryptoEncode

ScryptoEvent

ScryptoSbor

ScryptoSborEnum

ScryptoSborTuple

Send

Types that can be transferred across thread boundaries.

Signed

Useful functions for signed numbers (i.e. numbers that can be negative).

SingleMetadataVal

Sized

Types with a constant size known at compile time.

Sqrt

Sync

Types for which it is safe to share references between threads.

ToMetadataEntry

ToOwned

A generalization of Clone to borrowed data.

ToPrimitive

A generic trait for converting a value to a number.

ToRoleEntry

ToString

A trait for converting a value to a String.

Truncate

TryFrom

Simple and safe type conversions that may fail in a controlled way under some circumstances. It is the reciprocal of TryInto.

TryInto

An attempted conversion that consumes self, which may or may not be expensive.

Unpin

Types that can be safely moved after being pinned.

VariantNames

A trait for retrieving the names of each variant in Enum. This trait can be autoderived by strum_macros.

Zero

Defines an additive identity element for Self.

Functions

basic_decode

Decode an instance of T from a slice.

basic_encode

Encode a T into byte array.

bech32_encode_to_fmt

NOTE: The below code is copied with minor alterations from the bech32 crate. These alterations are to avoid using std for allocations, and fit with the sbor no-alloc options.

blake2b_256_hash

check_fungible_amount

check_non_fungible_amount

combine

Combines a u8 with a u8 slice.

component_address

copy_u8_array

Copies a slice to a fixed-sized array.

drop

Disposes of a value.

from_manifest_value

global_caller

A requirement for the global ancestor of the actor who made the latest global call to either be:

hash

Computes the hash digest of a message.

hash_map_new

Creates an empty map with capacity 0 and default Hasher

hash_map_with_capacity

Creates an empty map with given capacity and default Hasher

hash_public_key_bytes

hash_set_new

Creates an empty set with capacity 0 and default Hasher

hash_set_with_capacity

Creates an empty set with given capacity and default Hasher

index_map_new

This is safe for std and no-std use cases (unlike IndexMap::new which disappears when std is not in the toolchain - see this article for deep technical reasons)

index_map_with_capacity

This is safe for std and no-std use cases (unlike IndexMap::with_capacity which disappears when std is not in the toolchain - see this article for deep technical reasons)

index_set_new

This is safe for std and no-std use cases (unlike IndexSet::new which disappears when std is not in the toolchain - see this article for deep technical reasons)

index_set_with_capacity

This is safe for std and no-std use cases (unlike IndexSet::with_capacity which disappears when std is not in the toolchain - see this article for deep technical reasons)

local_address

manifest_decode

manifest_encode

package_address

package_of_direct_caller

A requirement for the immediate caller’s package to equal the given package.

replace_self_package_address

require

require_all_of

require_amount

require_any_of

require_n_of

resource_address

scrypto_decode

Decodes a data structure from a byte array.

scrypto_encode

Encodes a data structure into byte array.

to_manifest_value

Type Definitions

AccessRulesCreateOutput

AccessRulesGetRoleOutput

AccessRulesLockOwnerRoleOutput

AccessRulesSetOwnerRoleOutput

AccessRulesSetRoleOutput

AuthZoneClearOutput

AuthZoneClearVirtualProofsOutput

AuthZoneCreateProofOfAllOutput

AuthZoneCreateProofOfAmountOutput

AuthZoneCreateProofOfNonFungiblesOutput

AuthZoneDrainOutput

AuthZoneDropOutput

AuthZonePopOutput

AuthZonePushOutput

Blake2b256

BucketCreateProofOfAllOutput

BucketGetAmountOutput

BucketGetNonFungibleLocalIdsOutput

BucketGetResourceAddressOutput

BucketPutOutput

BucketTakeAdvancedOutput

BucketTakeNonFungiblesOutput

BucketTakeOutput

BufferId

FieldKey

FungibleBucketCreateProofOfAmountOutput

FungibleBucketLockAmountOutput

FungibleBucketUnlockAmountOutput

FungibleResourceManagerCreateOutput

FungibleResourceManagerCreateWithInitialSupplyOutput

FungibleResourceManagerMintOutput

FungibleVaultCreateProofOfAmountOutput

FungibleVaultLockFeeOutput

FungibleVaultLockFungibleAmountOutput

FungibleVaultUnlockFungibleAmountOutput

HashMap

HashSet

IndexMap

IndexSet

LockHandle

ManifestDecoder

ManifestEncoder

ManifestEnumVariantValue

ManifestOwnedRawPayload

ManifestOwnedRawValue

ManifestRawPayload

ManifestRawValue

ManifestTraverser

ManifestValue

ManifestValueKind

MapKey

MetadataCreateOutput

MetadataCreateWithDataOutput

MetadataGetOutput

MetadataInit

MetadataLockOutput

MetadataRemoveOutput

MetadataSetOutput

NonFungibleBucketCreateProofOfNonFungiblesOutput

NonFungibleBucketLockNonFungiblesOutput

NonFungibleBucketUnlockNonFungiblesOutput

NonFungibleProofGetLocalIdsOutput

NonFungibleResourceManagerCreateOutput

NonFungibleResourceManagerCreateRuidWithInitialSupplyOutput

NonFungibleResourceManagerCreateWithInitialSupplyOutput

NonFungibleResourceManagerExistsOutput

NonFungibleResourceManagerGetNonFungibleOutput

NonFungibleResourceManagerMintOutput

NonFungibleResourceManagerMintRuidOutput

NonFungibleResourceManagerMintSingleRuidOutput

NonFungibleResourceManagerUpdateDataOutput

NonFungibleVaultBurnNonFungiblesOutput

NonFungibleVaultCreateProofOfNonFungiblesOutput

NonFungibleVaultGetNonFungibleLocalIdsOutput

NonFungibleVaultLockNonFungiblesOutput

NonFungibleVaultRecallNonFungiblesOutput

NonFungibleVaultTakeNonFungiblesOutput

NonFungibleVaultUnlockNonFungiblesOutput

ProofCloneOutput

ProofDropOutput

ProofGetAmountOutput

ProofGetResourceAddressOutput

ResourceManagerBurnOutput

ResourceManagerCreateEmptyBucketOutput

ResourceManagerCreateEmptyVaultOutput

ResourceManagerDropEmptyBucketOutput

ResourceManagerGetAmountForWithdrawalOutput

ResourceManagerGetResourceTypeOutput

ResourceManagerGetTotalSupplyOutput

ResourceManagerPackageBurnOutput

RoleDefinition

ScryptoDecoder

ScryptoEncoder

ScryptoOwnedRawPayload

ScryptoOwnedRawValue

ScryptoRawPayload

ScryptoRawValue

ScryptoSchema

ScryptoTraverser

ScryptoTypeData

ScryptoTypeKind

ScryptoValue

ScryptoValueKind

SortedU16Key

ValidatorIndex

An index of a specific validator within the current validator set. To be exact: a ValidatorIndex equal to k references the k-th element returned by the iterator of the IndexMap<ComponentAddress, Validator> in this epoch’s active validator set (which is expected to be sorted by stake, descending). This uniquely identifies the validator, while being shorter than ComponentAddress (we do care about the constant factor of the space taken by LeaderProposalHistory under prolonged liveness break scenarios).

VaultBurnOutput

VaultFreezeOutput

VaultGetAmountOutput

VaultPutOutput

VaultRecallOutput

VaultTakeAdvancedOutput

VaultTakeOutput

VaultUnfreezeOutput

WorktopAssertContainsAmountOutput

WorktopAssertContainsNonFungiblesOutput

WorktopAssertContainsOutput

WorktopDrainOutput

WorktopDropOutput

WorktopPutOutput

WorktopTakeAllOutput

WorktopTakeNonFungiblesOutput

WorktopTakeOutput

Attribute Macros

blueprint

Declares a blueprint.

Derive Macros

AsRefStr

Converts enum variants to &'static str.

AsStaticStr

BasicCategorize

Derive code that returns the value kind - specifically for Basic SBOR.

BasicDecode

Derive code that decodes this data structure from a byte array - specifically for Basic SBOR.

BasicDescribe

Derive code that describes the SBOR type - specifically for Basic SBOR.

BasicEncode

Derive code that encodes this data structure - specifically for Basic SBOR.

BasicSbor

Derive code that implements BasicCategorize, BasicEncode, BasicDecode, and BasicDescribe traits for this struct or enum.

Categorize

Derive code that returns the value kind.

Clone

Derive macro generating an impl of the trait Clone.

Copy

Derive macro generating an impl of the trait Copy.

Debug

Derive macro generating an impl of the trait Debug.

Decode

Derive code that decodes this data structure from a byte array.

Default

Derive macro generating an impl of the trait Default.

Display

Converts enum variants to strings.

Encode

Derive code that encodes this data structure

EnumCount

Add a constant usize equal to the number of variants.

EnumDiscriminants

Generate a new type with only the discriminant names.

EnumIter

Creates a new type that iterates of the variants of an enum.

EnumMessage

Add a verbose message to an enum variant.

EnumProperty

Add custom properties to enum variants.

EnumString

Converts strings to enum variants based on their name.

EnumVariantNames

Implements Strum::VariantNames which adds an associated constant VARIANTS which is an array of discriminant names.

Eq

Derive macro generating an impl of the trait Eq.

FromRepr

Add a function to enum that allows accessing variants by its discriminant

IntoStaticStr

Implements From<MyEnum> for &'static str on an enum.

ManifestCategorize

Derives code for categorizing a struct or enum with Manifest value model.

ManifestDecode

Derives code for decoding a struct or enum with Manifest value model.

ManifestEncode

Derives code for encoding a struct or enum with Manifest value model.

ManifestSbor

Derive code that implements ManifestCategorize, ManifestEncode and ManifestDecode traits for this struct or enum.

NonFungibleData

Derive code that describe a non-fungible data structure.

Ord

Derive macro generating an impl of the trait Ord. The behavior of this macro is described in detail here.

PartialEq

Derive macro generating an impl of the trait PartialEq. The behavior of this macro is described in detail here.

PartialOrd

Derive macro generating an impl of the trait PartialOrd. The behavior of this macro is described in detail here.

Sbor

Derive code that implements Categorize, Encode, Decode, and Describe traits for this struct or enum.

ScryptoCategorize

Derives code for categorizing a struct or enum with Scrypto value model.

ScryptoDecode

Derives code for decoding a struct or enum with Scrypto value model.

ScryptoEncode

Derives code for encoding a struct or enum with Scrypto value model.

ScryptoEvent

Derive code for implementing the required logic to mark a type as being an event.

ScryptoSbor

Derive code that implements ScryptoCategorize, ScryptoEncode, ScryptoDecode, and ScryptoDescribe traits for this struct or enum.

ToString

implements std::string::ToString on en enum