Module scrypto::prelude

Scrypto preludes.

Re-exports

pub use crate::component::*;

pub use crate::constants::*;

pub use crate::core::*;

pub use crate::crypto::*;

pub use crate::math::*;

pub use crate::misc::*;

pub use crate::resource::*;

pub use crate::args;

pub use crate::rule;

pub use crate::access_and_or;

pub use crate::access_rule_node;

pub use crate::borrow_component;

pub use crate::borrow_package;

pub use crate::borrow_resource_manager;

pub use crate::compile_package;

pub use crate::debug;

pub use crate::dec;

pub use crate::error;

pub use crate::include_package;

pub use crate::info;

pub use crate::resource_list;

pub use crate::trace;

pub use crate::warn;

Modules

vec

A contiguous growable array type with heap-allocated contents, written Vec<T>.

Macros

blueprint

Declares a blueprint.

import

Imports a blueprint from its ABI.

vec

Creates a Vec containing the arguments.

Structs

BTreeMap

An ordered map based on a B-Tree.

BTreeSet

An ordered set based on a B-Tree.

HashMap

A hash map implemented with quadratic probing and SIMD lookup.

HashSet

A hash set implemented as a HashMap where the value is ().

String

A UTF-8–encoded, growable string.

Vec

A contiguous growable array type, written as Vec<T>, short for ‘vector’.

Traits

Decode

A data structure that can be decoded from a byte array using SBOR.

Describe

A data structure that can be described using SBOR types.

Encode

A data structure that can be serialized into a byte array using SBOR.

FromStr

Parse a value from a string

ToOwned

A generalization of Clone to borrowed data.

ToString

A trait for converting a value to a String.

TypeId

A SBOR type ID.

Functions

scrypto_decode

Decodes an instance of T from a slice.

scrypto_encode

Encodes a data structure into byte array.

Derive Macros

Decode

Derive code that decodes this data structure from a byte array.

Describe

Derive code that describes this data structure.

Encode

Derive code that encodes this data structure

NonFungibleData

Derive code that describe a non-fungible data structure.

TypeId

Derive code that returns the type ID.