Module scrypto::prelude

Scrypto preludes.

Re-exports

pub use crate::core::*;

pub use crate::resource::*;

pub use crate::types::*;

pub use crate::utils::*;

pub use crate::args;

pub use crate::bdec;

pub use crate::debug;

pub use crate::dec;

pub use crate::error;

pub use crate::include_code;

pub use crate::info;

pub use crate::trace;

pub use crate::warn;

Modules

vec

A contiguous growable array type with heap-allocated contents, written Vec<T>.

Macros

blueprint

Declares a blueprint.

import

Imports a blueprint from its ABI.

vec

Creates a Vec containing the arguments.

Structs

BTreeMap

A map based on a B-Tree.

BTreeSet

A set based on a B-Tree.

HashMap

A hash map implemented with quadratic probing and SIMD lookup.

HashSet

A hash set implemented as a HashMap where the value is ().

String

A UTF-8–encoded, growable string.

Vec

A contiguous growable array type, written as Vec<T> and pronounced ‘vector’.

Enums

LogLevel

Represents the level of a log message.

NewSupply

Represents some supply of resource.

ResourceType

Represents the type of a resource.

Traits

FromStr

Parse a value from a string

ToOwned

A generalization of Clone to borrowed data.

ToString

A trait for converting a value to a String.

Functions

call_engine

Utility function for making a radix engine call.

scrypto_decode

Decodes an instance of T from a slice.

scrypto_encode

Encodes a data structure into byte array.

Attribute Macros

auth

Defines the authorization rule for a method.

Derive Macros

NonFungibleData

Derive code that describe a non-fungible data structure.