Module scrypto::resource

Scrypto resource library.

Re-exports

pub use resource_manager::Mutability::*;

pub use resource_manager::ResourceMethod::*;

Structs

AccessRules

Method authorization rules for a component

Bucket

Represents a transient resource container.

ComponentAuthZone

Represents the auth zone, which is used by system for checking if this component is allowed to

NonFungible

Represents a non-fungible unit.

NonFungibleAddress

Identifier for a non-fungible unit.

NonFungibleId

Represents a key for a non-fungible resource

Proof

Represents a proof of owning some resource.

ResourceAddress

Represents a resource address.

ResourceBuilder

Utility for setting up a new resource.

ResourceManager

Represents a resource manager.

ResourceSystem

Represents the Radix Engine resource subsystem.

SchemaPath

Describes a value located in some sbor given a schema for that sbor

Vault

Represents a persistent resource container on ledger state.

Enums

AccessRule

AccessRuleNode

MintParams

Represents the minting parameters

Mutability

ParseBucketError

Represents an error when decoding bucket.

ParseNonFungibleAddressError

Represents an error when parsing non-fungible address.

ParseNonFungibleIdError

Represents an error when decoding non-fungible id.

ParseProofError

Represents an error when decoding proof.

ParseResourceAddressError

Represents an error when decoding resource address.

ParseVaultError

Represents an error when decoding vault.

ProofRule

Resource Proof Rules

ResourceMethod

ResourceType

Represents the type of a resource.

SoftCount

SoftDecimal

TODO: add documentation for public types once they’re stable.

SoftResource

SoftResourceOrNonFungible

SoftResourceOrNonFungibleList

Constants

DIVISIBILITY_MAXIMUM

The maximum divisibility supported.

DIVISIBILITY_NONE

Not divisible.

Traits

NonFungibleData

Represents the data structure of a non-fungible.

Functions

init_resource_system

Initializes resource subsystem.

require

require_all_of

require_amount

require_any_of

require_n_of

resource_system

Returns the resource subsystem.