Module scrypto::resource

Scrypto resource abstraction.

Re-exports

pub use non_fungible::NonFungible;

pub use bucket::*;

Modules

auth_zone 🔒

bucket

non_fungible

proof 🔒

proof_rule 🔒

resource_builder 🔒

resource_manager 🔒

system 🔒

vault 🔒

Macros

resource_list

Structs

ComponentAuthZone

Represents the auth zone, which is used by system for checking if this component is allowed to

ResourceBuilder

Utility for setting up a new resource.

ResourceManager

Represents a resource manager.

ResourceSystem

Represents the Radix Engine resource subsystem.

ValidatedProof

Represents a proof of owning some resource that has had its resource address validated.

Vault

Constants

DIVISIBILITY_MAXIMUM

The maximum divisibility supported.

DIVISIBILITY_NONE

Not divisible.

Traits

CreateWithNoSupplyBuilder

ScryptoProof

ScryptoVault

SetOwnerBuilder

UpdateAuthBuilder

UpdateMetadataBuilder

PUBLIC TRAITS AND METHODS All public methods first - these all need good rust docs

UpdateNonFungibleAuthBuilder

Functions

init_resource_system

Initializes resource subsystem.

resource_system

Returns the resource subsystem.